### 2022-10-27 #114 @whedon
[PRE REVIEW]: High-performance xPU Stencil Computations in Julia
**Submitting author:** @omlins (<a href="http://orcid.org/0000-0001-8969-5619">Samuel Omlin</a>)
**Repository:** <a href="https://github.com/omlins/ParallelStencil.jl" target ="_blank">https://github.com/omlins/ParallelStencil.jl</a>
**Version:**
**Managing EiC:** Mathieu Besançon
