+++ paper/paper.yml
@@ -11,7 +11,7 @@ keywords:
   - Supercomputing
 authors:
   - name: Samuel Omlin
-    orcid: 0000-0000-0000-000X
+    orcid: 0000-0001-8969-5619
     affiliation: 1
   - name: Ludovic R\"ass
     orcid: 0000-0002-1136-899X
