### 2022-12-09 8f1e640
adjust incremental compilation unit test
