### 2022-12-09 2ef7333
adjust incremental compilation unit test
