### 2022-12-12 7a0c376
fix AMDGPU allocater tests
