### 2022-12-13 #114 @omlins
The paper is in the branch https://github.com/omlins/ParallelStencil.jl/tree/JuliaConProceeding2022
