### 2022-12-19 9851fd2
move macros into code and fix range start
