+++ src/kernel_language.jl
@@ -1,58 +1,6 @@
 #TODO: add ParallelStencil.ParallelKernel. in front of all kernel lang in macros! Later: generalize more for z?
 
 ##
-macro nx_l(args...) check_initialized(); checksinglearg(args...); esc(nx_l(args...)); end
-
-
-##
-macro ny_l(args...) check_initialized(); checksinglearg(args...); esc(ny_l(args...)); end
-
-
-##
-macro nz_l(args...) check_initialized(); checksinglearg(args...); esc(nz_l(args...)); end
-
-
-##
-macro t_h(args...) check_initialized(); checknoargs(args...); esc(t_h(args...)); end
-
-
-##
-macro t_h2(args...) check_initialized(); checktwoargs(args...); esc(t_h2(args...)); end
-
-
-##
-macro tx_h(args...) check_initialized(); checksinglearg(args...); esc(tx_h(args...)); end
-
-
-##
-macro ty_h(args...) check_initialized(); checksinglearg(args...); esc(ty_h(args...)); end
-
-
-##
-macro tx_h2(args...) check_initialized(); checktwoargs(args...); esc(tx_h2(args...)); end
-
-
-##
-macro ty_h2(args...) check_initialized(); checktwoargs(args...); esc(ty_h2(args...)); end
-
-
-##
-macro ix_h(args...) check_initialized(); checksinglearg(args...); esc(ix_h(args...)); end
-
-
-##
-macro iy_h(args...) check_initialized(); checktwoargs(args...); esc(iy_h(args...)); end
-
-
-##
-macro ix_h2(args...) check_initialized(); checktwoargs(args...); esc(ix_h2(args...)); end
-
-
-##
-macro iy_h2(args...) check_initialized(); checktwoargs(args...); esc(iy_h2(args...)); end
-
-
-##
 macro loop(args...) check_initialized(); checkargs_loop(args...); esc(loop(args...)); end
 
 
@@ -87,74 +35,6 @@ function checkargs_loopopt(args...)
 end
 
 
-## FUNCTIONS FOR INDEXING AND DIMENSIONS
-
-function nx_l(hx; package::Symbol=get_package())
-    if (package ∉ SUPPORTED_PACKAGES) @KeywordArgumentError("$ERRMSG_UNSUPPORTED_PACKAGE (obtained: $package).") end
-    return :(@blockDim().x + (2*$hx))
-end
-
-function ny_l(hy; package::Symbol=get_package())
-    if (package ∉ SUPPORTED_PACKAGES) @KeywordArgumentError("$ERRMSG_UNSUPPORTED_PACKAGE (obtained: $package).") end
-    return :(@blockDim().y + (2*$hy))
-end
-
-function nz_l(hz; package::Symbol=get_package())
-    if (package ∉ SUPPORTED_PACKAGES) @KeywordArgumentError("$ERRMSG_UNSUPPORTED_PACKAGE (obtained: $package).") end
-    return :(@blockDim().z + (2*$hz))
-end
-
-function t_h(args...; package::Symbol=get_package())
-    if (package ∉ SUPPORTED_PACKAGES) @KeywordArgumentError("$ERRMSG_UNSUPPORTED_PACKAGE (obtained: $package).") end
-    return :((@threadIdx().y-1)*@blockDim().x + @threadIdx().x)
-end
-
-function t_h2(hx, hy; package::Symbol=get_package())
-    if (package ∉ SUPPORTED_PACKAGES) @KeywordArgumentError("$ERRMSG_UNSUPPORTED_PACKAGE (obtained: $package).") end
-    return :(ParallelStencil.@t_h() + ParallelStencil.@nx_l($hx)*ParallelStencil.@ny_l($hy) - @blockDim().x*@blockDim().y)
-end
-
[…]
