### 2022-12-19 e74adc1
add error if too many positional args in loopopt
