### 2022-12-19 991c70a
add error if maxsize not dividable without rest by nthreads
