### 2022-12-19 a654ca6
remove possibility for smaller block size at end
