### 2022-12-19 4feb2a7
add error if maxsize not dividable without rest by nthreads
