### 2022-12-20 0286434
introduce stencilsize
