+++ src/kernel_language.jl
@@ -52,33 +52,34 @@ function loop(index::Symbol, optdim::Integer, loopsize, body; package::Symbol=ge
 end
 
 
-#function add_loopopt(body::Expr, indices::Array{<:Union{Expr,Symbol}}, optvars::Union{Expr,Symbol}, optdim::Integer, loopsize::Union{Expr,Symbol,Integer}, halosize::Union{Integer,NTuple{N,Integer} where N})
+#function add_loopopt(body::Expr, indices::Array{<:Union{Expr,Symbol}}, optvars::Union{Expr,Symbol}, optdim::Integer, loopsize::Union{Expr,Symbol,Integer}, stencilsize::Union{Integer,NTuple{N,Integer} where N})
 #TODO: see what to do with global consts as SHMEM_HALO_X,... Support later multiple vars for opt (now just A=T...)
 #TODO: add input check and errors
 #TODO: maybe gensym with macro @gensym
 # TODO: create a run time check for requirement: 
 # In order to be able to read the data into shared memory in only two statements, the number of threats must be at least half of the size of the shared memory block plus halo; thus, the total number of threads in each dimension must equal the range length, as else there would be smaller thread blocks at the boundaries (threads overlapping the range are sent home). These smaller blocks would be likely not to match the criteria for a correct reading of the data to shared memory. In summary the following requirements must be matched: @gridDim().x*@blockDim().x - $rangelength_x == 0; @gridDim().y*@blockDim().y - $rangelength_y > 0
-function loopopt(caller::Module, indices, optvars, optdim::Integer, loopsize, halosize, indices_shift, body; package::Symbol=get_package())
+function loopopt(caller::Module, indices, optvars, optdim::Integer, loopsize, stencilsize, indices_shift, body; package::Symbol=get_package())
     if !isa(optvars, Symbol) @KeywordArgumentError("at present, only one optvar is supported.") end
     A = optvars 
     if (package ∉ SUPPORTED_PACKAGES) @KeywordArgumentError("$ERRMSG_UNSUPPORTED_PACKAGE (obtained: $package).") end
     if isa(indices,Expr) indices = indices.args else indices = (indices,) end
-    halosize = eval_arg(caller, halosize)
+    stencilsize = eval_arg(caller, stencilsize)
     noexpr = :(begin end)
     if optdim == 3
         ranges         = RANGES_VARNAME
         rangelength_z  = RANGELENGTHS_VARNAMES[3]
         tz_g           = THREADIDS_VARNAMES[3]
-        hx, hy         = halosize
-        shmem          = (hx>0 || hy>0)
+        sx1,sx2, sy1,sy2 ,sz1,sz2 = stencilsize
+        hx1,hx2, hy1,hy2 = stencilsize
+        shmem          = (hx1>0 || hx2>0 || hy1>0 || hy2>0)
         _ix, _iy, _iz  = indices
         _i             = gensym_world("i", @__MODULE__)
-        ix             = (indices_shift[1] > 0) ? :(($_ix + $(indices_shift[1]))) : _ix
-        iy             = (indices_shift[2] > 0) ? :(($_iy + $(indices_shift[2]))) : _iy
-        iz             = (indices_shift[3] > 0) ? :(($_iz + $(indices_shift[3]))) : _iz
-        i              = (indices_shift[3] > 0) ? :(($_i  - $(indices_shift[3]))) : _i
[…]
