### 2022-12-20 5dd5798
introduce stencilsize
