+++ src/parallel.jl
@@ -70,13 +70,13 @@ parallel_async(caller::Module, args::Union{Symbol,Expr}...; package::Symbol=get_
 function parallel(caller::Module, args::Union{Symbol,Expr}...; package::Symbol=get_package(), async::Bool=false)
     if is_kernel(args[end])
         posargs, kwargs_expr, kernelarg = split_parallel_args(args, is_call=false)
-        kwargs     = extract_kwargs(caller, kwargs_expr, (:loopopt, :optvars, :optdim, :loopsize, :halosize), "@parallel <kernel>"; eval_args=(:loopopt, :optdim, :halosize))
+        kwargs     = extract_kwargs(caller, kwargs_expr, (:loopopt, :optvars, :optdim, :loopsize, :stencilsize), "@parallel <kernel>"; eval_args=(:loopopt, :optdim, :stencilsize))
         numbertype = get_numbertype()
         ndims      = get_ndims()
         parallel_kernel(caller, package, numbertype, ndims, kernelarg, posargs...; kwargs)
     elseif is_call(args[end])
         posargs, kwargs_expr, kernelarg = split_parallel_args(args)
-        kwargs, backend_kwargs_expr = extract_kwargs(caller, kwargs_expr, (:loopopt, :optvars, :optdim, :loopsize, :halosize), "@parallel <kernelcall>", true; eval_args=(:loopopt, :optdim, :halosize))
+        kwargs, backend_kwargs_expr = extract_kwargs(caller, kwargs_expr, (:loopopt, :optvars, :optdim, :loopsize, :stencilsize), "@parallel <kernelcall>", true; eval_args=(:loopopt, :optdim, :stencilsize))
         if haskey(kwargs, :loopopt) && kwargs.loopopt
             if (length(posargs) > 1) @ArgumentError("maximum one positional argument (ranges) is allowed in a @parallel loopopt=true call.") end
             parallel_call_loopopt(posargs..., kernelarg, backend_kwargs_expr, async; kwargs...)
@@ -90,7 +90,7 @@ end
 function parallel_indices(caller::Module, args::Union{Symbol,Expr}...; package::Symbol=get_package())
     numbertype = get_numbertype()
     posargs, kwargs_expr = split_args(args)
-    kwargs = extract_kwargs(caller, kwargs_expr, (:loopopt, :optvars, :optdim, :loopsize, :halosize), "@parallel_indices"; eval_args=(:loopopt, :optdim, :halosize))
+    kwargs = extract_kwargs(caller, kwargs_expr, (:loopopt, :optvars, :optdim, :loopsize, :stencilsize), "@parallel_indices"; eval_args=(:loopopt, :optdim, :stencilsize))
     if haskey(kwargs, :loopopt) && kwargs.loopopt
         parallel_indices_loopopt(package, numbertype, posargs...; kwargs...)
     else
@@ -101,13 +101,14 @@ end
 
 ## @PARALLEL KERNEL FUNCTIONS
 
-function parallel_indices_loopopt(package::Symbol, numbertype::DataType, indices::Union{Symbol,Expr}, kernel::Expr; loopopt::Bool=false, optvars::Union{Expr,Symbol}=Symbol(""), optdim::Integer=determine_optdim(), loopsize::Union{Expr,Symbol,Integer}=compute_loopsize(), halosize::Union{Integer,NTuple{N,Integer} where N}=compute_halosize(), indices_shift::Tuple{<:Integer,<:Integer,<:Integer}=(0,0,0))
[…]
