### 2022-12-20 ed60173
move from stencilsize to stencilranges
