### 2022-12-20 333f720
move from stencilsize to stencilranges
