### 2022-12-20 20e7456
fix y-dimension register usage
