### 2022-12-22 4051d6e
generalize substitute and in expression check
