### 2022-12-22 3b726c1
evaluate stencil offsets at parse time
