### 2022-12-22 aebe72d
evaluate stencil offsets at parse time
