### 2022-12-22 808f4f0
extract offset
