### 2022-12-22 deed435
extract offset
