+++ src/kernel_language.jl
@@ -62,14 +62,20 @@ function loopopt(caller::Module, indices, optvars, optdim::Integer, loopsize, st
     if !isa(optvars, Symbol) @KeywordArgumentError("at present, only one optvar is supported.") end
     A = optvars 
     if (package ∉ SUPPORTED_PACKAGES) @KeywordArgumentError("$ERRMSG_UNSUPPORTED_PACKAGE (obtained: $package).") end
+    if     (package == PKG_CUDA)    int_type = INT_CUDA
+    elseif (package == PKG_AMDGPU)  int_type = INT_AMDGPU
+    elseif (package == PKG_THREADS) int_type = INT_THREADS
+    end
     if isa(indices,Expr) indices = indices.args else indices = (indices,) end
     stencilranges = eval_arg(caller, stencilranges)
     rx, ry ,rz    = stencilranges
     rx1, ry1, rz1 = rx.start, ry.start ,rz.start
     rx2, ry2, rz2 = rx.stop, ry.stop ,rz.stop
     noexpr        = :(begin end)
-    body          = eval_offsets(caller, body, indices)
-    offsets       = extract_offsets(caller, body, indices, optdim)
+    body          = eval_offsets(caller, body, indices, int_type)
+    offsets       = extract_offsets(caller, body, indices, int_type, optdim)
+    regqueue_head, regqueue_tail, optdim_offset_max  = define_regqueue(offsets, stencilranges, A, indices, int_type, optdim)
+    @show regqueue_head, regqueue_tail, optdim_offset_max
     if optdim == 3
         ranges         = RANGES_VARNAME
         rangelength_z  = RANGELENGTHS_VARNAMES[3]
@@ -100,22 +106,32 @@ function loopopt(caller::Module, indices, optvars, optdim::Integer, loopsize, st
         iy_h2          = gensym_world("iy_h2", @__MODULE__)
         loopoffset     = gensym_world("loopoffset", @__MODULE__)
         A_izp1         = gensym_world(string(A, "_izp1"), @__MODULE__)
-        A_ix_iy_izm1   = gensym_world(string(A, "_ix_iy_izm1"), @__MODULE__)
-        A_ix_iy_iz     = gensym_world(string(A, "_ix_iy_iz"), @__MODULE__)
-        A_ix_iy_izp1   = gensym_world(string(A, "_ix_iy_izp1"), @__MODULE__)
-        A_ixm1_iy_iz   = gensym_world(string(A, "_ixm1_iy_iz"), @__MODULE__)
-        A_ixp1_iy_iz   = gensym_world(string(A, "_ixp1_iy_iz"), @__MODULE__)
-        A_ix_iym1_iz   = gensym_world(string(A, "_ix_iym1_iz"), @__MODULE__)
-        A_ix_iyp1_iz   = gensym_world(string(A, "_ix_iyp1_iz"), @__MODULE__)
 
-        
-        if (rz1 < 0) body = substitute(body, :($A[$ix,$iy,$iz-1]), A_ix_iy_izm1) end
-        body = substitute(body, :($A[$ix,$iy,$iz  ]), A_ix_iy_iz  )
-        if (rz2 > 0) body = substitute(body, :($A[$ix,$iy,$iz+1]), A_ix_iy_izp1) end
-        if (rx1 < 0) body = substitute(body, :($A[$ix-1,$iy,$iz]), A_ixm1_iy_iz) end
-        if (rx2 > 0) body = substitute(body, :($A[$ix+1,$iy,$iz]), A_ixp1_iy_iz) end
[…]
