### 2023-01-05 c145fb3
generalize register queuing
