+++ src/kernel_language.jl
@@ -68,19 +68,19 @@ function loopopt(caller::Module, indices, optvars, optdim::Integer, loopsize, st
     end
     if isa(indices,Expr) indices = indices.args else indices = (indices,) end
     stencilranges = eval_arg(caller, stencilranges)
-    rx, ry ,rz    = stencilranges
-    rx1, ry1, rz1 = rx.start, ry.start ,rz.start
-    rx2, ry2, rz2 = rx.stop, ry.stop ,rz.stop
     noexpr        = :(begin end)
     body          = eval_offsets(caller, body, indices, int_type)
     offsets       = extract_offsets(caller, body, indices, int_type, optdim)
-    regqueue_head, regqueue_tail, optdim_offset_max  = define_regqueue(offsets, stencilranges, A, indices, int_type, optdim)
-    @show regqueue_head, regqueue_tail, optdim_offset_max
+    regqueue_head, regqueue_tail, offset_min, offset_max = define_regqueue(offsets, stencilranges, A, indices, int_type, optdim)
+    @show regqueue_head, regqueue_tail, offset_min, offset_max
     if optdim == 3
         ranges         = RANGES_VARNAME
         rangelength_z  = RANGELENGTHS_VARNAMES[3]
         tz_g           = THREADIDS_VARNAMES[3]
-        hx1,hx2, hy1,hy2 = -rx1,rx2, -ry1,ry2
+        oz_max         = offset_max[3]
+        hx1, hy1       = -1 .* offset_min[1:2]
+        hx2, hy2       = offset_max[1:2]
+        # hx1,hx2, hy1,hy2 = -rx1,rx2, -ry1,ry2
         shmem          = (hx1+hx2>0 || hy1+hy2>0)
         _ix, _iy, _iz  = indices
         _i             = gensym_world("i", @__MODULE__)
@@ -225,17 +225,17 @@ $((shmem ?
 $((:(
                         $(regs[oz])           = $(regs[oz+1])
     )
-    for regs in values(regqueue_tail) for oz in sort(keys(regs)) if oz<=optdim_offset_max-2
-        # for (oxy, regs) in regqueue_tail for (oz, reg) in regs if oz<=optdim_offset_max-2
+    for regs in values(regqueue_tail) for oz in sort(keys(regs)) if oz<=oz_max-2
+        # for (oxy, regs) in regqueue_tail for (oz, reg) in regs if oz<=oz_max-2
   )...
 )
                         # $A_ix_iy_izm1 = $A_ix_iy_iz
 $((:(                   $reg           = $(regsource(A_izp1, oxy, (tx, ty)))
     )
-    for (oxy, regs) in regqueue_tail for (oz, reg) in regs if oz==optdim_offset_max-1 && !(haskey(regqueue_head, oxy) && haskey(regqueue_head[oxy], oz+1))
+    for (oxy, regs) in regqueue_tail for (oz, reg) in regs if oz==oz_max-1 && !(haskey(regqueue_head, oxy) && haskey(regqueue_head[oxy], oz+1))
   )...
 )
-# for oxy in keys(regqueue_tail) for (reg, oz) in zip(regqueue_tail[oxy], filter(oz -> oz==optdim_offset_max-1, keys(regqueue_tail[oxy])))
+# for oxy in keys(regqueue_tail) for (reg, oz) in zip(regqueue_tail[oxy], filter(oz -> oz==oz_max-1, keys(regqueue_tail[oxy])))
 # $(rx1<0 ? :(            $A_ixm1_iy_iz = $A_izp1[$tx-1,$ty]                                ) : noexpr)
[…]
