### 2023-01-08 9ddd3e0
handle incoherent stencilranges
