+++ src/kernel_language.jl
@@ -82,6 +82,9 @@ function loopopt(caller::Module, indices, optvars, optdim::Integer, loopsize, st
         hx2, hy2       = offset_max[1:2]
         # hx1,hx2, hy1,hy2 = -rx1,rx2, -ry1,ry2
         shmem          = (hx1+hx2>0 || hy1+hy2>0)
+        offset_span    = offset_max .- offset_min
+        oz_span        = offset_span[3]
+        loopstart      = 1 - oz_span
         _ix, _iy, _iz  = indices
         _i             = gensym_world("i", @__MODULE__)
         ix             = _ix # (indices_shift[1] > 0) ? :(($_ix + $(indices_shift[1]))) : _ix
@@ -105,7 +108,7 @@ function loopopt(caller::Module, indices, optvars, optdim::Integer, loopsize, st
         iy_h           = gensym_world("iy_h", @__MODULE__)
         iy_h2          = gensym_world("iy_h2", @__MODULE__)
         loopoffset     = gensym_world("loopoffset", @__MODULE__)
-        A_izp1         = gensym_world(string(A, "_izp1"), @__MODULE__)
+        A_head         = gensym_world(varname(A, (oz_max,); i="iz"), @__MODULE__)
 
         # A_ix_iy_izm1   = gensym_world(string(A, "_ix_iy_izm1"), @__MODULE__)
         # A_ix_iy_iz     = gensym_world(string(A, "_ix_iy_iz"), @__MODULE__)
@@ -145,8 +148,7 @@ function loopopt(caller::Module, indices, optvars, optdim::Integer, loopsize, st
                         end
                     end
         end
-        loopstart = 0 #TODO: if in z neighbors beyond iz-1 are accessed, then this needs to be bigger (also range_z_start-...!). NOTE: not the same as stencilranges which is for shmemhalo
-        
+
         return quote
                     $tx            = @threadIdx().x + $hx1
                     $ty            = @threadIdx().y + $hy1
@@ -167,7 +169,7 @@ $(shmem ? quote
         noexpr
 )
                     $loopoffset    = (@blockIdx().z-1)*$loopsize #TODO: MOVE UP - see no perf change! interchange other lines!
-$(shmem ? :(        $A_izp1        = @sharedMem(eltype($A), ($nx_l, $ny_l))              ) : noexpr)
+$(shmem ? :(        $A_head        = @sharedMem(eltype($A), ($nx_l, $ny_l))              ) : noexpr)
 $((:(               $reg           = 0.0
     ) 
     for regs in values(regqueue_tail) for reg in values(regs)
@@ -179,7 +181,7 @@ $((:(               $reg           = 0.0
   )...
 )
                     # TODO: the following line should be removed later and replaced with an earlier start of the loop.
-                    $(regqueue_tail[(0,0)][0]) = (@blockIdx().z>1) ? $A[$ix,$iy,$range_z_start-1+$loopoffset] : 0.0
+                    #$(regqueue_tail[(0,0)][0]) = (@blockIdx().z>1) ? $A[$ix,$iy,$range_z_start-1+$loopoffset] : 0.0
 # $(       (:(        $(regqueue_head[oxy][oz]) = 0.0                                      ) for oxy in keys(regqueue_head), oz in keys(regqueue_head[oxy]))...)
 #                     $A_ix_iy_izm1  = 0.0
 #                     $A_ix_iy_iz    = (@blockIdx().z>1) ? $A[$ix,$iy,$range_z_start-1+$loopoffset] : 0.0
@@ -192,35 +194,35 @@ $((:(               $reg           = 0.0
[…]
