### 2023-01-19 842acfb
generalize register queuing
