### 2023-01-20 70b652c
improve check package error message
