+++ src/kernel_language.jl
@@ -58,7 +58,7 @@ end
 #TODO: maybe gensym with macro @gensym
 # TODO: create a run time check for requirement: 
 # In order to be able to read the data into shared memory in only two statements, the number of threats must be at least half of the size of the shared memory block plus halo; thus, the total number of threads in each dimension must equal the range length, as else there would be smaller thread blocks at the boundaries (threads overlapping the range are sent home). These smaller blocks would be likely not to match the criteria for a correct reading of the data to shared memory. In summary the following requirements must be matched: @gridDim().x*@blockDim().x - $rangelength_x == 0; @gridDim().y*@blockDim().y - $rangelength_y > 0
-function loopopt(caller::Module, indices, optvars, optdim::Integer, loopsize, stencilranges::Union{Nothing,NTuple{N,UnitRange} where N}, indices_shift, body; package::Symbol=get_package())
+function loopopt(caller::Module, indices, optvars, optdim::Integer, loopsize, stencilranges, body; package::Symbol=get_package())
     if !isa(optvars, Symbol) @KeywordArgumentError("at present, only one optvar is supported.") end
     A = optvars 
     if (package ∉ SUPPORTED_PACKAGES) @KeywordArgumentError("$ERRMSG_UNSUPPORTED_PACKAGE (obtained: $package).") end
@@ -69,7 +69,6 @@ function loopopt(caller::Module, indices, optvars, optdim::Integer, loopsize, st
     if isa(indices,Expr) indices = indices.args else indices = (indices,) end
     fullrange     = typemin(int_type):typemax(int_type)
     stencilranges = isnothing(stencilranges) ? (fullrange, fullrange, fullrange) : eval_arg(caller, stencilranges)
-    noexpr        = :(begin end)
     body          = eval_offsets(caller, body, indices, int_type)
     offsets       = extract_offsets(caller, body, indices, int_type, optdim)
     regqueue_head, regqueue_tail, offset_min, offset_max = define_regqueue(offsets, stencilranges, A, indices, int_type, optdim)
@@ -81,19 +80,14 @@ function loopopt(caller::Module, indices, optvars, optdim::Integer, loopsize, st
         oz_max         = offset_max[3]
         hx1, hy1       = -1 .* offset_min[1:2]
         hx2, hy2       = offset_max[1:2]
-        # hx1,hx2, hy1,hy2 = -rx1,rx2, -ry1,ry2
         shmem          = (hx1+hx2>0 || hy1+hy2>0)
         offset_span    = offset_max .- offset_min
         oz_span        = offset_span[3]
-        loopstart      = 1 - oz_span
-        _ix, _iy, _iz  = indices
-        _i             = gensym_world("i", @__MODULE__)
-        ix             = _ix # (indices_shift[1] > 0) ? :(($_ix + $(indices_shift[1]))) : _ix
-        iy             = _iy # (indices_shift[2] > 0) ? :(($_iy + $(indices_shift[2]))) : _iy
-        iz             = _iz # (indices_shift[3] > 0) ? :(($_iz + $(indices_shift[3]))) : _iz
-        i              = _i # (indices_shift[3] > 0) ? :(($_i  - $(indices_shift[3]))) : _i
-        range_z        = :(($ranges[3])[$tz_g]) # (indices_shift[3] > 0) ? :(($ranges[3])[$tz_g] - $(indices_shift[3])) : :(($ranges[3])[$tz_g])
[…]
