### 2023-01-20 2900eac
add unit tests for loopopt
