### 2023-01-27 80eee91
generalize register queuing - for multiple arrays
