+++ src/kernel_language.jl
@@ -51,7 +51,7 @@ function loop(index::Symbol, optdim::Integer, loopsize, body; package::Symbol=ge
     end
 end
 
-
+# stencilranges=(A=(1:1, 0:2, 1:1), B=(1:1, 0:2, 1:1))
 #function add_loopopt(body::Expr, indices::Array{<:Union{Expr,Symbol}}, optvars::Union{Expr,Symbol}, optdim::Integer, loopsize::Union{Expr,Symbol,Integer}, stencilranges::Union{Integer,NTuple{N,Integer} where N})
 #TODO: see what to do with global consts as SHMEM_HALO_X,... Support later multiple vars for opt (now just A=T...)
 #TODO: add input check and errors
@@ -59,70 +59,102 @@ end
 # TODO: create a run time check for requirement: 
 # In order to be able to read the data into shared memory in only two statements, the number of threats must be at least half of the size of the shared memory block plus halo; thus, the total number of threads in each dimension must equal the range length, as else there would be smaller thread blocks at the boundaries (threads overlapping the range are sent home). These smaller blocks would be likely not to match the criteria for a correct reading of the data to shared memory. In summary the following requirements must be matched: @gridDim().x*@blockDim().x - $rangelength_x == 0; @gridDim().y*@blockDim().y - $rangelength_y > 0
 function loopopt(caller::Module, indices, optvars, optdim::Integer, loopsize, stencilranges, body; package::Symbol=get_package())
+    #TODO: the two following lines have to be removed later
     if !isa(optvars, Symbol) @KeywordArgumentError("at present, only one optvar is supported.") end
-    A = optvars 
+    stencilranges = nothing
+
+    optvars = Tuple(extract_tuple(optvars)) #TODO: make this function actually return directly a tuple rather than an array
+    indices = Tuple(extract_tuple(indices))
+    readonlyvars  = find_readonlyvars(body, indices)
+    if optvars == (Symbol(""),)
+        optvars = Tuple(keys(readonlyvars))
+    else
+        for A in optvars
+            if !haskey(readonlyvars, A) @IncoherentArgumentError("incoherent argument optvars in @loopopt: optimization can only be applied to arrays that are only read within the kernel (not applicable to: $A).") end
+        end
+    end
     if (package ∉ SUPPORTED_PACKAGES) @KeywordArgumentError("$ERRMSG_UNSUPPORTED_PACKAGE (obtained: $package).") end
     if     (package == PKG_CUDA)    int_type = INT_CUDA
     elseif (package == PKG_AMDGPU)  int_type = INT_AMDGPU
     elseif (package == PKG_THREADS) int_type = INT_THREADS
     end
-    if isa(indices,Expr) indices = indices.args else indices = (indices,) end
     fullrange     = typemin(int_type):typemax(int_type)
-    stencilranges = isnothing(stencilranges) ? (fullrange, fullrange, fullrange) : eval_arg(caller, stencilranges)
+    fullranges    = (fullrange, fullrange, fullrange)
+    stencilranges = isnothing(stencilranges) ? (A = fullranges for A in optvars) : eval_arg(caller, stencilranges)
+    stencilranges = Dict(A => (A ∈ keys(stencilranges)) ? stencilranges.A : fullranges for A in optvars)
[…]
