### 2023-01-27 29cfda4
generalize register queuing - for multiple arrays
