+++ src/kernel_language.jl
@@ -91,17 +91,18 @@ function loopopt(caller::Module, indices, optvars, optdim::Integer, loopsize, st
     # regqueue_head, regqueue_tail, offset_min, offset_max = regqueue_heads[A], regqueue_tails[A], offset_mins[A], offset_maxs[A]
     # @show stencilranges, regqueue_head, regqueue_tail, offset_min, offset_max
     if optdim == 3
-        oz_maxs, hx1s, hy1s, hx2s, hy2s, shmems, offset_spans, oz_spans, loopstarts = define_helper_variables(offset_mins, offset_maxs, optvars, optdim)
-        loopstart = loopstarts[optvars[1]]
-        # oz_max, hx1, hy1, hx2, hy2, shmem, offset_span, oz_span, loopstart = oz_maxs[A], hx1s[A], hy1s[A], hx2s[A], hy2s[A], shmems[A], offset_spans[A], oz_spans[A], loopstarts[A]
+        oz_maxs, hx1s, hy1s, hx2s, hy2s, use_shmems, offset_spans, oz_spans, loopentrys = define_helper_variables(offset_mins, offset_maxs, optvars, optdim)
+        # loopentry = loopentrys[optvars[1]]
+        loopstart = minimum(values(loopentrys)) #TODO:
+        # oz_max, hx1, hy1, hx2, hy2, use_shmem, offset_span, oz_span, loopentry = oz_maxs[A], hx1s[A], hy1s[A], hx2s[A], hy2s[A], use_shmems[A], offset_spans[A], oz_spans[A], loopentrys[A]
         # oz_max         = offset_max[3]
         # hx1, hy1       = -1 .* offset_min[1:2]
         # hx2, hy2       = offset_max[1:2]
-        # shmem          = (hx1+hx2>0 || hy1+hy2>0)
+        # use_shmem          = (hx1+hx2>0 || hy1+hy2>0)
         # offset_span    = offset_max .- offset_min
         # oz_span        = offset_span[3]
-        # loopstart      = 1 - oz_span #TODO: make possibility to do first and last read in z dimension directly into registers without halo
-        shmem_symbols = define_shmem_symbols(oz_maxs, optvars, optdim) #TODO: if shmems[A] can be removed in the loops if this is defined only for arrays it effectively must use shared memory
+        # loopentry      = 1 - oz_span #TODO: make possibility to do first and last read in z dimension directly into registers without halo
+        shmem_symbols = define_shmem_symbols(oz_maxs, optvars, optdim) #TODO: if use_shmems[A] can be removed in the loops if this is defined only for arrays it effectively must use shared memory
         # s=shmem_symbols[A]; tx, ty, nx_l, ny_l, t_h, t_h2, tx_h, tx_h2, ty_h, ty_h2, ix_h, ix_h2, iy_h, iy_h2, A_head = s[:tx], s[:ty], s[:nx_l], s[:ny_l], s[:t_h], s[:t_h2], s[:tx_h], s[:tx_h2], s[:ty_h], s[:ty_h2], s[:ix_h], s[:ix_h2], s[:iy_h], s[:iy_h2], s[:A_head]
         # tx             = gensym_world("tx", @__MODULE__)
         # ty             = gensym_world("ty", @__MODULE__)
@@ -146,12 +147,7 @@ function loopopt(caller::Module, indices, optvars, optdim::Integer, loopsize, st
[…]
