### 2023-02-20 707d99c
accept offset argument in shared memory macro
