+++ src/ParallelKernel/kernel_language.jl
@@ -53,11 +53,11 @@ macro sync_threads(args...) check_initialized(); checknoargs(args...); esc(sync_
 
 
 ##
-# NOTE: the optional offset parameter of cuDynamicSharedMem is currently not exposed.
 const SHAREDMEM_DOC = """
-    @sharedMem(T, dims)
+    @sharedMem(T, dims, offset::Integer=0)
 
-Create an array that is *shared* between the threads of a block (i.e. accessible only by the threads of a same block), with element type `T` and size specified by `dims`.
+Create an array that is *shared* between the threads of a block (i.e. accessible only by the threads of a same block), with element type `T` and size specified by `dims`. 
+When multiple shared memory arrays are created within a kernel, then all arrays except for the first one typically need to define the `offset` to the base shared memory pointer in bytes.
 
 !!! note "Note"
     The amount of shared memory needs to specified when launching the kernel (keyword argument `shmem`).
@@ -93,7 +93,7 @@ function checknoargs(args...)
 end
 
 function checkargs_sharedMem(args...)
-    if (length(args) != 2) @ArgumentError("wrong number of arguments.") end
+    if !(2 <= length(args) <= 3) @ArgumentError("wrong number of arguments.") end
 end
 
 
