### 2023-02-20 5377e71
add global loop optimization keyword argument
