+++ src/init_parallel_stencil.jl
@@ -40,13 +40,14 @@ macro init_parallel_stencil(args...)
     if (length(posargs) == 3) package, numbertype_val, ndims_val = extract_posargs_init(__module__, posargs...)
     else                      package, numbertype_val, ndims_val = extract_kwargs_init(__module__, kwargs)
     end
+    loopopt_val = extract_kwargs_optional(__module__, kwargs)
     if (package == PKG_NONE) @ArgumentError("the package argument cannot be ommited.") end #TODO: this error message will disappear, once the package can be defined at runtime.
     if (ndims == NDIMS_NONE) @ArgumentError("the ndims argument cannot be ommited.") end #TODO: this error message will disappear, once the ndims can be defined at runtime.
-    check_already_initialized(package, numbertype_val, ndims_val)
-    esc(init_parallel_stencil(__module__, package, numbertype_val, ndims_val))
+    check_already_initialized(package, numbertype_val, ndims_val, loopopt_val)
+    esc(init_parallel_stencil(__module__, package, numbertype_val, ndims_val, loopopt_val))
 end
 
-function init_parallel_stencil(caller::Module, package::Symbol, numbertype::DataType, ndims::Integer)
+function init_parallel_stencil(caller::Module, package::Symbol, numbertype::DataType, ndims::Integer, loopopt::Bool)
     if (numbertype == NUMBERTYPE_NONE) datadoc_call = :(@doc replace(ParallelStencil.ParallelKernel.DATA_DOC_NUMBERTYPE_NONE, "@init_parallel_kernel" => "@init_parallel_stencil") Data)
     else                               datadoc_call = :(@doc replace(ParallelStencil.ParallelKernel.DATA_DOC,                 "@init_parallel_kernel" => "@init_parallel_stencil") Data)
     end
@@ -54,6 +55,7 @@ function init_parallel_stencil(caller::Module, package::Symbol, numbertype::Data
     set_package(package)
     set_numbertype(numbertype)
     set_ndims(ndims)
+    set_loopopt(loopopt)
     set_initialized(true)
     return nothing
 end
@@ -63,12 +65,14 @@ macro is_initialized() is_initialized() end
 macro get_package() get_package() end
 macro get_numbertype() get_numbertype() end
 macro get_ndims() get_ndims() end
+macro get_loopopt() get_loopopt() end
 let
-    global is_initialized, set_initialized, set_package, get_package, set_numbertype, get_numbertype, set_ndims, get_ndims, check_initialized, check_already_initialized
+    global is_initialized, set_initialized, set_package, get_package, set_numbertype, get_numbertype, set_ndims, get_ndims, set_loopopt, get_loopopt, check_initialized, check_already_initialized
     _is_initialized::Bool       = false
     package::Symbol             = PKG_NONE
     numbertype::DataType        = NUMBERTYPE_NONE
     ndims::Integer              = NDIMS_NONE
+    loopopt::Bool               = false
     set_initialized(flag::Bool) = (_is_initialized = flag)
     is_initialized()            = _is_initialized
     set_package(pkg::Symbol)    = (package = pkg)
@@ -77,11 +81,13 @@ let
     get_numbertype()            = numbertype
     set_ndims(n::Integer)       = (ndims = n)
     get_ndims()                 = ndims
+    set_loopopt(loopopt::Bool)  = (loopopt = loopopt)
+    get_loopopt()               = loopopt
     check_initialized()         = if !is_initialized() @NotInitializedError("no macro or function of the module can be called before @init_parallel_stencil.") end
 
-    function check_already_initialized(package::Symbol, numbertype::DataType, ndims::Integer)
[…]
