+++ src/kernel_language.jl
@@ -5,11 +5,15 @@ macro loop(args...) check_initialized(); checkargs_loop(args...); esc(loop(args.
 
 
 ##
-macro loopopt(args...) check_initialized(); checkargs_loopopt(args...); esc(loopopt(__module__, args...)); end
+macro loopopt(args...) check_initialized(); checkargs_loopopt(args...); esc(loopopt(args[1], __module__, args[2:end]...)); end
 
 
 ##
 macro shortif(args...) check_initialized(); checktwoargs(args...); esc(shortif(args...)); end
+
+
+##
+macro return_nothing(args...) check_initialized(); checknoargs(args...); esc(return_nothing(args...)); end
 
 
 ## ARGUMENT CHECKS
@@ -22,7 +26,7 @@ function checksinglearg(args...)
     if (length(args) != 1) @ArgumentError("wrong number of arguments.") end
 end
 
-function checktwoargs(args...)
+function checktwoargs(args...)  
     if (length(args) != 2) @ArgumentError("wrong number of arguments.") end
 end
 
@@ -31,7 +35,7 @@ function checkargs_loop(args...)
 end
 
 function checkargs_loopopt(args...)
-    if (length(args) != 7 && length(args) != 6 && length(args) != 3) @ArgumentError("wrong number of arguments.") end
+    if (length(args) != 8 && length(args) != 7 && length(args) != 4) @ArgumentError("wrong number of arguments.") end
 end
 
 
@@ -51,18 +55,14 @@ function loop(index::Symbol, optdim::Integer, loopsize, body; package::Symbol=ge
     end
 end
 
-# stencilranges=(A=(1:1, 0:2, 1:1), B=(1:1, 0:2, 1:1))
-#function add_loopopt(body::Expr, indices::Array{<:Union{Expr,Symbol}}, optvars::Union{Expr,Symbol}, optdim::Integer, loopsize::Union{Expr,Symbol,Integer}, stencilranges::Union{Integer,NTuple{N,Integer} where N})
+# optranges=(A=(1:1, 0:2, 1:1), B=(1:1, 0:2, 1:1))
+#function add_loopopt(body::Expr, indices::Array{<:Union{Expr,Symbol}}, optvars::Union{Expr,Symbol}, optdim::Integer, loopsize::Union{Expr,Symbol,Integer}, optranges::Union{Integer,NTuple{N,Integer} where N})
 #TODO: see what to do with global consts as SHMEM_HALO_X,... Support later multiple vars for opt (now just A=T...)
 #TODO: add input check and errors
 #TODO: maybe gensym with macro @gensym
 # TODO: create a run time check for requirement: 
 # In order to be able to read the data into shared memory in only two statements, the number of threats must be at least half of the size of the shared memory block plus halo; thus, the total number of threads in each dimension must equal the range length, as else there would be smaller thread blocks at the boundaries (threads overlapping the range are sent home). These smaller blocks would be likely not to match the criteria for a correct reading of the data to shared memory. In summary the following requirements must be matched: @gridDim().x*@blockDim().x - $rangelength_x == 0; @gridDim().y*@blockDim().y - $rangelength_y > 0
-function loopopt(caller::Module, indices, optvars, optdim::Integer, loopsize, stencilranges, body; package::Symbol=get_package())
-    #TODO: the two following lines have to be removed later
-    if !isa(optvars, Symbol) @KeywordArgumentError("at present, only one optvar is supported.") end
-    stencilranges = nothing
-
[…]
