### 2023-02-20 ac07238
add tests for two and three arrays
