+++ test/test_parallel.jl
@@ -16,6 +16,8 @@ end
     if !AMDGPU.functional() TEST_PACKAGES = filter!(x->x≠PKG_AMDGPU, TEST_PACKAGES) end
 end
 
+import ParallelStencil.@gorgeousexpand
+
 @static for package in TEST_PACKAGES  eval(:(
     @testset "$(basename(@__FILE__)) (package: $(nameof($package)))" begin
         @testset "1. parallel macros" begin
@@ -121,23 +123,23 @@ end
                         @testset "@parallel_indices <kernel> (3D, loopopt, stencilranges=0:0)" begin
                             A  = @zeros(nx, ny, nz);
                             A2 = @zeros(nx, ny, nz);
-                            copy!(A, [ix + (iy-1)*size(A2,1) + (iz-1)*size(A2,1)*size(A2,2) for ix=1:size(A2,1), iy=1:size(A2,2), iz=1:size(A2,3)].^2)
-                            @parallel_indices (ix,iy,iz) loopopt=true optvars=A stencilranges=0:0 loopsize=3 function copy_loopopt!(A2, A)
+                            copy!(A, [ix + (iy-1)*size(A,1) + (iz-1)*size(A,1)*size(A,2) for ix=1:size(A,1), iy=1:size(A,2), iz=1:size(A,3)].^3);
+                            @parallel_indices (ix,iy,iz) loopopt=true loopsize=3 function copy_loopopt!(A2, A)
                                 A2[ix,iy,iz] = A[ix,iy,iz]
                                 return
                             end
-                            @parallel loopopt=true loopsize=3 copy_loopopt!(A2, A);
+                            @parallel loopopt=true copy_loopopt!(A2, A);
                             @test all(Array(A2) .== Array(A))
                         end
                         @testset "@parallel <kernel> (3D, loopopt, stencilranges=0:0)" begin
                             A  = @zeros(nx, ny, nz);
                             A2 = @zeros(nx, ny, nz);
-                            copy!(A, [ix + (iy-1)*size(A2,1) + (iz-1)*size(A2,1)*size(A2,2) for ix=1:size(A2,1), iy=1:size(A2,2), iz=1:size(A2,3)].^2)
-                            @parallel loopopt=true optvars=A stencilranges=0:0 loopsize=3 function copy_loopopt!(A2, A)
+                            copy!(A, [ix + (iy-1)*size(A,1) + (iz-1)*size(A,1)*size(A,2) for ix=1:size(A,1), iy=1:size(A,2), iz=1:size(A,3)].^3);
+                            @parallel loopopt=true loopsize=3 function copy_loopopt!(A2, A)
                                 @all(A2) = @all(A)
                                 return
                             end
-                            @parallel loopopt=true loopsize=3 copy_loopopt!(A2, A);
+                            @parallel loopopt=true copy_loopopt!(A2, A);
                             @test all(Array(A2) .== Array(A))
                         end
                         # Add here also single difference, where with parallel we have a negative halo: stencilranges=(0:0, 0:0, 1:2)
@@ -145,56 +147,56 @@ end
                             A      = @zeros(nx, ny, nz);
                             A2     = @zeros(nx, ny, nz);
                             A2_ref = @zeros(nx, ny, nz);
[…]
