### 2023-02-20 e1cb5de
add shared memory offset handling for CPU
