### 2023-02-21 d16175d
fix shared memory offset and do minor improvements
