+++ src/kernel_language.jl
@@ -70,7 +70,7 @@ function loopopt(metadata_module::Module, caller::Module, indices::Union{Symbol,
         optvars = Tuple(keys(readonlyvars))
     else
         for A in optvars
-            if !haskey(readonlyvars, A) @IncoherentArgumentError("incoherent argument optvars in @loopopt: optimization can only be applied to arrays that are only read within the kernel (not applicable to: $A).") end
+            if !haskey(readonlyvars, A) @IncoherentArgumentError("incoherent argument optvars in loopopt: optimization can only be applied to arrays that are only read within the kernel (not applicable to: $A).") end
         end
     end
     if (package ∉ SUPPORTED_PACKAGES) @KeywordArgumentError("$ERRMSG_UNSUPPORTED_PACKAGE (obtained: $package).") end
@@ -81,7 +81,7 @@ function loopopt(metadata_module::Module, caller::Module, indices::Union{Symbol,
     fullrange  = typemin(int_type):typemax(int_type)
     fullranges = (fullrange, fullrange, fullrange)
     optranges  = isnothing(optranges) ? (A = fullranges for A in optvars) : eval_arg(caller, optranges)
-    optranges  = Dict(A => (A ∈ keys(optranges)) ? optranges.A : fullranges for A in optvars)
+    optranges  = Dict(A => (A ∈ keys(optranges)) ? getproperty(optranges, A) : fullranges for A in optvars)
     body       = eval_offsets(caller, body, indices, int_type)
     offsets    = extract_offsets(caller, body, indices, int_type, optvars, optdim)
     regqueue_heads, regqueue_tails, offset_mins, offset_maxs = define_regqueues(offsets, optranges, optvars, indices, int_type, optdim)
@@ -102,7 +102,7 @@ function loopopt(metadata_module::Module, caller::Module, indices::Union{Symbol,
         # offset_span    = offset_max .- offset_min
         # oz_span        = offset_span[3]
         # loopentry      = 1 - oz_span #TODO: make possibility to do first and last read in z dimension directly into registers without halo
-        shmem_symbols = define_shmem_symbols(oz_maxs, optvars, optdim) #TODO: if use_shmems[A] can be removed in the loops if this is defined only for arrays it effectively must use shared memory
+        shmem_symbols = define_shmem_symbols(oz_maxs, optvars, use_shmems, optdim)
         # s=shmem_symbols[A]; tx, ty, nx_l, ny_l, t_h, t_h2, tx_h, tx_h2, ty_h, ty_h2, ix_h, ix_h2, iy_h, iy_h2, A_head = s[:tx], s[:ty], s[:nx_l], s[:ny_l], s[:t_h], s[:t_h2], s[:tx_h], s[:tx_h2], s[:ty_h], s[:ty_h2], s[:ix_h], s[:ix_h2], s[:iy_h], s[:iy_h2], s[:A_head]
         # tx             = gensym_world("tx", @__MODULE__)
         # ty             = gensym_world("ty", @__MODULE__)
@@ -119,7 +119,7 @@ function loopopt(metadata_module::Module, caller::Module, indices::Union{Symbol,
         # iy_h           = gensym_world("iy_h", @__MODULE__)
         # iy_h2          = gensym_world("iy_h2", @__MODULE__)
         # A_head         = gensym_world(varname(A, (oz_max,); i="iz"), @__MODULE__)
[…]
