### 2023-02-21 ea74c35
add unit test for shared memory offset on cpu
