### 2023-02-21 90b7502
add unit tests for optimization keyword arguments
