+++ test/test_parallel.jl
@@ -142,7 +142,6 @@ import ParallelStencil.@gorgeousexpand
                             @parallel loopopt=true copy_loopopt!(A2, A);
                             @test all(Array(A2) .== Array(A))
                         end
-                        # Add here also single difference, where with parallel we have a negative halo: stencilranges=(0:0, 0:0, 1:2)
                         @testset "@parallel_indices <kernel> (3D, loopopt, stencilranges=(0:0, 0:0, -1:1); z-stencil)" begin
                             A      = @zeros(nx, ny, nz);
                             A2     = @zeros(nx, ny, nz);
@@ -464,6 +463,156 @@ import ParallelStencil.@gorgeousexpand
                             copy!(B, 2 .* [ix + (iy-1)*size(A,1) + (iz-1)*size(A,1)*size(A,2) for ix=1:size(A,1), iy=1:size(A,2), iz=1:size(A,3)].^3);
                             copy!(C, 3 .* [ix + (iy-1)*size(A,1) + (iz-1)*size(A,1)*size(A,2) for ix=1:size(A,1), iy=1:size(A,2), iz=1:size(A,3)].^3);
                             @parallel_indices (ix,iy,iz) loopopt=true loopsize=3 function higher_order_loopopt!(A2, B2, C2, A, B, C)
+                                if (ix-4>1 && ix-1<size(A2,1) && iy+2>1 && iy+2<=size(A2,2) && iz-2>=1 && iz+3<=size(A2,3))
+                                    A2[ix-1,iy+2,iz] = A[ix-1,iy+2,iz+3] - 2.0*A[ix-3,iy+2,iz] + A[ix-4,iy+2,iz-2]
+                                end
+                                if (ix-4>1 && ix+1<size(B2,1) && iy+2>1 && iy+2<=size(B2,2) && iz+1>=1 && iz+2<=size(B2,3))
+                                    B2[ix+1,iy+2,iz+1] = B[ix+1,iy+2,iz+2] - 2.0*B[ix-3,iy+2,iz+1] + B[ix-4,iy+2,iz+1]
+                                end
+                                if (ix-1>1 && ix-1<size(C2,1) && iy+2>1 && iy+2<=size(C2,2) && iz-1>=1 && iz<=size(C2,3))
+                                    C2[ix-1,iy+2,iz-1] = C[ix-1,iy+2,iz] - 2.0*C[ix-1,iy+2,iz-1] + C[ix-1,iy+2,iz-1]
+                                end
+                                return
+                            end
+                            @parallel loopopt=true higher_order_loopopt!(A2, B2, C2, A, B, C);
+                            A2_ref[5:end-1,3:end,3:end-3] .= A[5:end-1,3:end,6:end] .- 2.0.*A[3:end-3,3:end,3:end-3] .+ A[2:end-4,3:end,1:end-5];
+                            B2_ref[7:end-1,3:end,2:end-1] .= B[7:end-1,3:end,3:end] .- 2.0.*B[3:end-5,3:end,2:end-1] .+ B[2:end-6,3:end,2:end-1];
+                            C2_ref[2:end-1,3:end,1:end-1] .= C[2:end-1,3:end,2:end] .- 2.0.*C[2:end-1,3:end,1:end-1] .+ C[2:end-1,3:end,1:end-1];
+                            @test all(Array(A2) .== Array(A2_ref))
+                            @test all(Array(B2) .== Array(B2_ref))
+                            @test all(Array(C2) .== Array(C2_ref))
+                        end
[…]
