### 2023-03-14 929dc3c
add shared memory loop entry and exit
