+++ src/kernel_language.jl
@@ -78,12 +78,12 @@ function loopopt(metadata_module::Module, caller::Module, indices::Union{Symbol,
     elseif (package == PKG_AMDGPU)  int_type = INT_AMDGPU
     elseif (package == PKG_THREADS) int_type = INT_THREADS
     end
-    fullrange  = typemin(int_type):typemax(int_type)
-    fullranges = (fullrange, fullrange, fullrange)
-    optranges  = isnothing(optranges) ? (A = fullranges for A in optvars) : eval_arg(caller, optranges)
-    optranges  = Dict(A => (A ∈ keys(optranges)) ? getproperty(optranges, A) : fullranges for A in optvars)
-    body       = eval_offsets(caller, body, indices, int_type)
-    offsets    = extract_offsets(caller, body, indices, int_type, optvars, optdim)
+    fullrange             = typemin(int_type):typemax(int_type)
+    fullranges            = (fullrange, fullrange, fullrange)
+    optranges             = isnothing(optranges) ? (A = fullranges for A in optvars) : eval_arg(caller, optranges)
+    optranges             = Dict(A => (A ∈ keys(optranges)) ? getproperty(optranges, A) : fullranges for A in optvars)
+    body                  = eval_offsets(caller, body, indices, int_type)
+    offsets, offsets_by_z = extract_offsets(caller, body, indices, int_type, optvars, optdim)
     regqueue_heads, regqueue_tails, offset_mins, offset_maxs = define_regqueues(offsets, optranges, optvars, indices, int_type, optdim)
     
     
@@ -94,6 +94,7 @@ function loopopt(metadata_module::Module, caller::Module, indices::Union{Symbol,
         oz_maxs, hx1s, hy1s, hx2s, hy2s, use_shmems, offset_spans, oz_spans, loopentrys = define_helper_variables(offset_mins, offset_maxs, optvars, optdim)
         # loopentry = loopentrys[optvars[1]]
         loopstart = minimum(values(loopentrys)) #TODO:
+        loopend   = loopsize
         # oz_max, hx1, hy1, hx2, hy2, use_shmem, offset_span, oz_span, loopentry = oz_maxs[A], hx1s[A], hy1s[A], hx2s[A], hy2s[A], use_shmems[A], offset_spans[A], oz_spans[A], loopentrys[A]
         # oz_max         = offset_max[3]
         # hx1, hy1       = -1 .* offset_min[1:2]
@@ -120,6 +121,9 @@ function loopopt(metadata_module::Module, caller::Module, indices::Union{Symbol,
         # iy_h2          = gensym_world("iy_h2", @__MODULE__)
         # A_head         = gensym_world(varname(A, (oz_max,); i="iz"), @__MODULE__)
         shmem_exprs    = define_shmem_exprs(shmem_symbols, optdim)
+        @show shmem_z_ranges = define_shmem_z_ranges(offsets_by_z, use_shmems, optdim)
+        @show shmem_loopentrys = define_shmem_loopentrys(loopentrys, shmem_z_ranges, offset_mins, optdim)
+        @show shmem_loopexits  = define_shmem_loopexits(loopend, shmem_z_ranges, offset_maxs, optdim)
         ix, iy, iz     = indices
         tz_g           = THREADIDS_VARNAMES[3]
         rangelength_z  = RANGELENGTHS_VARNAMES[3]
@@ -181,11 +185,11 @@ $((:(               $reg           = 0.0                                        
     for A in optvars for regs in values(regqueue_heads[A]) for reg in values(regs)
[…]
