### 2023-03-16 759d98b
make hide communication compatible with incremental compilation
