+++ src/ParallelKernel/shared.jl
@@ -180,8 +180,11 @@ is_call(arg)        = ( isa(arg, Expr) && (arg.head == :call) )
 is_block(arg)       = ( isa(arg, Expr) && (arg.head == :block) )
 is_parallel_call(x) = isexpr(x, :macrocall) && (x.args[1] == Symbol("@parallel") || x.args[1] == :(@parallel))
 
-macro get_args(args...) return args end
-extract_args(call::Expr, macroname::Symbol) = eval(substitute(deepcopy(call), macroname, Symbol("@get_args")))
+function extract_args(call::Expr, macroname::Symbol)
+    if (call.head == :macrocall) @ModuleInternalError("argument is not a macro call.") end
+    if (call.args[1] != macroname) @ModuleInternalError("unexpected macro name.") end
+    return (call.args[3:end]...,)
+end
 
 extract_kernelcall_args(call::Expr)         = split_args(call.args[2:end]; in_kernelcall=true)
 
