### 2023-03-16 cd2d3d0
make hide communication compatible with incremental compilation
