### 2023-03-20 6d3ce51
add halo read optimization
