+++ src/kernel_language.jl
@@ -62,7 +62,7 @@ end
 #TODO: maybe gensym with macro @gensym
 # TODO: create a run time check for requirement: 
 # In order to be able to read the data into shared memory in only two statements, the number of threats must be at least half of the size of the shared memory block plus halo; thus, the total number of threads in each dimension must equal the range length, as else there would be smaller thread blocks at the boundaries (threads overlapping the range are sent home). These smaller blocks would be likely not to match the criteria for a correct reading of the data to shared memory. In summary the following requirements must be matched: @gridDim().x*@blockDim().x - $rangelength_x == 0; @gridDim().y*@blockDim().y - $rangelength_y > 0
-function loopopt(metadata_module::Module, caller::Module, indices::Union{Symbol,Expr}, optvars::Union{Expr,Symbol}, optdim::Integer, loopsize::Integer, optranges::Union{Nothing, NamedTuple{t, <:NTuple{N,NTuple{3,UnitRange}} where N} where t}, body::Expr; package::Symbol=get_package())
+function loopopt(metadata_module::Module, caller::Module, indices::Union{Symbol,Expr}, optvars::Union{Expr,Symbol}, optdim::Integer, loopsize::Integer, optranges::Union{Nothing, NamedTuple{t, <:NTuple{N,NTuple{3,UnitRange}} where N} where t}, optimize_halo_read::Bool, body::Expr; package::Symbol=get_package())
     optvars = Tuple(extract_tuple(optvars)) #TODO: make this function actually return directly a tuple rather than an array
     indices = Tuple(extract_tuple(indices))
     readonlyvars  = find_readonlyvars(body, indices)
@@ -121,9 +121,11 @@ function loopopt(metadata_module::Module, caller::Module, indices::Union{Symbol,
         # iy_h2          = gensym_world("iy_h2", @__MODULE__)
         # A_head         = gensym_world(varname(A, (oz_max,); i="iz"), @__MODULE__)
         shmem_exprs    = define_shmem_exprs(shmem_symbols, optdim)
-        @show shmem_z_ranges = define_shmem_z_ranges(offsets_by_z, use_shmems, optdim)
+        @show shmem_z_ranges   = define_shmem_z_ranges(offsets_by_z, use_shmems, optdim)
         @show shmem_loopentrys = define_shmem_loopentrys(loopentrys, shmem_z_ranges, offset_mins, optdim)
         @show shmem_loopexits  = define_shmem_loopexits(loopend, shmem_z_ranges, offset_maxs, optdim)
+        mainloopstart          = (optimize_halo_read && !isempty(shmem_loopentrys)) ? minimum(values(shmem_loopentrys)) : loopstart
+        mainloopend            = loopend # TODO: the second loop split leads to wrong results, probably due to a compiler bug. # mainloopend            = (optimize_halo_read && !isempty(shmem_loopexits) ) ? maximum(values(shmem_loopexits) ) : loopend
         ix, iy, iz     = indices
         tz_g           = THREADIDS_VARNAMES[3]
         rangelength_z  = RANGELENGTHS_VARNAMES[3]
[…]
