### 2023-03-20 5bcb0a4
add halo read optimization
