+++ src/kernel_language.jl
@@ -81,55 +81,26 @@ function loopopt(metadata_module::Module, caller::Module, indices::Union{Symbol,
     body                  = eval_offsets(caller, body, indices, int_type)
     offsets, offsets_by_z = extract_offsets(caller, body, indices, int_type, optvars, optdim)
     regqueue_heads, regqueue_tails, offset_mins, offset_maxs = define_regqueues(offsets, optranges, optvars, indices, int_type, optdim)
-    
-    
-    # A = optvars[1]
-    # regqueue_head, regqueue_tail, offset_min, offset_max = regqueue_heads[A], regqueue_tails[A], offset_mins[A], offset_maxs[A]
-    # @show optranges, regqueue_head, regqueue_tail, offset_min, offset_max
+
     if optdim == 3
         oz_maxs, hx1s, hy1s, hx2s, hy2s, use_shmems, offset_spans, oz_spans, loopentrys = define_helper_variables(offset_mins, offset_maxs, optvars, optdim)
-        # loopentry = loopentrys[optvars[1]]
-        loopstart = minimum(values(loopentrys)) #TODO:
-        loopend   = loopsize
-        # oz_max, hx1, hy1, hx2, hy2, use_shmem, offset_span, oz_span, loopentry = oz_maxs[A], hx1s[A], hy1s[A], hx2s[A], hy2s[A], use_shmems[A], offset_spans[A], oz_spans[A], loopentrys[A]
-        # oz_max         = offset_max[3]
-        # hx1, hy1       = -1 .* offset_min[1:2]
-        # hx2, hy2       = offset_max[1:2]
-        # use_shmem          = (hx1+hx2>0 || hy1+hy2>0)
-        # offset_span    = offset_max .- offset_min
-        # oz_span        = offset_span[3]
-        # loopentry      = 1 - oz_span #TODO: make possibility to do first and last read in z dimension directly into registers without halo
-        shmem_symbols = define_shmem_symbols(oz_maxs, optvars, use_shmems, optdim)
-        # s=shmem_symbols[A]; tx, ty, nx_l, ny_l, t_h, t_h2, tx_h, tx_h2, ty_h, ty_h2, ix_h, ix_h2, iy_h, iy_h2, A_head = s[:tx], s[:ty], s[:nx_l], s[:ny_l], s[:t_h], s[:t_h2], s[:tx_h], s[:tx_h2], s[:ty_h], s[:ty_h2], s[:ix_h], s[:ix_h2], s[:iy_h], s[:iy_h2], s[:A_head]
-        # tx             = gensym_world("tx", @__MODULE__)
-        # ty             = gensym_world("ty", @__MODULE__)
-        # nx_l           = gensym_world("nx_l", @__MODULE__)
-        # ny_l           = gensym_world("ny_l", @__MODULE__)
-        # t_h            = gensym_world("t_h", @__MODULE__)
-        # t_h2           = gensym_world("t_h2", @__MODULE__)
-        # tx_h           = gensym_world("tx_h", @__MODULE__)
-        # tx_h2          = gensym_world("tx_h2", @__MODULE__)
-        # ty_h           = gensym_world("ty_h", @__MODULE__)
-        # ty_h2          = gensym_world("ty_h2", @__MODULE__)
-        # ix_h           = gensym_world("ix_h", @__MODULE__)
-        # ix_h2          = gensym_world("ix_h2", @__MODULE__)
-        # iy_h           = gensym_world("iy_h", @__MODULE__)
-        # iy_h2          = gensym_world("iy_h2", @__MODULE__)
-        # A_head         = gensym_world(varname(A, (oz_max,); i="iz"), @__MODULE__)
[…]
