### 2023-03-20 fbda13d
make hide communication compatible with incremental compilation
