+++ src/ParallelKernel/shared.jl
@@ -96,8 +96,11 @@ is_call(arg)        = ( isa(arg, Expr) && (arg.head == :call) )
 is_block(arg)       = ( isa(arg, Expr) && (arg.head == :block) )
 is_parallel_call(x) = isexpr(x, :macrocall) && (x.args[1] == Symbol("@parallel") || x.args[1] == :(@parallel))
 
-macro get_args(args...) return args end
-extract_args(call::Expr, macroname::Symbol) = eval(substitute(deepcopy(call), macroname, Symbol("@get_args")))
+function extract_args(call::Expr, macroname::Symbol)
+    if (call.head != :macrocall) @ModuleInternalError("argument is not a macro call.") end
+    if (call.args[1] != macroname) @ModuleInternalError("unexpected macro name.") end
+    return (call.args[3:end]...,)
+end
 
 function split_args(args)
     posargs   = [x for x in args[1:end-1] if !(isa(x,Expr) && x.head == :(=))]
