+++ .github/workflows/ci.yml
@@ -14,7 +14,7 @@ jobs:
         version:
           - '1.6' # Minimum required Julia version (due to CUDA 3 dependency).
           - '1'   # Latest stable 1.x release of Julia
-          - 'nightly'
+          # - 'nightly'
         os:
           - ubuntu-latest
           - macOS-latest
