+++ src/kernel_language.jl
@@ -84,23 +84,24 @@ function loopopt(metadata_module::Module, caller::Module, indices::Union{Symbol,
 
     if optdim == 3
         oz_maxs, hx1s, hy1s, hx2s, hy2s, use_shmems, offset_spans, oz_spans, loopentrys = define_helper_variables(offset_mins, offset_maxs, optvars, optdim)
-        loopstart        = minimum(values(loopentrys))
-        loopend          = loopsize
-        shmem_symbols    = define_shmem_symbols(oz_maxs, optvars, use_shmems, optdim)
-        shmem_exprs      = define_shmem_exprs(shmem_symbols, optdim)
-        shmem_z_ranges   = define_shmem_z_ranges(offsets_by_z, use_shmems, optdim)
-        shmem_loopentrys = define_shmem_loopentrys(loopentrys, shmem_z_ranges, offset_mins, optdim)
-        shmem_loopexits  = define_shmem_loopexits(loopend, shmem_z_ranges, offset_maxs, optdim)
-        mainloopstart    = (optimize_halo_read && !isempty(shmem_loopentrys)) ? minimum(values(shmem_loopentrys)) : loopstart
-        mainloopend      = loopend # TODO: the second loop split leads to wrong results, probably due to a compiler bug. # mainloopend            = (optimize_halo_read && !isempty(shmem_loopexits) ) ? maximum(values(shmem_loopexits) ) : loopend
-        ix, iy, iz       = indices
-        tz_g             = THREADIDS_VARNAMES[3]
-        rangelength_z    = RANGELENGTHS_VARNAMES[3]
-        ranges           = RANGES_VARNAME
-        range_z          = :(($ranges[3])[$tz_g])
-        range_z_start    = :(($ranges[3])[1])
-        i                = gensym_world("i", @__MODULE__)
-        loopoffset       = gensym_world("loopoffset", @__MODULE__)
+        loopstart          = minimum(values(loopentrys))
+        loopend            = loopsize
+        shmem_index_groups = define_shmem_index_groups(hx1s, hy1s, hx2s, hy2s, optvars, use_shmems, optdim)
+        shmem_symbols      = define_shmem_symbols(oz_maxs, optvars, use_shmems, shmem_index_groups, optdim)
+        shmem_exprs        = define_shmem_exprs(shmem_symbols, optdim)
+        shmem_z_ranges     = define_shmem_z_ranges(offsets_by_z, use_shmems, optdim)
+        shmem_loopentrys   = define_shmem_loopentrys(loopentrys, shmem_z_ranges, offset_mins, optdim)
+        shmem_loopexits    = define_shmem_loopexits(loopend, shmem_z_ranges, offset_maxs, optdim)
+        mainloopstart      = (optimize_halo_read && !isempty(shmem_loopentrys)) ? minimum(values(shmem_loopentrys)) : loopstart
+        mainloopend        = loopend # TODO: the second loop split leads to wrong results, probably due to a compiler bug. # mainloopend            = (optimize_halo_read && !isempty(shmem_loopexits) ) ? maximum(values(shmem_loopexits) ) : loopend
+        ix, iy, iz         = indices
+        tz_g               = THREADIDS_VARNAMES[3]
+        rangelength_z      = RANGELENGTHS_VARNAMES[3]
+        ranges             = RANGES_VARNAME
[…]
