### 2023-03-24 1516fc1
fix shared memory cpu macro
