### 2023-03-24 ebc884d
enable macro name simplification
