### 2023-03-24 65db8eb
fix loop optimization initialization
