### 2023-03-24 040e68a
enable on-the-fly assignments in parallel kernels
