### 2023-03-24 0766ea0
enable on-the-fly assignments in parallel kernels
