+++ src/parallel.jl
@@ -74,7 +74,7 @@ function parallel(source::LineNumberNode, caller::Module, args::Union{Symbol,Exp
         numbertype = get_numbertype()
         ndims      = get_ndims()
         if !haskey(kwargs, :metadata_module)
-            @show get_name(kernelarg)
+            get_name(kernelarg)
             metadata_module, metadata_function = create_metadata_storage(source, caller, kernelarg)
         else
             metadata_module, metadata_function = kwargs.metadata_module, kwargs.metadata_function
@@ -100,7 +100,7 @@ function parallel_indices(source::LineNumberNode, caller::Module, args::Union{Sy
     kwargs = extract_kwargs(caller, kwargs_expr, (:loopopt, :optvars, :optdim, :loopsize, :optranges, :optimize_halo_read, :metadata_module, :metadata_function), "@parallel_indices"; eval_args=(:loopopt, :optdim, :optranges, :optimize_halo_read, :metadata_module))
     kernelarg = args[end]
     if !haskey(kwargs, :metadata_module)
-        @show get_name(kernelarg)
+        get_name(kernelarg)
         metadata_module, metadata_function = create_metadata_storage(source, caller, kernelarg)
     else
         metadata_module, metadata_function = kwargs.metadata_module, kwargs.metadata_function
@@ -138,6 +138,19 @@ function parallel_kernel(metadata_module::Module, metadata_function::Expr, calle
     body = get_body(kernel)
     body = remove_return(body)
     validate_body(body)
+    kernelargs = splitarg.(extract_kernel_args(kernel)[1])
+    argvars = (arg[1] for arg in kernelargs)
+    onthefly_vars, onthefly_exprs, write_vars, body = extract_onthefly_arrays!(body, argvars)
+    check_mask_macro(caller)
+    body = apply_masks(body, indices)
+    if length(onthefly_vars) > 0
+        body = macroexpand(caller, body)
+        onthefly_syms  = gensym_world.(onthefly_vars, (@__MODULE__,))
+        onthefly_exprs = macroexpand.((caller,), onthefly_exprs)
+        body           = insert_onthefly!(body, onthefly_vars, onthefly_syms, indices)
+        onthefly_exprs = insert_onthefly!.(onthefly_exprs, (onthefly_vars,), (onthefly_syms,), (indices,))
+        create_onthefly_macro.((caller,), onthefly_syms, onthefly_exprs, onthefly_vars, (indices,))
+    end
     if isgpu(package)
         kernel = substitute(kernel, :(Data.Array),      :(Data.DeviceArray))
         kernel = substitute(kernel, :(Data.Cell),       :(Data.DeviceCell))
@@ -166,8 +179,6 @@ function parallel_kernel(metadata_module::Module, metadata_function::Expr, calle
             @ArgumentError("$ERRMSG_UNSUPPORTED_PACKAGE (obtained: $package).")
         end
     end
-    check_mask_macro(caller)
-    body = apply_masks(body, indices)
     body = add_return(body)
     set_body!(kernel, body)
     if loopopt 
@@ -326,3 +337,61 @@ end
 
 get_kernelid(kernelname, file, line) = Symbol("$(kernelname)_$(file)_$(line)")
 get_meta_function(kernelname)        = Symbol("$(META_FUNCTION_PREFIX)$(GENSYM_SEPARATOR)$(kernelname)")
+
+
+## FUNCTIONS TO DEAL WITH ON-THE-FLY ASSIGNMENTS
+
+function extract_onthefly_arrays!(body, argvars)
+    onthefly_vars  = ()
+    onthefly_exprs = ()
+    write_vars     = ()
+    statements     = get_statements(body)
[…]
