### 2023-03-24 b853e3e
move stencil access tests to shared
