### 2023-03-24 fc592f2
add tests for loopopt in initialization
