### 2023-03-24 284fcf1
add tests for on-the-fly statements in parallel kernels
