### 2023-03-30 db258d3
add register counting
