### 2023-03-30 646fb20
add explicit shared memory and launch handling
