### 2023-04-04 125d291
add unit tests for nx, ny, nz != x .* threads
