### 2023-04-12 d39d379
add support for shared memory without halo
