### 2023-04-13 da1fb9b
fix kernel indexing
