+++ src/kernel_language.jl
@@ -142,10 +142,6 @@ $((quote
                     $ix_h2         = $ix - @threadIdx().x + $tx_h2 - $hx1    # ...
                     $iy_h          = $iy - @threadIdx().y + $ty_h  - $hy1    # ...
                     $iy_h2         = $iy - @threadIdx().y + $ty_h2 - $hy1    # ...
-                    # $ix_h          = (@blockIdx().x-1)*@blockDim().x + $tx_h  - $hx1    # NOTE: here it must be @blockDim().x, not bx
-                    # $ix_h2         = (@blockIdx().x-1)*@blockDim().x + $tx_h2 - $hx1    # ...
-                    # $iy_h          = (@blockIdx().y-1)*@blockDim().y + $ty_h  - $hy1    # ...
-                    # $iy_h2         = (@blockIdx().y-1)*@blockDim().y + $ty_h2 - $hy1    # ...
     end
     for vars in values(shmem_index_groups) for A in (vars[1],) if use_shmemindices[A] for s in (shmem_vars[A],) for (shmem_offset,  hx1, hx2, hy1, hy2,  tx, ty, nx_l, ny_l, t_h, t_h2, tx_h, tx_h2, ty_h, ty_h2, ix_h, ix_h2, iy_h, iy_h2, A_head) = ((shmem_exprs[A][:offset],  hx1s[A], hx2s[A], hy1s[A], hy2s[A],  s[:tx], s[:ty], s[:nx_l], s[:ny_l], s[:t_h], s[:t_h2], s[:tx_h], s[:tx_h2], s[:ty_h], s[:ty_h2], s[:ix_h], s[:ix_h2], s[:iy_h], s[:iy_h2], s[:A_head]),)
   )...
