### 2023-04-13 a24a9ed
fix kernel indexing
