### 2023-04-13 abd417a
fix kernel indexing
