### 2023-04-13 4177324
improve optimization variable and ranges handling
