### 2023-04-27 68e3939
change boolean array type
