### 2023-05-31 cd22a1c
ad optimization defaults
