### 2023-05-31 df8d227
add documentation about number type ommission
