### 2023-06-07 3d5a5e3
add shared memory support for AMDGPU
