### 2023-06-07 2feb70b
introduce loopdim
