### 2023-06-07 4b5bb32
fixing boolean tests to be compatible with CUDA boolean array implementation
