### 2023-06-07 4a75dc5
add AMDGPU shared memory support and partial support for 32-bit indexing
