### 2023-06-07 e37c7bf
adjust thread computation for memopt
