+++ src/parallel.jl
@@ -2,13 +2,42 @@ import .ParallelKernel: get_name, set_name, get_body, set_body!, add_return, rem
 
 const PARALLEL_DOC = """
     @parallel kernel
+    @parallel memopt=... kernel
 
 Declare the `kernel` parallel and containing stencil computations be performed with one of the submodules `ParallelStencil.FiniteDifferences{1D|2D|3D}` (or with a compatible custom module or set of macros).
+
+# Optional keyword arguments
+- `memopt::Bool=false`: whether to perform advanced stencil-specific on-chip memory optimisations. If `memopt=true` is set, then it must also be set in the corresponding kernel call(s).
 
 See also: [`@init_parallel_stencil`](@ref)
 
 --------------------------------------------------------------------------------
-$(replace(ParallelKernel.PARALLEL_DOC, "@init_parallel_kernel" => "@init_parallel_stencil"))
+    @parallel kernelcall
+    @parallel memopt=... kernelcall
+
+!!! note "Advanced"
+        @parallel ranges kernelcall
+        @parallel nblocks nthreads kernelcall
+        @parallel ranges nblocks nthreads kernelcall
+        @parallel (...) kwargs... kernelcall
+
+Declare the `kernelcall` parallel. The kernel will automatically be called as required by the package for parallelization selected with [`@init_parallel_kernel`](@ref). Synchronizes at the end of the call (if a stream is given via keyword arguments, then it synchronizes only this stream).
+
+# Arguments
+- `kernelcall`: a call to a kernel that is declared parallel.
+!!! note "Advanced optional arguments"
+    - `ranges::Tuple{UnitRange{},UnitRange{},UnitRange{}} | Tuple{UnitRange{},UnitRange{}} | Tuple{UnitRange{}} | UnitRange{}`: the ranges of indices in each dimension for which computations must be performed.
+    - `nblocks::Tuple{Integer,Integer,Integer}`: the number of blocks to be used if the package CUDA or AMDGPU was selected with [`@init_parallel_kernel`](@ref).
+    - `nthreads::Tuple{Integer,Integer,Integer}`: the number of threads to be used if the package CUDA or AMDGPU was selected with [`@init_parallel_kernel`](@ref).
+    - `kwargs...`: keyword arguments to be passed further to CUDA or AMDGPU (ignored for Threads).
+
+# Optional keyword arguments
+- `memopt::Bool=false`: whether the kernel to be launched was generated with `memopt=true` (meaning the keyword was set in the kernel declaration).
+
+!!! note "Performance note"
+    Kernel launch parameters are automatically defined with heuristics, where not defined with optional kernel arguments. For CUDA and AMDGPU, `nthreads` is typically set to (32,8,1) and `nblocks` accordingly to ensure that enough threads are launched.
+
+See also: [`@init_parallel_kernel`](@ref)
 """
 @doc PARALLEL_DOC
 macro parallel(args...) check_initialized(); checkargs_parallel(args...); esc(parallel(__source__, __module__, args...)); end
@@ -70,7 +99,7 @@ parallel_async(source::LineNumberNode, caller::Module, args::Union{Symbol,Expr}.
 function parallel(source::LineNumberNode, caller::Module, args::Union{Symbol,Expr}...; package::Symbol=get_package(), async::Bool=false)
     if is_kernel(args[end])
         posargs, kwargs_expr, kernelarg = split_parallel_args(args, is_call=false)
[…]
