### 2023-06-07 c4504b7
complete AMDGPU unit tests
