### 2023-06-08 e53d462
add AMDGPU shared memory support and partial support for 32-bit indexing
