+++ test/test_projects/Diffusion3D_CUDA/src/Diffusion3D_CUDA.jl
@@ -3,4 +3,5 @@ module Diffusion3D_CUDA
     using ParallelStencil.FiniteDifferences3D
     @init_parallel_stencil(CUDA, Float64, 3)
     include(joinpath(@__DIR__, "..", "..", "shared", "diffusion3D.jl"))
+    ParallelStencil.@reset_parallel_stencil()
 end
