### 2023-06-09 e43b628
fix shared memory integer promotion issue
