### 2023-06-09 38e1f1e
remove kernel print tests not available in AMDGPU
