### 2023-06-09 42e89a2
make incremental compilation test compatible with julia 1.9
