+++ test/test_incremental_compilation.jl
@@ -1,7 +1,4 @@
-push!(LOAD_PATH, "@stdlib")  # NOTE: this is needed to enable this test to run from the Pkg manager
-push!(LOAD_PATH, joinpath(@__DIR__, ".."))
 using Test
-using Pkg
 import ParallelStencil: SUPPORTED_PACKAGES, PKG_CUDA, PKG_AMDGPU
 TEST_PACKAGES = SUPPORTED_PACKAGES
 @static if PKG_CUDA in TEST_PACKAGES
@@ -12,14 +9,19 @@ end
     import AMDGPU
     if !AMDGPU.functional() TEST_PACKAGES = filter!(x->x≠PKG_AMDGPU, TEST_PACKAGES) end
 end
+exename = joinpath(Sys.BINDIR, Base.julia_exename())
 
 @static for package in TEST_PACKAGES  eval(:(
     @testset "$(basename(@__FILE__)) (package: $(nameof($package)))" begin
         @testset "incremental compilation" begin
-            Pkg.activate(joinpath(@__DIR__, "test_projects", "Diffusion3D_$(nameof($package))"))
-            Pkg.instantiate()
-            using $(Symbol("Diffusion3D_$package"))
-            @test $(Symbol("Diffusion3D_$package")).diffusion3D()
+            test_script = joinpath(@__DIR__, "test_projects", "Diffusion3D_$(nameof($package))", "src", "test_diffusion.jl")
+            was_success = true
+            try
+                run(`$exename -O3 --startup-file=no --check-bounds=no $test_script`)
+            catch ex
+                was_success = false
+            end
+            @test was_success
         end;
     end;
 )) end == nothing || true;
