+++ test/test_projects/Diffusion3D_Threads/src/Diffusion3D_Threads.jl
@@ -3,5 +3,4 @@ module Diffusion3D_Threads
     using ParallelStencil.FiniteDifferences3D
     @init_parallel_stencil(Threads, Float64, 3)
     include(joinpath(@__DIR__, "..", "..", "shared", "diffusion3D.jl"))
-    ParallelStencil.@reset_parallel_stencil()
 end
