+++ test/test_projects/Diffusion3D_CUDA/src/test_diffusion.jl
@@ -1,0 +1,8 @@
+push!(LOAD_PATH, "@stdlib")  # NOTE: this is needed to enable this test to run from the Pkg manager
+push!(LOAD_PATH, joinpath(@__DIR__, ".."))
+using Test
+using Pkg
+Pkg.activate(joinpath(@__DIR__, "test_projects", "Diffusion3D_CUDA"))
+Pkg.instantiate()
+using Diffusion3D_CUDA
+@test Diffusion3D_CUDA.diffusion3D()
