+++ test/test_projects/Diffusion3D_AMDGPU/src/Diffusion3D_AMDGPU.jl
@@ -2,6 +2,5 @@ module Diffusion3D_AMDGPU
     using ParallelStencil
     using ParallelStencil.FiniteDifferences3D
     @init_parallel_stencil(AMDGPU, Float64, 3)
-    include(joinpath(@__DIR__, "..", "..", "shared", "diffusion3D.jl"))
-    ParallelStencil.@reset_parallel_stencil()
+    include(joinpath(@__DIR__, "..", "..", "shared", "diffusion3D.jl"))    
 end
