### 2023-06-09 ca10e62
make incremental compilation test compatible with julia 1.9
