+++ .github/workflows/ci.yml
@@ -13,8 +13,8 @@ jobs:
       matrix:
         version:
           - '1.7' # Minimum required Julia version (due to CellArrays' AMDGPU dependency).
-          #- '1'   # Latest stable 1.x release of Julia
-          #- 'nightly'
+          - '1'   # Latest stable 1.x release of Julia
+          - 'nightly'
         os:
           - ubuntu-latest
           - macOS-latest
