### 2023-06-09 0469330
make incremental compilation test compatible with julia 1.9
