### 2023-06-12 d88be9b
remove the nightly CI failing due to AMDGPU not compatible with LLVM
