+++ .github/workflows/ci.yml
@@ -14,7 +14,7 @@ jobs:
         version:
           - '1.7' # Minimum required Julia version (due to CellArrays' AMDGPU dependency).
           - '1'   # Latest stable 1.x release of Julia
-          - 'nightly'
+          # - 'nightly'
         os:
           - ubuntu-latest
           - macOS-latest
