### 2023-06-12 0e81db4
add documentation on CellArrays and memopt
