+++ README.md
@@ -13,12 +13,13 @@ ParallelStencil relies on the native kernel programming capabilities of [CUDA.jl
 A particularity of ParallelStencil is that it enables writing a single high-level Julia code that can be deployed both on a CPU or a GPU. In conjuction with [ImplicitGlobalGrid.jl] the same Julia code can even run on a single CPU thread or on thousands of GPUs/CPUs.
 
 ## Contents
-* [Parallelization with one macro call](#parallelization-with-one-macro-call)
+* [Parallelization and optimization with one macro call](#parallelization-with-one-macro-call)
 * [Stencil computations with math-close notation](#stencil-computations-with-math-close-notation)
 * [50-lines example deployable on GPU and CPU](#50-lines-example-deployable-on-GPU-and-CPU)
 * [50-lines multi-xPU example](#50-lines-multi-xpu-example)
 * [Seamless interoperability with communication packages and hiding communication](#seamless-interoperability-with-communication-packages-and-hiding-communication)
 * [Support for architecture-agnostic low level kernel programming](#support-for-architecture-agnostic-low-level-kernel-programming)
+* [Support for logical arrays of small arrays / structs](#support-for-logical-arrays-of-small-arrays--structs)
 * [Module documentation callable from the Julia REPL / IJulia](#module-documentation-callable-from-the-julia-repl--ijulia)
 * [Concise single/multi-xPU miniapps](#concise-singlemulti-xpu-miniapps)
 * [Dependencies](#dependencies)
@@ -26,11 +27,11 @@ A particularity of ParallelStencil is that it enables writing a single high-leve
 * [Questions, comments and discussions](#questions-comments-and-discussions)
 * [References](#references)
 
-## Parallelization with one macro call
-A simple call to `@parallel` is enough to parallelize a function and to launch it. The package used underneath for parallelization is defined in a call to `@init_parallel_stencil` beforehand. Supported are [CUDA.jl] for running on GPU and [Base.Threads] for CPU. The following example outlines how to run parallel computations on a GPU using the native kernel programming capabilities of [CUDA.jl] underneath (omitted lines are represented with `#(...)`, omitted arguments with `...`):
+## Parallelization and optimization with one macro call
+A simple call to `@parallel` is enough to parallelize and optimize a function and to launch it. The package used underneath for parallelization is defined in a call to `@init_parallel_stencil` beforehand. Supported are [CUDA.jl] for running on GPU and [Base.Threads] for CPU. The following example outlines how to run parallel computations on a GPU using the native kernel programming capabilities of [CUDA.jl] underneath (omitted lines are represented with `#(...)`, omitted arguments with `...`):
 ```julia
 #(...)
-@init_parallel_stencil(CUDA,...);
+@init_parallel_stencil(CUDA,...)
 #(...)
 @parallel function diffusion3D_step!(...)
     #(...)
@@ -38,10 +39,21 @@ end
 #(...)
 @parallel diffusion3D_step!(...)
 ```
+Automatic advanced fast memory usage optimization (of shared memory and registers) can be activated with the keyword argument `memopt=true`:
+```julia
+@parallel memopt=true function diffusion3D_step!(...)
+    #(...)
+end
+#(...)
+@parallel memopt=true diffusion3D_step!(...)
+```
 Note that arrays are automatically allocated on the hardware chosen for the computations (GPU or CPU) when using the provided allocation macros:
 - `@zeros`
 - `@ones`
 - `@rand`
+- `@falses`
+- `@trues`
+- `@fill`
 
 ## Stencil computations with math-close notation
[…]
