### 2023-06-13 29c5aec
fix AMDGPU shared memory allocation
