### 2023-06-13 f34e690
to work around for AMDGPU random allocation of Enums
