### 2023-06-14 d5a55d7
add error message for not supported memopt cases
