+++ test/ParallelKernel/test_parallel.jl
@@ -7,11 +7,11 @@ import ParallelStencil.ParallelKernel: checkargs_parallel, checkargs_parallel_in
 using ParallelStencil.ParallelKernel.Exceptions
 TEST_PACKAGES = SUPPORTED_PACKAGES
 @static if PKG_CUDA in TEST_PACKAGES
-    import CUDA
+    import ParallelStencil.ParallelKernel.CUDA
     if !CUDA.functional() TEST_PACKAGES = filter!(x->x≠PKG_CUDA, TEST_PACKAGES) end
 end
 @static if PKG_AMDGPU in TEST_PACKAGES
-    import AMDGPU
+    import ParallelStencil.ParallelKernel.AMDGPU
     if !AMDGPU.functional() TEST_PACKAGES = filter!(x->x≠PKG_AMDGPU, TEST_PACKAGES) end
 end
 
