### 2023-06-14 765795e
add documentation for memopt, CellArrays and AMDGPU
