### 2023-06-14 0434f1a
add documentation for memopt, CellArrays and AMDGPU
