### 2023-06-14 e66b758
add documentation for memopt, CellArrays and AMDGPU
