+++ Project.toml
@@ -5,16 +5,17 @@ version = "0.7.1"
 
 [deps]
 AMDGPU = "21141c5a-9bdb-4563-92ae-f87d6854732e"
-CellArrays = "d35fcfd7-7af4-4c67-b1aa-d78070614af4"
 CUDA = "052768ef-5323-5732-b1bb-66c8b64840ba"
+CellArrays = "d35fcfd7-7af4-4c67-b1aa-d78070614af4"
+Enzyme = "7da242da-08ed-463a-9acd-ee780be4f1d9"
 MacroTools = "1914dd2f-81c6-5fcd-8719-6d5c9610ff09"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 
 [compat]
 AMDGPU = "0.4.14"
-CellArrays = "0.1"
 CUDA = "3.12, 4"
+CellArrays = "0.1"
 MacroTools = "0.5"
 StaticArrays = "1"
 julia = "1.7"
