+++ Project.toml
@@ -16,6 +16,7 @@ StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 AMDGPU = "0.4.14"
 CUDA = "3.12, 4"
 CellArrays = "0.1"
+Enzyme = "0.11"
 MacroTools = "0.5"
 StaticArrays = "1"
 julia = "1.7"
