### 2023-07-07 0298e06
add ranges macros
