### 2023-07-13 9131112
extend @parallel doc strings
