### 2023-07-13 8bb9b24
add keyword argument recognition for AD in PS parallel function
