### 2023-07-13 f9ce077
fix async flag usage in AD
