### 2023-07-13 bd27856
extend @parallel doc strings
