### 2023-07-13 a906312
extend @parallel doc strings
