### 2023-07-13 f635ab4
add documentation on automatic differentiation
