### 2023-07-13 ad36f3e
add documentation on automatic differentiation
