### 2023-07-13 c8f2edb
add documentation on automatic differentiation
