### 2023-07-14 463c0a5
add documentation on automatic differentiation
