### 2023-07-14 e3a0412
add documentation on automatic differentiation
