### 2023-07-14 cb4650c
add error message for invalidate ad keyword if not AD triggered
