### 2023-07-14 fc3c03f
add error message for invalidate ad keyword if not AD triggered
