### 2023-07-14 cac3736
add unit tests for keyword configcall
