### 2023-07-14 55feed6
add AD low-level unit test
