### 2023-07-14 59fe0ed
deactivate low-level unit test for AD on GPU as failing with pkg manager
