### 2023-07-14 9f48667
deactivate low-level unit test for AD on GPU as failing with pkg manager
