### 2023-07-14 970270a
deactivate low-level unit test for AD on GPU as failing with pkg manager
