### 2023-07-14 1eab49e
adjust julia compatibility due to Enzyme and add TOML test dep
