+++ Project.toml
@@ -19,10 +19,11 @@ CellArrays = "0.1"
 Enzyme = "0.11"
 MacroTools = "0.5"
 StaticArrays = "1"
-julia = "1.7"
+julia = "1.8"
 
 [extras]
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+TOML = "fa267f1f-6049-4f14-aa54-33bafae1ed76"
 
 [targets]
 test = ["Test"]
