### 2023-07-18 2346c48
set Enzyme runtime activity to true for threads by default
