### 2023-07-18 bab4602
set Enzyme runtime activity to true for threads by default
