### 2023-07-18 7ef282b
set Enzyme runtime activity to true for threads by default
