+++ src/ParallelKernel/AD.jl
@@ -10,6 +10,9 @@ Provides GPU-compatible wrappers for automatic differentiation functions of the 
 - `autodiff_deferred!`: wraps function `autodiff_deferred`.
 - `autodiff_deferred_thunk!`: wraps function `autodiff_deferred_thunk`.
 
+!!! note "Enzyme runtime activity default"
+    If ParallelKernel is initialized with Threads, then `Enzyme.API.runtimeActivity!(true)` is called at module load time to ensure correct behavior of Enzyme. If you want to disable this behavior, then call `Enzyme.API.runtimeActivity!(false)` after loading ParallelStencil.
+
 To see a description of a function type `?<functionname>`.
 """
 module AD
