### 2023-07-19 d2d9b54
restore Julia 1.7 compatibility
