+++ Project.toml
@@ -19,7 +19,7 @@ CellArrays = "0.1"
 Enzyme = "0.11"
 MacroTools = "0.5"
 StaticArrays = "1"
-julia = "1.8"
+julia = "1.7"
 
 [extras]
 TOML = "fa267f1f-6049-4f14-aa54-33bafae1ed76"
