### 2023-07-20 d0dae39
Adapt to transition to HIP backend in AMDGPU
