+++ src/ParallelKernel/kernel_language.jl
@@ -170,7 +170,7 @@ end
 
 function pk_println(args...; package::Symbol=get_package())
     if     (package == PKG_CUDA)    return :(CUDA.@cuprintln($(args...)))
-    elseif (package == PKG_AMDGPU)  @KeywordArgumentError("this functionality is not yet supported in AMDGPU.jl.")
+    elseif (package == PKG_AMDGPU)  return :(AMDGPU.@rocprintln($(args...)))
     elseif (package == PKG_THREADS) return :(Base.println($(args...)))
     else                            @KeywordArgumentError("$ERRMSG_UNSUPPORTED_PACKAGE (obtained: $package).")
     end
