### 2023-07-20 e4665cb
Fix gridGroupDim name
