### 2023-07-20 ee39261
HOTFIX add device_sync for AMDGPU
