### 2023-07-20 a276878
Fixup - compare arrays on CPU
