### 2023-07-21 d730f80
Compare on CPU
