+++ test/test_FiniteDifferences3D.jl
@@ -53,15 +53,15 @@ end
                 @parallel d2_xi!(R, Axxyyzz) = (@all(R) = @d2_xi(Axxyyzz); return)
                 @parallel d2_yi!(R, Axxyyzz) = (@all(R) = @d2_yi(Axxyyzz); return)
                 @parallel d2_zi!(R, Axxyyzz) = (@all(R) = @d2_zi(Axxyyzz); return)
-                R.=0; @parallel  d_xa!(R, Ax);       @test all(R .== Ax[2:end,    :,    :].-Ax[1:end-1,      :,      :])
-                R.=0; @parallel  d_ya!(R, Ay);       @test all(R .== Ay[    :,2:end,    :].-Ay[      :,1:end-1,      :])
-                R.=0; @parallel  d_za!(R, Az);       @test all(R .== Az[    :,    :,2:end].-Az[      :,      :,1:end-1])
-                R.=0; @parallel  d_xi!(R, Axyyzz);   @test all(R .== Axyyzz[2:end  ,2:end-1,2:end-1].-Axyyzz[1:end-1,2:end-1,2:end-1])
-                R.=0; @parallel  d_yi!(R, Axxyzz);   @test all(R .== Axxyzz[2:end-1,2:end  ,2:end-1].-Axxyzz[2:end-1,1:end-1,2:end-1])
-                R.=0; @parallel  d_zi!(R, Axxyyz);   @test all(R .== Axxyyz[2:end-1,2:end-1,2:end  ].-Axxyyz[2:end-1,2:end-1,1:end-1])
-                R.=0; @parallel d2_xi!(R, Axxyyzz);  @test all(R .== (Axxyyzz[3:end,2:end-1,2:end-1].-Axxyyzz[2:end-1,2:end-1,2:end-1]).-(Axxyyzz[2:end-1,2:end-1,2:end-1].-Axxyyzz[1:end-2,2:end-1,2:end-1]))
-                R.=0; @parallel d2_yi!(R, Axxyyzz);  @test all(R .== (Axxyyzz[2:end-1,3:end,2:end-1].-Axxyyzz[2:end-1,2:end-1,2:end-1]).-(Axxyyzz[2:end-1,2:end-1,2:end-1].-Axxyyzz[2:end-1,1:end-2,2:end-1]))
-                R.=0; @parallel d2_zi!(R, Axxyyzz);  @test all(R .== (Axxyyzz[2:end-1,2:end-1,3:end].-Axxyyzz[2:end-1,2:end-1,2:end-1]).-(Axxyyzz[2:end-1,2:end-1,2:end-1].-Axxyyzz[2:end-1,2:end-1,1:end-2]))
+                R.=0; @parallel  d_xa!(R, Ax);       @test all(Array(R .== Ax[2:end,    :,    :].-Ax[1:end-1,      :,      :]))
+                R.=0; @parallel  d_ya!(R, Ay);       @test all(Array(R .== Ay[    :,2:end,    :].-Ay[      :,1:end-1,      :]))
+                R.=0; @parallel  d_za!(R, Az);       @test all(Array(R .== Az[    :,    :,2:end].-Az[      :,      :,1:end-1]))
+                R.=0; @parallel  d_xi!(R, Axyyzz);   @test all(Array(R .== Axyyzz[2:end  ,2:end-1,2:end-1].-Axyyzz[1:end-1,2:end-1,2:end-1]))
+                R.=0; @parallel  d_yi!(R, Axxyzz);   @test all(Array(R .== Axxyzz[2:end-1,2:end  ,2:end-1].-Axxyzz[2:end-1,1:end-1,2:end-1]))
[…]
