### 2023-07-22 a732f2b
Add new GPU figure
To be presented at JuliaCon 23
