+++ paper/data2/out_diffusion3D_clean_benchm_explicit_cuda.jl_sustained_1_V100.txt
@@ -1,0 +1,20 @@
+1 512 512 512 90 0.5845770835876465 3.221225472 0.0064953009287516275 495.93167542725496
+1 512 512 512 90 0.584571123123169 3.221225472 0.006495234701368544 495.9367320970386
+1 512 512 512 90 0.5845739841461182 3.221225472 0.006495266490512424 495.93430488267336
+1 512 512 512 90 0.5844860076904297 3.221225472 0.0064942889743381076 496.00895259335215
+1 512 512 512 90 0.5846059322357178 3.221225472 0.006495621469285753 495.9072026027712
+1 512 512 512 90 0.5846347808837891 3.221225472 0.006495942009819879 495.88273219349736
+1 512 512 512 90 0.5845370292663574 3.221225472 0.006494855880737305 495.96565823017494
+1 512 512 512 90 0.5845851898193359 3.221225472 0.006495390997992622 495.9247985218301
+1 512 512 512 90 0.5845210552215576 3.221225472 0.0064946783913506405 495.97921219469504
+1 512 512 512 90 0.5844588279724121 3.221225472 0.006493986977471246 496.03201903160317
+1 512 512 512 90 0.5845499038696289 3.221225472 0.006494998931884766 495.95473467849234
+1 512 512 512 90 0.5846281051635742 3.221225472 0.006495867835150825 495.8883945527823
+1 512 512 512 90 0.5845239162445068 3.221225472 0.00649471018049452 495.9767845645007
+1 512 512 512 90 0.5845770835876465 3.221225472 0.0064953009287516275 495.93167542725496
+1 512 512 512 90 0.5845630168914795 3.221225472 0.0064951446321275495 495.9436093334315
+1 512 512 512 90 0.5845351219177246 3.221225472 0.006494834687974718 495.96727657505227
+1 512 512 512 90 0.5846359729766846 3.221225472 0.006495955255296495 495.88172107151826
+1 512 512 512 90 0.5845880508422852 3.221225472 0.006495422787136502 495.9223714242738
+1 512 512 512 90 0.5846710205078125 3.221225472 0.006496344672309028 495.8519959278983
+1 512 512 512 90 0.5845420360565186 3.221225472 0.0064949115117390955 495.9614101251205
