+++ paper/data/out_diffusion3D_clean_benchm_gpuarrays_5eqs.jl_sustained_1_V100.txt
@@ -1,0 +1,20 @@
+1 512 512 512 90 4.397403001785278 3.221225472 0.04886003335316976 65.92761508606348
+1 512 512 512 90 4.397604942321777 3.221225472 0.04886227713690864 65.9245876522364
+1 512 512 512 90 4.396752119064331 3.221225472 0.04885280132293701 65.93737482332654
+1 512 512 512 90 4.396445035934448 3.221225472 0.04884938928816054 65.94198042063788
+1 512 512 512 90 4.3966169357299805 3.221225472 0.048851299285888675 65.93940220809014
+1 512 512 512 90 4.401278972625732 3.221225472 0.04890309969584147 65.86955616381758
+1 512 512 512 90 4.396696090698242 3.221225472 0.04885217878553602 65.93821508230722
+1 512 512 512 90 4.397148132324219 3.221225472 0.0488572014702691 65.93143641189111
+1 512 512 512 90 4.396762132644653 3.221225472 0.048852912584940594 65.93722465163674
+1 512 512 512 90 4.397141933441162 3.221225472 0.04885713259379069 65.93152935891676
+1 512 512 512 90 4.399523973464966 3.221225472 0.048883599705166286 65.89583196467349
+1 512 512 512 90 4.398624897003174 3.221225472 0.04887360996670193 65.90930103576659
+1 512 512 512 90 4.396697044372559 3.221225472 0.048852189381917314 65.93820077984753
+1 512 512 512 90 4.3993988037109375 3.221225472 0.048882208930121525 65.8977068038155
+1 512 512 512 90 4.397149085998535 3.221225472 0.04885721206665039 65.93142211237196
+1 512 512 512 90 4.40133810043335 3.221225472 0.04890375667148166 65.86867126873435
+1 512 512 512 90 4.39637017250061 3.221225472 0.04884855747222901 65.94310331131693
+1 512 512 512 90 4.397564888000488 3.221225472 0.048861832088894316 65.92518811287357
+1 512 512 512 90 4.395910024642944 3.221225472 0.048843444718254934 65.95000599529965
+1 512 512 512 90 4.396588087081909 3.221225472 0.048850978745354545 65.93983487600687
