+++ paper/data2/out_diffusion3D_clean_benchm_5eqs.jl_benchmarktools_1_P100.txt
@@ -1,0 +1,20 @@
+1 512 512 512 1 0.00665944 3.221225472 0.00665944 483.7081604459234
+1 512 512 512 1 0.006664362 3.221225472 0.006664362 483.35091521138855
+1 512 512 512 1 0.006679108 3.221225472 0.006679108 482.28378280453023
+1 512 512 512 1 0.00666383 3.221225472 0.00666383 483.3895030335408
+1 512 512 512 1 0.006678198 3.221225472 0.006678198 482.34950086834806
+1 512 512 512 1 0.006686697 3.221225472 0.006686697 481.7364196403695
+1 512 512 512 1 0.006696313 3.221225472 0.006696313 481.0446393410822
+1 512 512 512 1 0.006691248 3.221225472 0.006691248 481.408770381848
+1 512 512 512 1 0.006692653 3.221225472 0.006692653 481.307707421855
+1 512 512 512 1 0.006670923 3.221225472 0.006670923 482.875528918562
+1 512 512 512 1 0.006677394 3.221225472 0.006677394 482.40757876500925
+1 512 512 512 1 0.006678594 3.221225472 0.006678594 482.32090047695664
+1 512 512 512 1 0.006678926 3.221225472 0.006678926 482.2969249846457
+1 512 512 512 1 0.00669054 3.221225472 0.00669054 481.45971356572113
+1 512 512 512 1 0.006697849 3.221225472 0.006697849 480.93432264597186
+1 512 512 512 1 0.006669891 3.221225472 0.006669891 482.9502419154976
+1 512 512 512 1 0.006672377 3.221225472 0.006672377 482.7703038961977
+1 512 512 512 1 0.006704336 3.221225472 0.006704336 480.4689788817267
+1 512 512 512 1 0.006652313 3.221225472 0.006652313 484.2263844169689
+1 512 512 512 1 0.006674679 3.221225472 0.006674679 482.60380341886105

[…]
