### 2023-07-22 b656841
Fix CI Julia matrix
