+++ .github/workflows/ci.yml
@@ -12,9 +12,9 @@ jobs:
       fail-fast: false
       matrix:
         version:
-          - '1.8' # Minimum required Julia version (due to CellArrays' AMDGPU dependency 1.7 and due to Enzyme 1.8).
+          # - '1.8' # Minimum required Julia version (due to CellArrays' AMDGPU dependency 1.7 and due to Enzyme 1.8).
           - '1'   # Latest stable 1.x release of Julia
-          # - 'nightly'
+          - 'nightly'
         os:
           - ubuntu-latest
           - macOS-latest
