### 2023-07-29 13f95ea
Modify AMDGPU threads heurisitics
