+++ src/init_parallel_stencil.jl
@@ -89,7 +89,7 @@ let
     function check_already_initialized(package::Symbol, numbertype::DataType, ndims::Integer, memopt::Bool)
         if is_initialized()
             if package==get_package() && numbertype==get_numbertype() && ndims==get_ndims() && memopt==get_memopt()
-                @warn "ParallelStencil has already been initialized, with the same arguments. If you are using ParallelStencil interactively in the REPL, then you can ignore this message. If you are using ParallelStencil non-interactively, then you are likely using ParallelStencil in an inconsistent way: @init_parallel_stencil should only be called once, right after 'using ParallelStencil'."
+                @info "ParallelStencil has already been initialized, with the same arguments."
             else
                 @IncoherentCallError("ParallelStencil has already been initialized, with different arguments. If you are using ParallelStencil interactively in the REPL and want to avoid restarting Julia, then you can call ParallelStencil.@reset_parallel_stencil() and rerun all parts of your code that use ParallelStencil features (including kernel definitions and array allocations). If you are using ParallelStencil non-interactively, then you are using ParallelStencil in an invalid way: @init_parallel_stencil should only be called once, right after 'using ParallelStencil'.")
             end
