### 2023-08-03 7f54318
Update AMDGPU nthreads
