+++ src/ParallelKernel/parallel.jl
@@ -157,6 +157,11 @@ function parallel_kernel(package::Symbol, numbertype::DataType, indices::Union{S
     indices = extract_tuple(indices)
     body = get_body(kernel)
     body = remove_return(body)
+    if !all(indices .== INDICES[1:length(indices)])
+        indices_aliases = indices
+        indices = INDICES[1:length(indices)]
+        body = substitute(body, indices_aliases, indices)
+    end
     if isgpu(package)
         kernel = substitute(kernel, :(Data.Array),      :(Data.DeviceArray))
         kernel = substitute(kernel, :(Data.Cell),       :(Data.DeviceCell))
