### 2023-08-25 a3fe7fd
treat parallel indices as aliases
