### 2023-08-25 92ad775
add tests for parallel indices aliases
