+++ test/ParallelKernel/test_parallel.jl
@@ -3,7 +3,7 @@ import ParallelStencil
 using ParallelStencil.ParallelKernel
 using ParallelStencil.ParallelKernel.Enzyme
 import ParallelStencil.ParallelKernel.AD
-import ParallelStencil.ParallelKernel: @reset_parallel_kernel, @is_initialized, SUPPORTED_PACKAGES, PKG_CUDA, PKG_AMDGPU, PKG_THREADS
+import ParallelStencil.ParallelKernel: @reset_parallel_kernel, @is_initialized, SUPPORTED_PACKAGES, PKG_CUDA, PKG_AMDGPU, PKG_THREADS, INDICES
 import ParallelStencil.ParallelKernel: @require, @prettystring, @gorgeousstring, @isgpu
 import ParallelStencil.ParallelKernel: checkargs_parallel, checkargs_parallel_indices, parallel_indices
 using ParallelStencil.ParallelKernel.Exceptions
@@ -16,6 +16,8 @@ end
     import ParallelStencil.ParallelKernel.AMDGPU
     if !AMDGPU.functional() TEST_PACKAGES = filter!(x->x≠PKG_AMDGPU, TEST_PACKAGES) end
 end
+macro compute(A)              esc(:($(INDICES[1]) + ($(INDICES[2])-1)*size($A,1))) end
+macro compute_with_aliases(A) esc(:(ix            + (iz           -1)*size($A,1))) end
 
 @static for package in TEST_PACKAGES  eval(:(
     @testset "$(basename(@__FILE__)) (package: $(nameof($package)))" begin
@@ -182,6 +184,24 @@ end
                     @parallel (1:size(A,1), 1:size(A,3)) write_indices!(A);
                     @test all(Array(A)[:,end,:] .== [ix + (iz-1)*size(A,1) for ix=1:size(A,1), iz=1:size(A,3)])
                 end;
+                @testset "@parallel_indices (2D in 3D with macro)" begin
+                    A  = @zeros(4, 5, 6)
+                    @parallel_indices (ix,iz) function write_indices!(A)
+                        A[ix,end,iz] = @compute(A);
+                        return
+                    end
+                    @parallel (1:size(A,1), 1:size(A,3)) write_indices!(A);
+                    @test all(Array(A)[:,end,:] .== [ix + (iz-1)*size(A,1) for ix=1:size(A,1), iz=1:size(A,3)])
+                end;
+                @testset "@parallel_indices (2D in 3D with macro with aliases)" begin
+                    A  = @zeros(4, 5, 6)
+                    @parallel_indices (ix,iz) function write_indices!(A)
+                        A[ix,end,iz] = @compute_with_aliases(A);
+                        return
+                    end
+                    @parallel (1:size(A,1), 1:size(A,3)) write_indices!(A);
+                    @test all(Array(A)[:,end,:] .== [ix + (iz-1)*size(A,1) for ix=1:size(A,1), iz=1:size(A,3)])
+                end;
             end;
             @testset "@parallel_async" begin
                 @static if @isgpu($package)
@@ -279,25 +299,25 @@ end
             @init_parallel_kernel($package, Float64)
             @require @is_initialized
             @testset "arguments @parallel" begin
-                @test_throws ArgumentError checkargs_parallel();                                                    # Error: isempty(args)
-                @test_throws ArgumentError checkargs_parallel(:(f()), :(something));                                # Error: last arg is not function call.
-                @test_throws ArgumentError checkargs_parallel(:(f()=99));                                           # Error: last arg is not function call.
+                @test_throws ArgumentError checkargs_parallel();                                                        # Error: isempty(args)
+                @test_throws ArgumentError checkargs_parallel(:(f()), :(something));                                    # Error: last arg is not function call.
[…]
