### 2023-08-25 d1a53a5
improve already initialized warning
