### 2023-08-25 cffc15b
improve already initialized warning
