### 2023-08-25 2006689
improve already initialized warning
