### 2023-08-25 a88d609
improve already initialized warning
