### 2023-08-25 9aede66
improve already initialized warning
