### 2023-08-29 793fa7c
remove data module warning if interactive
