### 2023-08-29 dd135c7
remove already initialized warning if interactive
