### 2023-08-29 b9b21bc
remove already initialized warning if interactive
