### 2023-08-29 a82c5cd
remove data module warning if interactive
