### 2023-08-29 515acd4
add AD unit tests with expr kwarg
