### 2023-08-29 efe1092
add AD unit tests with expr kwarg
