### 2023-09-04 2d30b21
enable multiple computation calls in hide communication
