+++ src/ParallelKernel/hide_communication.jl
@@ -1,6 +1,7 @@
-const ERRMSG_HIDE_COMMUNICATION = "@hide_communication: the code block must start with exactly one @parallel computation call and be followed by code to set boundary conditions and to perform communication. Type `?@hide_communication` for more information."
+const ERRMSG_HIDE_COMMUNICATION = "@hide_communication: the code block must normally start with one @parallel computation call (for exceptions, see keyword `computation_calls`) and be followed by code to set boundary conditions and to perform communication. Type `?@hide_communication` for more information."
 const ERRMSG_INVALID_STREAM = "@hide_communication: invalid keyword argument in @parallel <kernelcall>: stream. Specifying the stream is not allowed in a @parallel call inside the code block argument of @hide_communication."
-const ERRMSG_INVALID_COMP = "@hide_communication: invalid positional arguments in @parallel <kernelcall>: no positional argument is allowed in the @parallel call to perform computations inside the code block argument of @hide_communication."
+const ERRMSG_INVALID_COMP = "@hide_communication: invalid positional arguments in @parallel <kernelcall>: zero or one positional argument (ranges) is allowed in the @parallel call to perform computations inside the code block argument of @hide_communication <boundary_width> <block>."
+const ERRMSG_INVALID_COMP_2 = "@hide_communication: invalid positional arguments in @parallel <kernelcall>: no positional argument is allowed in the @parallel call to perform computations inside the code block argument of @hide_communication <ranges_outer> <ranges_inner> <block>."
 const ERRMSG_INVALID_BC_COMM = "@hide_communication: invalid positional arguments in @parallel <kernelcall>: exactly one positional argument argument (ranges) must be passed in @parallel calls to set boundary conditions and to perform communication inside the code block argument of @hide_communication."
 
 ##
@@ -8,17 +9,22 @@ const HIDE_COMMUNICATION_DOC = """
     @hide_communication boundary_width block
 
 !!! note "Advanced"
+        @hide_communication boundary_width block computation_calls=...
         @hide_communication ranges_outer ranges_inner block
+        @hide_communication ranges_outer ranges_inner block computation_calls=...
 
 Hide the communication behind the computation within the code `block`.
 
 # Arguments
 - `boundary_width::Tuple{Integer,Integer,Integer} | Tuple{Integer,Integer} | Tuple{Integer}`: width of the boundaries in each dimension. The boundaries must include (at least) all the data that is accessed in the communcation performed.
[…]
