+++ src/ParallelKernel/hide_communication.jl
@@ -235,12 +235,13 @@ promote_boundary_width(boundary_width::BOUNDARY_WIDTH_TYPE_2D)       = (boundary
 promote_boundary_width(boundary_width::BOUNDARY_WIDTH_TYPE)          = boundary_width
 promote_boundary_width(boundary_width)                               = @ArgumentError("@hide_communication: boundary_width must be a Tuple of Integer of size 1, 2 or 3 (obtained: $boundary_width; its type is: $(typeof(boundary_width))).")
 
-function get_ranges_outer(boundary_width, ranges...::RANGES_TYPE)
-    if length(ranges) > 1
-        ranges = ranges[1]
-    else
-        ranges = ranges[1]
-    end
+function determine_common_ranges(ranges::RANGES_TYPE...)
+    if (length(ranges) > 1) && !all(map(x->x==ranges[1], ranges)) @ArgumentError("@hide_communication: the ranges of the computation calls are not all equal (obtained: $ranges). Make them equal setting the ranges manually or remove incompatible computation calls.") end
+    ranges = ranges[1]
+end
+
+function get_ranges_outer(boundary_width, ranges::RANGES_TYPE...)
+    ranges = determine_common_ranges(ranges...)
     boundary_width = promote_boundary_width(boundary_width)
     validate_ranges_args(boundary_width, ranges)
     ms = length.(ranges)
@@ -269,7 +270,8 @@ function get_ranges_outer(boundary_width, ranges...::RANGES_TYPE)
     return Tuple([r for r in ranges_outer if all(length.(r) .!= 0)])
 end
 
-function get_ranges_inner(boundary_width, ranges::RANGES_TYPE)
+function get_ranges_inner(boundary_width, ranges::RANGES_TYPE...)
+    ranges = determine_common_ranges(ranges...)
     boundary_width = promote_boundary_width(boundary_width)
     validate_ranges_args(boundary_width, ranges)
     ms = length.(ranges)
