### 2023-09-04 4f5c470
add unit tests for computation_calls and explicit ranges in computations
