+++ test/ParallelKernel/test_hide_communication.jl
@@ -26,7 +26,27 @@ end
                     block = string(@gorgeousexpand(1, @hide_communication(boundary_width, begin @parallel f(A, B); communication(); end)))
                     @test occursin("ranges_outer = ParallelStencil.ParallelKernel.get_ranges_outer(boundary_width, ParallelStencil.ParallelKernel.get_ranges(A, B))", block)
                     @test occursin("ranges_inner = ParallelStencil.ParallelKernel.get_ranges_inner(boundary_width, ParallelStencil.ParallelKernel.get_ranges(A, B))", block)
-                    @test occursin("@hide_communication ranges_outer ranges_inner begin\n            @parallel f(A, B)\n            communication()\n        end\nend", block)
+                    @test occursin("@hide_communication ranges_outer ranges_inner computation_calls = 1 begin\n            @parallel f(A, B)\n            communication()\n        end\nend", block)
+
+                    block = string(@gorgeousexpand(1, @hide_communication(boundary_width, begin @parallel ranges f(A, B); communication(); end)))
+                    @test occursin("ranges_outer = ParallelStencil.ParallelKernel.get_ranges_outer(boundary_width, ranges)", block)
+                    @test occursin("ranges_inner = ParallelStencil.ParallelKernel.get_ranges_inner(boundary_width, ranges)", block)
+                    @test occursin("@hide_communication ranges_outer ranges_inner computation_calls = 1 begin\n            @parallel f(A, B)\n            communication()\n        end\nend", block)
+
+                    block = string(@gorgeousexpand(1, @hide_communication(boundary_width, computation_calls=2, begin @parallel f(A, B); @parallel g(B, C); communication(); end)))
+                    @test occursin("ranges_outer = ParallelStencil.ParallelKernel.get_ranges_outer(boundary_width, ParallelStencil.ParallelKernel.get_ranges(A, B), ParallelStencil.ParallelKernel.get_ranges(B, C))", block)
+                    @test occursin("ranges_inner = ParallelStencil.ParallelKernel.get_ranges_inner(boundary_width, ParallelStencil.ParallelKernel.get_ranges(A, B), ParallelStencil.ParallelKernel.get_ranges(B, C))", block)
+                    @test occursin("@hide_communication ranges_outer ranges_inner computation_calls = 2 begin\n            @parallel f(A, B)\n            @parallel g(B, C)\n            communication()\n        end\nend", block)
+
+                    block = string(@gorgeousexpand(1, @hide_communication(boundary_width, computation_calls=2, begin @parallel ranges1 f(A, B); @parallel g(B, C); communication(); end)))
+                    @test occursin("ranges_outer = ParallelStencil.ParallelKernel.get_ranges_outer(boundary_width, ranges1, ParallelStencil.ParallelKernel.get_ranges(B, C))", block)
+                    @test occursin("ranges_inner = ParallelStencil.ParallelKernel.get_ranges_inner(boundary_width, ranges1, ParallelStencil.ParallelKernel.get_ranges(B, C))", block)
+                    @test occursin("@hide_communication ranges_outer ranges_inner computation_calls = 2 begin\n            @parallel f(A, B)\n            @parallel g(B, C)\n            communication()\n        end\nend", block)
+
+                    block = string(@gorgeousexpand(1, @hide_communication(boundary_width, computation_calls=2, begin @parallel ranges1 f(A, B); @parallel ranges2 g(B, C); communication(); end)))
+                    @test occursin("ranges_outer = ParallelStencil.ParallelKernel.get_ranges_outer(boundary_width, ranges1, ranges2)", block)
+                    @test occursin("ranges_inner = ParallelStencil.ParallelKernel.get_ranges_inner(boundary_width, ranges1, ranges2)", block)
[…]
