### 2023-09-05 92ea00d
add support for splat index arguments
