+++ src/parallel.jl
@@ -52,7 +52,7 @@ macro parallel(args...) check_initialized(); checkargs_parallel(args...); esc(pa
 
 
 const PARALLEL_INDICES_DOC = """
-$(replace(ParallelKernel.PARALLEL_INDICES_DOC, "@init_parallel_kernel" => "@init_parallel_stencil"))
+$(replace(ParallelKernel.PARALLEL_INDICES_DOC, "@init_parallel_kernel" => "@init_parallel_stencil")) Using splat syntax for the `indices` (e.g., `@parallel_indices (I...)`) enables to generate a tuple of parallel indices (`I` in this example) of length `ndims` selected with [`@init_parallel_stencil`](@ref). This makes it possible to write kernels that are agnostic to the number of dimensions (writing, e.g., `A[I...]` to access elements of the array `A`).
 """
 @doc PARALLEL_INDICES_DOC
 macro parallel_indices(args...) check_initialized(); checkargs_parallel_indices(args...); esc(parallel_indices(__source__, __module__, args...)); end
@@ -96,6 +96,8 @@ end
 
 function checkargs_parallel_indices(args...)
     posargs, = split_args(args)
+    indices = posargs[1]
+    if (!isa(indices,Symbol) && !isa(indices.head,Symbol)) @ArgumentError("@parallel_indices: argument 'indices' must be a tuple of indices, a single index or a variable followed by the splat operator representing a tuple of indices (e.g. (ix, iy, iz) or (ix, iy) or ix or I...).") end
     ParallelKernel.checkargs_parallel_indices(posargs...)
 end
 
@@ -140,32 +142,48 @@ end
 function parallel_indices(source::LineNumberNode, caller::Module, args::Union{Symbol,Expr}...; package::Symbol=get_package())
     is_parallel_kernel = false
     numbertype = get_numbertype()
+    ndims      = get_ndims()
     posargs, kwargs_expr = split_args(args)
     kwargs = extract_kwargs(caller, kwargs_expr, (:memopt, :optvars, :loopdim, :loopsize, :optranges, :useshmemhalos, :optimize_halo_read, :metadata_module, :metadata_function), "@parallel_indices"; eval_args=(:memopt, :loopdim, :optranges, :useshmemhalos, :optimize_halo_read, :metadata_module))
     kernelarg = args[end]
-    if !haskey(kwargs, :metadata_module)
-        get_name(kernelarg)
-        metadata_module, metadata_function = create_metadata_storage(source, caller, kernelarg)
+    indices_expr = posargs[1]
+    if is_splatarg(indices_expr)
+        parallel_indices_splatarg(caller, package, ndims, posargs...; kwargs...)
     else
-        metadata_module, metadata_function = kwargs.metadata_module, kwargs.metadata_function
-    end
-    memopt = haskey(kwargs, :memopt) ? kwargs.memopt : get_memopt()
-    if memopt
-        quote
-            $(parallel_indices_memopt(metadata_module, metadata_function, is_parallel_kernel, caller, package, posargs...; kwargs...))  #TODO: the package and numbertype will have to be passed here further once supported as kwargs (currently removed from call: package, numbertype, )
-            $metadata_function
+        if !haskey(kwargs, :metadata_module)
+            get_name(kernelarg)
+            metadata_module, metadata_function = create_metadata_storage(source, caller, kernelarg)
+        else
+            metadata_module, metadata_function = kwargs.metadata_module, kwargs.metadata_function
         end
-    else
-        ParallelKernel.parallel_indices(caller, posargs...; package=package)
+        memopt = haskey(kwargs, :memopt) ? kwargs.memopt : get_memopt()
+        if memopt
+            quote
[…]
