### 2023-09-05 ccc4a8b
do splat argument substitution also in macros
