### 2023-09-06 34295e3
make in-kernel function definition returns implicit
