### 2023-09-06 5aa846d
add unit tests for nested functions
