+++ test/ParallelKernel/test_parallel.jl
@@ -203,6 +203,51 @@ macro compute_with_aliases(A) esc(:(ix            + (iz           -1)*size($A,1)
                     @parallel (1:size(A,1), 1:size(A,3)) write_indices!(A);
                     @test all(Array(A)[:,end,:] .== [ix + (iz-1)*size(A,1) for ix=1:size(A,1), iz=1:size(A,3)])
                 end;
+                @testset "function closure (long definition, array modification)" begin
+                    A  = @zeros(4, 5, 6)
+                    @parallel_indices (ix,iy,iz) function write_indices!(A)
+                        function compute_indices!(A)
+                            A[ix,iy,iz] = ix + (iy-1)*size(A,1) + (iz-1)*size(A,1)*size(A,2);
+                            return
+                        end
+                        compute_indices!(A)
+                        return
+                    end
+                    @parallel write_indices!(A);
+                    @test all(Array(A) .== [ix + (iy-1)*size(A,1) + (iz-1)*size(A,1)*size(A,2) for ix=1:size(A,1), iy=1:size(A,2), iz=1:size(A,3)])
+                end;
+                @testset "function closure (short definition, array modification)" begin
+                    A  = @zeros(4, 5, 6)
+                    @parallel_indices (ix,iy,iz) function write_indices!(A)
+                        compute_indices!(A) = (A[ix,iy,iz] = ix + (iy-1)*size(A,1) + (iz-1)*size(A,1)*size(A,2); return)
+                        compute_indices!(A)
+                        return
+                    end
+                    @parallel write_indices!(A);
+                    @test all(Array(A) .== [ix + (iy-1)*size(A,1) + (iz-1)*size(A,1)*size(A,2) for ix=1:size(A,1), iy=1:size(A,2), iz=1:size(A,3)])
+                end;
+                @testset "function closure (long definition, return value)" begin
+                    A  = @zeros(4, 5, 6)
+                    @parallel_indices (ix,iy,iz) function write_indices!(A)
+                        function compute_indices(A)
+                            return ix + (iy-1)*size(A,1) + (iz-1)*size(A,1)*size(A,2)
+                        end
+                        A[ix,iy,iz] = compute_indices(A)
+                        return
+                    end
+                    @parallel write_indices!(A);
+                    @test all(Array(A) .== [ix + (iy-1)*size(A,1) + (iz-1)*size(A,1)*size(A,2) for ix=1:size(A,1), iy=1:size(A,2), iz=1:size(A,3)])
+                end;
+                @testset "function closure (short definition, return value)" begin
+                    A  = @zeros(4, 5, 6)
+                    @parallel_indices (ix,iy,iz) function write_indices!(A)
+                        compute_indices(A) = return ix + (iy-1)*size(A,1) + (iz-1)*size(A,1)*size(A,2)
+                        A[ix,iy,iz] = compute_indices(A)
+                        return
+                    end
+                    @parallel write_indices!(A);
+                    @test all(Array(A) .== [ix + (iy-1)*size(A,1) + (iz-1)*size(A,1)*size(A,2) for ix=1:size(A,1), iy=1:size(A,2), iz=1:size(A,3)])
+                end;
             end;
             @testset "@parallel_async" begin
                 @static if @isgpu($package)
