### 2023-09-06 eb1552c
add unit tests for nested functions
