### 2023-09-12 54d3315
add global inbounds keyword
