+++ src/ParallelKernel/init_parallel_kernel.jl
@@ -5,25 +5,27 @@ Initialize the package ParallelKernel, giving access to its main functionality. 
 
 # Arguments
 - `package::Module`: the package used for parallelization (CUDA, AMDGPU or Threads).
-- `numbertype::DataType`: the type of numbers used by @zeros, @ones, @rand and @fill and in all array types of module `Data` (e.g. Float32 or Float64). It is contained in `Data.Number` after @init_parallel_stencil.
+- `numbertype::DataType`: the type of numbers used by @zeros, @ones, @rand and @fill and in all array types of module `Data` (e.g. Float32 or Float64). It is contained in `Data.Number` after @init_parallel_kernel.
+- `inbounds::Bool=false`: whether `@inbounds` should be applied to the kernels by default (overwritable in each kernel definition).
 
 See also: [`Data`](@ref)
 """
 macro init_parallel_kernel(args...)
     check_already_initialized()
     posargs, kwargs_expr = split_args(args)
-    if (length(args) > 2)            @ArgumentError("too many arguments.")
+    if (length(args) > 3)            @ArgumentError("too many arguments.")
     elseif (0 < length(posargs) < 2) @ArgumentError("there must be either two or zero positional arguments.")
     end
     kwargs = split_kwargs(kwargs_expr)
     if (length(posargs) == 2) package, numbertype_val = extract_posargs_init(__module__, posargs...)
     else                      package, numbertype_val = extract_kwargs_init(__module__, kwargs)
     end
+    inbounds_val = extract_kwargs_optional(__module__, kwargs)
     if (package == PKG_NONE) @ArgumentError("the package argument cannot be ommited.") end #TODO: this error message will disappear, once the package can be defined at runtime.
-    esc(init_parallel_kernel(__module__, package, numbertype_val))
+    esc(init_parallel_kernel(__module__, package, numbertype_val, inbounds_val))
 end
 
-function init_parallel_kernel(caller::Module, package::Symbol, numbertype::DataType; datadoc_call=:())
+function init_parallel_kernel(caller::Module, package::Symbol, numbertype::DataType, inbounds::Bool; datadoc_call=:())
     if package == PKG_CUDA
         data_module     = Data_cuda(numbertype)
         pkg_import_cmd  = :(import ParallelStencil.ParallelKernel.CUDA)
@@ -53,6 +55,7 @@ function init_parallel_kernel(caller::Module, package::Symbol, numbertype::DataT
     @eval(caller, $ad_import_cmd)
     set_package(package)
     set_numbertype(numbertype)
+    set_inbounds(inbounds)
     set_initialized(true)
     return nothing
 end
@@ -61,17 +64,21 @@ end
 macro is_initialized() is_initialized() end
 macro get_package() get_package() end
 macro get_numbertype() get_numbertype() end
+macro get_inbounds() get_inbounds() end
 let
-    global is_initialized, set_initialized, set_package, get_package, set_numbertype, get_numbertype, check_initialized, check_already_initialized
+    global is_initialized, set_initialized, set_package, get_package, set_numbertype, get_numbertype, set_inbounds, get_inbounds, check_initialized, check_already_initialized
     _is_initialized::Bool       = false
     package::Symbol             = PKG_NONE
     numbertype::DataType        = NUMBERTYPE_NONE
+    inbounds::Bool              = false
     set_initialized(flag::Bool) = (_is_initialized = flag)
     is_initialized()            = _is_initialized
     set_package(pkg::Symbol)    = (package = pkg)
     get_package()               = package
     set_numbertype(T::DataType) = (numbertype = T)
     get_numbertype()            = numbertype
+    set_inbounds(flag::Bool)    = (inbounds = flag)
[…]
