### 2023-09-12 516a142
add global inbounds keyword
