### 2023-09-12 6fd6b09
add global inbounds keyword
