+++ src/init_parallel_stencil.jl
@@ -21,7 +21,8 @@
 
 """
     @init_parallel_stencil(package, numbertype, ndims)
-    @init_parallel_stencil(package=..., ndims=...)
+    @init_parallel_stencil(package, numbertype, ndims, inbounds=...)
+    @init_parallel_stencil(package=..., ndims=..., inbounds=...)
 
 Initialize the package ParallelStencil, giving access to its main functionality. Creates a module `Data` in the module where `@init_parallel_stencil` is called from. The module `Data` contains the types as `Data.Number`, `Data.Array` and `Data.CellArray` (type `?Data` *after* calling `@init_parallel_stencil` to see the full description of the module).
 
@@ -29,33 +30,35 @@ Initialize the package ParallelStencil, giving access to its main functionality.
 - `package::Module`: the package used for parallelization (CUDA, AMDGPU or Threads).
 - `numbertype::DataType`: the type of numbers used by @zeros, @ones, @rand and @fill and in all array types of module `Data` (e.g. Float32 or Float64). It is contained in `Data.Number` after @init_parallel_stencil. The `numbertype` can be omitted if the other arguments are given as keyword arguments (in that case, the `numbertype` will have to be given explicitly when using the types provided by the module `Data`).
 - `ndims::Integer`: the number of dimensions used for the stencil computations in the kernels (1, 2 or 3).
+- `inbounds::Bool=false`: whether `@inbounds` should be applied to the kernels by default (overwritable in each kernel definition).
 
 See also: [`Data`](@ref)
 """
 macro init_parallel_stencil(args...)
     posargs, kwargs_expr = split_args(args)
-    if (length(args) > 4)            @ArgumentError("too many arguments.")
+    if (length(args) > 5)            @ArgumentError("too many arguments.")
     elseif (0 < length(posargs) < 3) @ArgumentError("there must be either three or zero positional arguments.")
     end
     kwargs = split_kwargs(kwargs_expr)
     if (length(posargs) == 3) package, numbertype_val, ndims_val = extract_posargs_init(__module__, posargs...)
     else                      package, numbertype_val, ndims_val = extract_kwargs_init(__module__, kwargs)
     end
-    memopt_val = extract_kwargs_optional(__module__, kwargs)
+    inbounds_val, memopt_val = extract_kwargs_optional(__module__, kwargs)
     if (package == PKG_NONE) @ArgumentError("the package argument cannot be ommited.") end #TODO: this error message will disappear, once the package can be defined at runtime.
     if (ndims == NDIMS_NONE) @ArgumentError("the ndims argument cannot be ommited.") end #TODO: this error message will disappear, once the ndims can be defined at runtime.
-    check_already_initialized(package, numbertype_val, ndims_val, memopt_val)
-    esc(init_parallel_stencil(__module__, package, numbertype_val, ndims_val, memopt_val))
+    check_already_initialized(package, numbertype_val, ndims_val, inbounds_val, memopt_val)
+    esc(init_parallel_stencil(__module__, package, numbertype_val, ndims_val, inbounds_val, memopt_val))
 end
 
-function init_parallel_stencil(caller::Module, package::Symbol, numbertype::DataType, ndims::Integer, memopt::Bool)
+function init_parallel_stencil(caller::Module, package::Symbol, numbertype::DataType, ndims::Integer, inbounds::Bool, memopt::Bool)
     if (numbertype == NUMBERTYPE_NONE) datadoc_call = :(@doc replace(ParallelStencil.ParallelKernel.DATA_DOC_NUMBERTYPE_NONE, "@init_parallel_kernel" => "@init_parallel_stencil") Data)
[…]
