### 2023-09-12 13e772e
improve in bounds docstring
