### 2023-09-12 904baa8
improve in bounds docstring
