+++ src/init_parallel_stencil.jl
@@ -30,7 +30,7 @@ Initialize the package ParallelStencil, giving access to its main functionality.
 - `package::Module`: the package used for parallelization (CUDA, AMDGPU or Threads).
 - `numbertype::DataType`: the type of numbers used by @zeros, @ones, @rand and @fill and in all array types of module `Data` (e.g. Float32 or Float64). It is contained in `Data.Number` after @init_parallel_stencil. The `numbertype` can be omitted if the other arguments are given as keyword arguments (in that case, the `numbertype` will have to be given explicitly when using the types provided by the module `Data`).
 - `ndims::Integer`: the number of dimensions used for the stencil computations in the kernels (1, 2 or 3).
-- `inbounds::Bool=false`: whether `@inbounds` should be applied to the kernels by default (overwritable in each kernel definition).
+- `inbounds::Bool=false`: whether to apply `@inbounds` to the kernels by default (overwritable in each kernel definition).
 
 See also: [`Data`](@ref)
 """
