### 2023-09-12 db589e5
add inbounds kwarg to parallel kernel definitions
