### 2023-09-12 0144633
add functions to add inbounds
