### 2023-09-12 e943403
add functions to add inbounds
