### 2023-09-12 1ce797c
extract kwargs in PK parallel_indices
