### 2023-09-15 41d7be9
add test for nested Data module type declarations
