### 2023-09-15 d5e32f3
add unit tests for maxsize
