### 2023-09-15 b31ef48
handle tuples and do bitstype check in maxsize
