### 2023-09-19 e83f205
add tuple type aliases in data module
