+++ src/ParallelKernel/Data.jl
@@ -172,6 +172,79 @@ function Data_threads(numbertype::DataType)
     end
 end
 
+function Data_shared(numbertype::DataType)
+    if numbertype == NUMBERTYPE_NONE
+        quote
+            NumberTuple{N_tuple, T}                                  = NTuple{N_tuple, T}
+            ArrayTuple{N_tuple, T, N}                                = NTuple{N_tuple, Array{T, N}}
+            DeviceArrayTuple{N_tuple, T, N}                          = NTuple{N_tuple, DeviceArray{T, N}}
+            CellTuple{N_tuple, T, S}                                 = NTuple{N_tuple, Cell{T, S}}
+            DeviceCellTuple{N_tuple, T, S}                           = NTuple{N_tuple, DeviceCell{T, S}}
+            CellArrayTuple{N_tuple, T_elem, N, B}                    = NTuple{N_tuple, CellArray{T_elem, N, B}}
+            DeviceCellArrayTuple{N_tuple, T_elem, N, B}              = NTuple{N_tuple, DeviceCellArray{T_elem, N, B}}
+            NamedNumberTuple{N_tuple, T, names}                      = NamedTuple{names, NumberTuple{N_tuple, T}}
+            NamedArrayTuple{N_tuple, T, N, names}                    = NamedTuple{names, ArrayTuple{N_tuple, T, N}}
+            NamedDeviceArrayTuple{N_tuple, T, N, names}              = NamedTuple{names, DeviceArrayTuple{N_tuple, T, N}}
+            NamedCellTuple{N_tuple, T, S, names}                     = NamedTuple{names, CellTuple{N_tuple, T, S}}
+            NamedDeviceCellTuple{N_tuple, T, S, names}               = NamedTuple{names, DeviceCellTuple{N_tuple, T, S}}
+            NamedCellArrayTuple{N_tuple, T_elem, N, B, names}        = NamedTuple{names, CellArrayTuple{N_tuple, T_elem, N, B}}
+            NamedDeviceCellArrayTuple{N_tuple, T_elem, N, B, names}  = NamedTuple{names, DeviceCellArrayTuple{N_tuple, T_elem, N, B}}
+            NumberCollection{N_tuple, T}                             = Union{NumberTuple{N_tuple, T}, NamedNumberTuple{N_tuple, T}}
+            ArrayCollection{N_tuple, T, N}                           = Union{ArrayTuple{N_tuple, T, N}, NamedArrayTuple{N_tuple, T, N}}
+            DeviceArrayCollection{N_tuple, T, N}                     = Union{DeviceArrayTuple{N_tuple, T, N}, NamedDeviceArrayTuple{N_tuple, T, N}}
+            CellCollection{N_tuple, T, S}                            = Union{CellTuple{N_tuple, T, S}, NamedCellTuple{N_tuple, T, S}}
+            DeviceCellCollection{N_tuple, T, S}                      = Union{DeviceCellTuple{N_tuple, T, S}, NamedDeviceCellTuple{N_tuple, T, S}}
+            CellArrayCollection{N_tuple, T_elem, N, B}               = Union{CellArrayTuple{N_tuple, T_elem, N, B}, NamedCellArrayTuple{N_tuple, T_elem, N, B}}
+            DeviceCellArrayCollection{N_tuple, T_elem, N, B}         = Union{DeviceCellArrayTuple{N_tuple, T_elem, N, B}, NamedDeviceCellArrayTuple{N_tuple, T_elem, N, B}}
+        end
+    else
+        quote
+            NumberTuple{N_tuple}                                     = NTuple{N_tuple, Number}
[…]
