### 2023-09-20 3c7c28c
add NamedTuple constructors and tuple documentation data module
