+++ src/ParallelKernel/Data.jl
@@ -24,6 +24,26 @@ Expands to `Data.CellArray{numbertype, ndims}`, where `numbertype` is the dataty
 Expands to `Union{StaticArrays.SArray{S, numbertype}, StaticArrays.FieldArray{S, numbertype}}`, where `numbertype` is the datatype selected with [`@init_parallel_kernel`](@ref).
 
 --------------------------------------------------------------------------------
+    Data.NumberTuple{N_tuple} | Data.NamedNumberTuple{N_tuple, names} | Data.NumberCollection{N_tuple}
+
+Expands to `NTuple{N_tuple, Data.Number}` | `NamedTuple{names, NTuple{N_tuple, Data.Number}}` | `Union{NTuple{N_tuple, Data.Number}, NamedTuple{names, NTuple{N_tuple, Data.Number}}}`.
+
+--------------------------------------------------------------------------------
+    Data.ArrayTuple{N_tuple, N} | Data.NamedArrayTuple{N_tuple, N, names} | Data.ArrayCollection{N_tuple, N}
+
+Expands to `NTuple{N_tuple, Data.Array{N}}` | `NamedTuple{names, NTuple{N_tuple, Data.Array{N}}}` | `Union{NTuple{N_tuple, Data.Array{N}}, NamedTuple{names, NTuple{N_tuple, Data.Array{N}}}}`.
+
+--------------------------------------------------------------------------------
+    Data.CellArrayTuple{N_tuple, N, B} | Data.NamedCellArrayTuple{N_tuple, N, B, names} | Data.CellArrayCollection{N_tuple, N, B}
+
+Expands to `NTuple{N_tuple, Data.CellArray{N, B}}` | `NamedTuple{names, NTuple{N_tuple, Data.CellArray{N, B}}}` | `Union{NTuple{N_tuple, Data.CellArray{N, B}}, NamedTuple{names, NTuple{N_tuple, Data.CellArray{N, B}}}}`.
+
+--------------------------------------------------------------------------------
+    Data.CellTuple{N_tuple, S} | Data.NamedCellTuple{N_tuple, S, names} | Data.CellCollection{N_tuple, S}
+
+Expands to `NTuple{N_tuple, Data.Cell{S}}` | `NamedTuple{names, NTuple{N_tuple, Data.Cell{S}}}` | `Union{NTuple{N_tuple, Data.Cell{S}}, NamedTuple{names, NTuple{N_tuple, Data.Cell{S}}}}`.
+
+--------------------------------------------------------------------------------
 !!! note "Advanced"
         Data.DeviceArray{ndims}
 
@@ -82,7 +102,7 @@ Expands to `Union{StaticArrays.SArray{S, numbertype}, StaticArrays.FieldArray{S,
 function Data_cuda(numbertype::DataType)
     if numbertype == NUMBERTYPE_NONE
         :(baremodule Data # NOTE: there cannot be any newline before 'module Data' or it will create a begin end block and the module creation will fail.
-            import ParallelStencil.ParallelKernel.CUDA, ParallelStencil.ParallelKernel.CellArrays, ParallelStencil.ParallelKernel.StaticArrays
+            import Base, ParallelStencil.ParallelKernel.CUDA, ParallelStencil.ParallelKernel.CellArrays, ParallelStencil.ParallelKernel.StaticArrays
             Array{T, N}                   = CUDA.CuArray{T, N}
             DeviceArray{T, N}             = CUDA.CuDeviceArray{T, N}
             Cell{T, S}                    = Union{StaticArrays.SArray{S, T}, StaticArrays.FieldArray{S, T}}
@@ -92,7 +112,7 @@ function Data_cuda(numbertype::DataType)
         end)
     else
[…]
