+++ src/ParallelKernel/Data.jl
@@ -26,22 +26,22 @@ Expands to `Union{StaticArrays.SArray{S, numbertype}, StaticArrays.FieldArray{S,
 --------------------------------------------------------------------------------
     Data.NumberTuple{N_tuple} | Data.NamedNumberTuple{N_tuple, names} | Data.NumberCollection{N_tuple}
 
-Expands to `NTuple{N_tuple, Data.Number}` | `NamedTuple{names, NTuple{N_tuple, Data.Number}}` | `Union{NTuple{N_tuple, Data.Number}, NamedTuple{names, NTuple{N_tuple, Data.Number}}}`.
+Expands to: `NTuple{N_tuple, Data.Number}` | `NamedTuple{names, NTuple{N_tuple, Data.Number}}` | `Union{Data.NumberTuple{N_tuple}, Data.NamedNumberTuple{N_tuple}}`
 
 --------------------------------------------------------------------------------
     Data.ArrayTuple{N_tuple, N} | Data.NamedArrayTuple{N_tuple, N, names} | Data.ArrayCollection{N_tuple, N}
 
-Expands to `NTuple{N_tuple, Data.Array{N}}` | `NamedTuple{names, NTuple{N_tuple, Data.Array{N}}}` | `Union{NTuple{N_tuple, Data.Array{N}}, NamedTuple{names, NTuple{N_tuple, Data.Array{N}}}}`.
+Expands to: `NTuple{N_tuple, Data.Array{N}}` | `NamedTuple{names, NTuple{N_tuple, Data.Array{N}}}` | `Union{Data.ArrayTuple{N_tuple, N}, Data.NamedArrayTuple{N_tuple, N}}`
 
 --------------------------------------------------------------------------------
     Data.CellArrayTuple{N_tuple, N, B} | Data.NamedCellArrayTuple{N_tuple, N, B, names} | Data.CellArrayCollection{N_tuple, N, B}
 
-Expands to `NTuple{N_tuple, Data.CellArray{N, B}}` | `NamedTuple{names, NTuple{N_tuple, Data.CellArray{N, B}}}` | `Union{NTuple{N_tuple, Data.CellArray{N, B}}, NamedTuple{names, NTuple{N_tuple, Data.CellArray{N, B}}}}`.
+Expands to: `NTuple{N_tuple, Data.CellArray{N, B}}` | `NamedTuple{names, NTuple{N_tuple, Data.CellArray{N, B}}}` | `Union{Data.CellArrayTuple{N_tuple, N, B}, Data.NamedCellArrayTuple{N_tuple, N, B}}`
 
 --------------------------------------------------------------------------------
     Data.CellTuple{N_tuple, S} | Data.NamedCellTuple{N_tuple, S, names} | Data.CellCollection{N_tuple, S}
 
-Expands to `NTuple{N_tuple, Data.Cell{S}}` | `NamedTuple{names, NTuple{N_tuple, Data.Cell{S}}}` | `Union{NTuple{N_tuple, Data.Cell{S}}, NamedTuple{names, NTuple{N_tuple, Data.Cell{S}}}}`.
+Expands to: `NTuple{N_tuple, Data.Cell{S}}` | `NamedTuple{names, NTuple{N_tuple, Data.Cell{S}}}` | `Union{Data.CellTuple{N_tuple, S}, Data.NamedCellTuple{N_tuple, S}}`
 
 --------------------------------------------------------------------------------
 !!! note "Advanced"
@@ -80,6 +80,26 @@ The datatype `Data.CellArray` is automatically chosen to be compatible with the 
     Data.Cell{numbertype, S}
 
[…]
