### 2023-09-21 bada460
remove device type constructors as redundant
