### 2023-09-21 7c6371d
allow complex arguments for compute macros
