### 2023-09-21 dd3dcef
allow complex arguments for compute macros
