+++ src/FiniteDifferences.jl
@@ -45,24 +45,25 @@ export @within
 @doc "`@maxloc(A)`: Compute the maximum between 2nd order adjacent elements of `A`, using a moving window of size 3." :(@maxloc)
 @doc "`@minloc(A)`: Compute the minimum between 2nd order adjacent elements of `A`, using a moving window of size 3." :(@minloc)
 
-import ..ParallelStencil: INDICES, WITHIN_DOC
+import ..ParallelStencil: INDICES, WITHIN_DOC, expandargs
 const ix = INDICES[1]
 const ixi = :($ix+1)
 
-macro      d(A::Symbol)  esc(:( $A[$ix+1] - $A[$ix] )) end
-macro     d2(A::Symbol)  esc(:( ($A[$ixi+1] - $A[$ixi])  -  ($A[$ixi] - $A[$ixi-1]) )) end
-macro    all(A::Symbol)  esc(:( $A[$ix  ] )) end
-macro    inn(A::Symbol)  esc(:( $A[$ixi ] )) end
-macro     av(A::Symbol)  esc(:(($A[$ix] + $A[$ix+1] )*0.5 )) end
-macro   harm(A::Symbol)  esc(:(1.0/(1.0/$A[$ix] + 1.0/$A[$ix+1])*2.0 )) end
-macro maxloc(A::Symbol)  esc(:( max( max($A[$ixi-1], $A[$ixi+1]), $A[$ixi] ) )) end
-macro minloc(A::Symbol)  esc(:( min( min($A[$ixi-1], $A[$ixi+1]), $A[$ixi] ) )) end
+macro      d(A)  A, = expandargs(__module__, A);  esc(:( $A[$ix+1] - $A[$ix] )) end
+macro     d2(A)  A, = expandargs(__module__, A);  esc(:( ($A[$ixi+1] - $A[$ixi])  -  ($A[$ixi] - $A[$ixi-1]) )) end
+macro    all(A)  A, = expandargs(__module__, A);  esc(:( $A[$ix  ] )) end
+macro    inn(A)  A, = expandargs(__module__, A);  esc(:( $A[$ixi ] )) end
+macro     av(A)  A, = expandargs(__module__, A);  esc(:(($A[$ix] + $A[$ix+1] )*0.5 )) end
+macro   harm(A)  A, = expandargs(__module__, A);  esc(:(1.0/(1.0/$A[$ix] + 1.0/$A[$ix+1])*2.0 )) end
+macro maxloc(A)  A, = expandargs(__module__, A);  esc(:( max( max($A[$ixi-1], $A[$ixi+1]), $A[$ixi] ) )) end
+macro minloc(A)  A, = expandargs(__module__, A);  esc(:( min( min($A[$ixi-1], $A[$ixi+1]), $A[$ixi] ) )) end
 
 @doc WITHIN_DOC
-macro within(macroname::String, A::Symbol)
+macro within(macroname::String, A)
+    A, = expandargs(__module__, A);
     if     macroname == "@all"  esc(  :($ix<=size($A,1)  )  )
     elseif macroname == "@inn"  esc(  :($ix<=size($A,1)-2)  )
[…]
