### 2023-09-21 8e14efe
allow complex arguments for compute macros
