+++ src/shared.jl
@@ -17,7 +17,7 @@ import .ParallelKernel: @require, @symbols, symbols, longnameof, @prettyexpand, 
 ## CONSTANTS
 
 const WITHIN_DOC = """
-    @within(macroname::String, A::Symbol)
+    @within(macroname::String, A)
 
 Return an expression that evaluates to `true` if the indices generated by @parallel (module ParallelStencil) point to elements in bounds of the selection of `A` by `macroname`.
 
@@ -77,6 +77,18 @@ function substitute(expr::Expr, A, m, indices::NTuple{N,<:Union{Symbol,Expr}} wh
         end
     end
 end
+
+
+## FUNCTIONS AND MACROS FOR USAGE IN CUSTOM MACRO DEFINITIONS
+
+function expandargs(caller, args...; valid_types::NTuple{N,Type}=(Symbol, Expr)) where N
+    for arg in args
+        if (typeof(arg) ∉ valid_types) @ArgumentError("argument $arg is not of type $(join(valid_types, ", ", " or ")).") end
+    end
+    args = macroexpand.((caller,), args)
+    return args
+end
+
 
 ## FUNCTIONS FOR ERROR HANDLING
 
