### 2023-09-21 f4f9b31
add tuple device type substitutions
