### 2023-09-21 547de4c
add tuple device type substitutions
