### 2023-09-21 6bd158a
add tuple device type substitutions
