### 2023-09-21 20ccbbd
add tuple device type substitutions
