### 2023-09-21 70193da
add tuple device type substitutions unit tests
