+++ test/ParallelKernel/test_parallel.jl
@@ -145,6 +145,60 @@ macro compute_with_aliases(A) esc(:(ix            + (iz           -1)*size($A,1)
                             @test occursin("f(A::Data.DeviceCellArray, B::Data.DeviceCellArray,", expansion)
                     end
                 end
+                @testset "Data.ArrayTuple to Data.DeviceArrayTuple" begin
+                    @static if @isgpu($package)
+                            expansion = @prettystring(1, @parallel_indices (ix,iy) f(A::Data.ArrayTuple, B::Data.ArrayTuple, c::T) where T <: Integer = return)
+                            @test occursin("f(A::Data.DeviceArrayTuple, B::Data.DeviceArrayTuple,", expansion)
+                    end
+                end
+                @testset "Data.CellTuple to Data.DeviceCellTuple" begin
+                    @static if @isgpu($package)
+                            expansion = @prettystring(1, @parallel_indices (ix,iy) f(A::Data.CellTuple, B::Data.CellTuple, c::T) where T <: Integer = return)
+                            @test occursin("f(A::Data.DeviceCellTuple, B::Data.DeviceCellTuple,", expansion)
+                    end
+                end
+                @testset "Data.CellArrayTuple to Data.DeviceCellArrayTuple" begin
+                    @static if @isgpu($package)
+                            expansion = @prettystring(1, @parallel_indices (ix,iy) f(A::Data.CellArrayTuple, B::Data.CellArrayTuple, c::T) where T <: Integer = return)
+                            @test occursin("f(A::Data.DeviceCellArrayTuple, B::Data.DeviceCellArrayTuple,", expansion)
+                    end
+                end
+                @testset "Data.NamedArrayTuple to Data.NamedDeviceArrayTuple" begin
+                    @static if @isgpu($package)
+                            expansion = @prettystring(1, @parallel_indices (ix,iy) f(A::Data.NamedArrayTuple, B::Data.NamedArrayTuple, c::T) where T <: Integer = return)
+                            @test occursin("f(A::Data.NamedDeviceArrayTuple, B::Data.NamedDeviceArrayTuple,", expansion)
+                    end
+                end
+                @testset "Data.NamedCellTuple to Data.NamedDeviceCellTuple" begin
+                    @static if @isgpu($package)
+                            expansion = @prettystring(1, @parallel_indices (ix,iy) f(A::Data.NamedCellTuple, B::Data.NamedCellTuple, c::T) where T <: Integer = return)
+                            @test occursin("f(A::Data.NamedDeviceCellTuple, B::Data.NamedDeviceCellTuple,", expansion)
+                    end
+                end
+                @testset "Data.NamedCellArrayTuple to Data.NamedDeviceCellArrayTuple" begin
+                    @static if @isgpu($package)
+                            expansion = @prettystring(1, @parallel_indices (ix,iy) f(A::Data.NamedCellArrayTuple, B::Data.NamedCellArrayTuple, c::T) where T <: Integer = return)
+                            @test occursin("f(A::Data.NamedDeviceCellArrayTuple, B::Data.NamedDeviceCellArrayTuple,", expansion)
+                    end
+                end
+                @testset "Data.ArrayCollection to Data.DeviceArrayCollection" begin
+                    @static if @isgpu($package)
+                            expansion = @prettystring(1, @parallel_indices (ix,iy) f!(A::Data.ArrayCollection, B::Data.ArrayCollection, c::T) where T <: Integer = (A[ix,iy] = B[ix,iy]^c; return))
+                            @test occursin("f!(A::Data.DeviceArrayCollection, B::Data.DeviceArrayCollection,", expansion)
+                    end
+                end
[…]
