### 2023-09-21 caaa4cb
implement in expression walk and symbol generation for expressions
