### 2023-09-21 5b52728
make on the fly variable detection suitable for expressions
