+++ src/ParallelKernel/Data.jl
@@ -223,13 +223,13 @@ function Data_shared(numbertype::DataType)
             CellArrayTuple{N_tuple, T_elem, N, B}                    = NTuple{N_tuple, CellArray{T_elem, N, B}}
             DeviceCellArrayTuple{N_tuple, T_elem, N, B}              = NTuple{N_tuple, DeviceCellArray{T_elem, N, B}}
 
-            NamedNumberTuple{N_tuple, T, names}                      = NamedTuple{names, NumberTuple{N_tuple, T}}
-            NamedArrayTuple{N_tuple, T, N, names}                    = NamedTuple{names, ArrayTuple{N_tuple, T, N}}
-            NamedDeviceArrayTuple{N_tuple, T, N, names}              = NamedTuple{names, DeviceArrayTuple{N_tuple, T, N}}
-            NamedCellTuple{N_tuple, T, S, names}                     = NamedTuple{names, CellTuple{N_tuple, T, S}}
-            NamedDeviceCellTuple{N_tuple, T, S, names}               = NamedTuple{names, DeviceCellTuple{N_tuple, T, S}}
-            NamedCellArrayTuple{N_tuple, T_elem, N, B, names}        = NamedTuple{names, CellArrayTuple{N_tuple, T_elem, N, B}}
-            NamedDeviceCellArrayTuple{N_tuple, T_elem, N, B, names}  = NamedTuple{names, DeviceCellArrayTuple{N_tuple, T_elem, N, B}}
+            NamedNumberTuple{N_tuple, T, names}                      = NamedTuple{names, <:NumberTuple{N_tuple, T}}
+            NamedArrayTuple{N_tuple, T, N, names}                    = NamedTuple{names, <:ArrayTuple{N_tuple, T, N}}
+            NamedDeviceArrayTuple{N_tuple, T, N, names}              = NamedTuple{names, <:DeviceArrayTuple{N_tuple, T, N}}
+            NamedCellTuple{N_tuple, T, S, names}                     = NamedTuple{names, <:CellTuple{N_tuple, T, S}}
+            NamedDeviceCellTuple{N_tuple, T, S, names}               = NamedTuple{names, <:DeviceCellTuple{N_tuple, T, S}}
+            NamedCellArrayTuple{N_tuple, T_elem, N, B, names}        = NamedTuple{names, <:CellArrayTuple{N_tuple, T_elem, N, B}}
+            NamedDeviceCellArrayTuple{N_tuple, T_elem, N, B, names}  = NamedTuple{names, <:DeviceCellArrayTuple{N_tuple, T_elem, N, B}}
 
             NumberCollection{N_tuple, T}                             = Union{NumberTuple{N_tuple, T}, NamedNumberTuple{N_tuple, T}}
             ArrayCollection{N_tuple, T, N}                           = Union{ArrayTuple{N_tuple, T, N}, NamedArrayTuple{N_tuple, T, N}}
@@ -261,20 +261,20 @@ function Data_shared(numbertype::DataType)
             TCellArrayTuple{N_tuple, T_elem, N, B}                   = NTuple{N_tuple, TCellArray{T_elem, N, B}}
             DeviceTCellArrayTuple{N_tuple, T_elem, N, B}             = NTuple{N_tuple, DeviceTCellArray{T_elem, N, B}}
 
-            NamedNumberTuple{N_tuple, names}                         = NamedTuple{names, NumberTuple{N_tuple}}
-            NamedArrayTuple{N_tuple, N, names}                       = NamedTuple{names, ArrayTuple{N_tuple, N}}
-            NamedDeviceArrayTuple{N_tuple, N, names}                 = NamedTuple{names, DeviceArrayTuple{N_tuple, N}}
[…]
