### 2023-11-06 #138 @whedon
[REVIEW]: High-performance xPU Stencil Computations in Julia
**Submitting author:** @omlins (<a href="http://orcid.org/0000-0001-8969-5619">Samuel Omlin</a>)
**Repository:** https://github.com/omlins/ParallelStencil.jl
**Branch with paper.md** (empty if default branch): JuliaConProceeding2022
**Version:** v1

- [x] **Page limit:** Is the page limit for extended abstracts respected by the submitted document?

- [x] **Page limit:** Is the page limit for extended abstracts respected by the submitted document?
