### 2023-11-13 omlins/ParallelStencil.jl#124 @benegee
[JuliaCon/proceedings-review] @parallel keyword argument `loopopt` deprecated?
Dear authors of ParallelStencil.jl!

I am just going through https://github.com/JuliaCon/proceedings-review/issues/138

The example provided in your repo, https://github.com/omlins/ParallelStencil.jl/tree/JuliaConProceeding2022#50-lines-example-deployable-on-gpu-and-cpu, works perfectly fine.

The snippet in your paper, however, does not work with the current version of PaperStencil.jl because the keyword `loopopt` is reported as invalid. I assume it is deprecated by now. Maybe you could update the snippet in the paper? Or provide the version of your package used to generate the results in the paper?
