### 2023-11-13 omlins/ParallelStencil.jl#126 @benegee
[JuliaCon/proceedings-review] DOI of paper by Besard et al.
The checklist at https://github.com/JuliaCon/proceedings-review/issues/138 tells me that archival references should have a DOI listed. Here is the DOI of the paper by Besard et al.: https://ieeexplore.ieee.org/document/8471188
