### 2023-11-13 omlins/ParallelStencil.jl#127 @benegee
[JuliaCon/proceedings-review] Community guidelines
According to the review guidelines https://github.com/JuliaCon/proceedings-review/issues/138 there should be

> clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software

As your package is hosted publicly on GitHub I am pretty sure that you would welcome any contributions or reports. Maybe you want to encourage people to do so through your README?
