### 2023-11-13 omlins/ParallelStencil.jl#126 @bgeihe
See also https://github.com/JuliaCon/proceedings-review/issues/138#issuecomment-1796282098
