### 2023-11-14 omlins/ParallelStencil.jl#124 @omlins
Thanks for highlighting this. We will update snippet in the paper: `loopopt` has been replaced with `memopt`.
