### 2023-11-14 omlins/ParallelStencil.jl#127 @omlins
We will gladly add some guidelines.
