### 2023-11-17 #138 @bgeihe
### Review

ParallelStencil.jl is a package which allows to implement efficient algorithms based on stencil computations. Its special feature is being architecture-agnostic, which means that the same code can be used for prototyping and high performance computations on either CPUs or GPUs.

The extended abstract is written clearly and concisely. Motivation is given, the approach for implementation is sketched, and numerical examples for 3D heat diffusion as well as nonlinear problems are presented. The impressive results demonstrate the claimed near optimal performance.

I have linked three issues with suggestions as to minor modifications. Once these have been addressed, I think the extended abstract should be accepted for publication.
