### 2023-11-25 #138 @georgebisbas
ParallelStencil.jl offers a high-level textbook-math-close interface and abstractions for solving PDEs for the HPC solution of PDEs. It can also support explicit stencil definition and computation. Users can benefit from seamlessly transitioning the high-level code to target both CPUs and GPUs, with only minor to no edits in the source code.

This work has a high overlap with  https://github.com/JuliaCon/proceedings-review/issues/137
I propose that they be merged into a paper covering both packages. However, this is not a firm requirement. Since ParallelStencil.jl does not claim distributed-memory parallelism benefits, mentions to DMP should be removed to reduce the overlap. For example:

Abstract: …and multiple dispatch enable writing a single code that is suitable for both productive proto typing on a single CPU thread and production runs on multi-GPU or CPU workstations or supercomputers….

Another suggestion is to avoid using the term xPU in the title. In the recent rise of “exotic” accelerators like Google’s TPU, Graphcore’s IPU, or expected QPUs, it probably would not harm just to use:

High-performance Stencil Computations in Julia

And just make it clear that this work is about CPU and GPU.

The package's functionality is confirmed, and the installation, testing and use documentation is present and well-defined.

The paper is well-written, clearly presenting the motivation and example usage of the package. This work also benefits from the math-close interface, which is only sometimes provided in the stencil-specific scientific community. Projects like OpenSBLI, Devito, and Exastencils provide such interfaces, primarily not in Julia. The paper would probably benefit by slightly “upgrading” the reference to the symbolic interface, as this helps interdisciplinary scientists take advantage of HPC solvers without the need to entangle with low-level HPC optimisations.

The diffusion example presented is well-written. Minor comments (not aiming to be addressed, but possibly interesting points for future work and further discussion):
Could the use of both “dx” and “_dx” possibly be avoided? As far as I understand, it only helps to reduce the text in equations present in the parallel loop (lines 5:12) while only adding more complexity (with no comments in line 26).
Additionally, T2 in line 30 is the stencil computation's second buffer. Does this mean the user would need to add additional copies if the PDE solvers require a higher accuracy order in time? (e.g. time discretisation order in acoustic wave propagation is 2 (needing T, T2, T3?))
- Should Figure 1 be presented as a Listing 1?
- Should Figure 2 be split into two subfigures?

Performance evaluation: Linked issue in the package repository.
