+++ Project.toml
@@ -14,7 +14,7 @@ StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 
 [compat]
 AMDGPU = "0.5"
-CUDA = "3.12, 4"
+CUDA = "3.12, 4, 5"
 CellArrays = "0.1"
 Enzyme = "0.11"
 MacroTools = "0.5"
