### 2023-11-30 0bf944b
Update Project.toml
Co-authored-by: Ludovic Räss <61313342+luraess@users.noreply.github.com>
