### 2023-11-30 944cd09
Update Project.toml
Co-authored-by: Ludovic Räss <61313342+luraess@users.noreply.github.com>
