### 2023-12-07 #138 @omlins
@georgebisbas : Thank you very much for the careful review. We are working on addressing the reviews. In the following you can find replies to two of the issues that you raised. We will respond to the remaining issues alongside with the improvements of the manuscript.

> This work has a high overlap with https://github.com/JuliaCon/proceedings-review/issues/137
I propose that they be merged into a paper covering both packages. However, this is not a firm requirement. Since ParallelStencil.jl does not claim distributed-memory parallelism benefits, mentions to DMP should be removed to reduce the overlap. For example:
Abstract: …and multiple dispatch enable writing a single code that is suitable for both productive proto typing on a single CPU thread and production runs on multi-GPU or CPU workstations or supercomputers….

Thank you for pointing out the overlaps concerning distributed memory parallelism. We will remove it as much as possible in order to publish the packages ParallelStencil.jl and ImplicitGlobalGrid.jl in separate papers. We would like to keep the format of one paper per package, given that they have a different focus and can be used independently.

> Performance evaluation: Linked issue in the package repository.

We have replied to this issue directly where it was opened.
