### 2023-12-12 c593840
add extensions unit tests
