### 2023-12-12 9c28831
add extensions unit tests
