+++ test/test_projects/Diffusion3D/Project.toml
@@ -14,5 +14,5 @@ AMDGPU = "21141c5a-9bdb-4563-92ae-f87d6854732e"
 CUDA = "052768ef-5323-5732-b1bb-66c8b64840ba"
 
 [extensions]
-#AMDGPUExt = "AMDGPU"
-CUDAExt = "CUDA"
+Diffusion3D_AMDGPUExt = "AMDGPU"
+Diffusion3D_CUDAExt = "CUDA"
