### 2023-12-13 adecfdb
add minimal extensions unit test
