+++ test/test_projects/Diffusion3D_minimal/src/Diffusion3D_minimal.jl
@@ -1,0 +1,4 @@
+module Diffusion3D_minimal
+    using ParallelStencil
+    f() = return
+end
