+++ test/test_projects/Diffusion3D_minimal/ext/D3D_CUDAExt.jl
@@ -1,0 +1,4 @@
+module D3D_CUDAExt
+    import Diffusion3D_minimal
+    g() = return
+end
