+++ test/test_projects/Diffusion3D_minimal/test/localtest_diffusion_CUDA.jl
@@ -1,0 +1,8 @@
+push!(LOAD_PATH, "@stdlib")  # NOTE: this is needed to enable this test to run from the Pkg manager
+push!(LOAD_PATH, joinpath(@__DIR__, ".."))
+using Test
+using Pkg
+Pkg.activate(joinpath(@__DIR__, "test_projects", "Diffusion3D_minimal"))
+Pkg.instantiate()
+import CUDA
+using Diffusion3D_minimal
