### 2023-12-13 7515982
Add @d2_...a macros
