### 2023-12-13 71481d0
And unit tests for d2_...a macros
