### 2023-12-14 eac88ad
update local test package directory
