+++ test/test_projects/Diffusion3D_minimal/test/localtest_diffusion_CUDA.jl
@@ -2,7 +2,7 @@ push!(LOAD_PATH, "@stdlib")  # NOTE: this is needed to enable this test to run f
 push!(LOAD_PATH, joinpath(@__DIR__, ".."))
 using Test
 using Pkg
-Pkg.activate(joinpath(@__DIR__, "test_projects", "Diffusion3D_minimal"))
+Pkg.activate(joinpath(@__DIR__, ".."))
 Pkg.instantiate()
 import CUDA
 using Diffusion3D_minimal
