### 2023-12-18 0dc4395
move to per module PS parameters
