+++ src/ParallelKernel/allocators.jl
@@ -19,10 +19,10 @@ See also: [`@ones`](@ref), [`@rand`](@ref), [`@falses`](@ref), [`@trues`](@ref),
 """
 @doc ZEROS_DOC
 macro zeros(args...)
-    check_initialized()
+    check_initialized(__module__)
     posargs, kwargs_expr = split_args(args)
     eltype, celldims, celltype, blocklength = extract_kwargvalues(kwargs_expr, (:eltype, :celldims, :celltype, :blocklength), "@zeros")
-    esc(_zeros(posargs...; eltype=eltype, celldims=celldims, celltype=celltype, blocklength=blocklength))
+    esc(_zeros(__module__, posargs...; eltype=eltype, celldims=celldims, celltype=celltype, blocklength=blocklength))
 end
 
 
@@ -47,10 +47,10 @@ See also: [`@zeros`](@ref), [`@rand`](@ref), [`@falses`](@ref), [`@trues`](@ref)
 """
 @doc ONES_DOC
 macro ones(args...)
-    check_initialized()
+    check_initialized(__module__)
     posargs, kwargs_expr = split_args(args)
     eltype, celldims, celltype, blocklength = extract_kwargvalues(kwargs_expr, (:eltype, :celldims, :celltype, :blocklength), "@ones")
-    esc(_ones(posargs...; eltype=eltype, celldims=celldims, celltype=celltype, blocklength=blocklength))
+    esc(_ones(__module__, posargs...; eltype=eltype, celldims=celldims, celltype=celltype, blocklength=blocklength))
 end
 
 ##
@@ -74,10 +74,10 @@ See also: [`@zeros`](@ref), [`@ones`](@ref), [`@falses`](@ref), [`@trues`](@ref)
 """
 @doc RAND_DOC
 macro rand(args...)
-    check_initialized()
+    check_initialized(__module__)
     posargs, kwargs_expr = split_args(args)
     eltype, celldims, celltype, blocklength = extract_kwargvalues(kwargs_expr, (:eltype, :celldims, :celltype, :blocklength), "@rand")
-    esc(_rand(posargs...; eltype=eltype, celldims=celldims, celltype=celltype, blocklength=blocklength))
+    esc(_rand(__module__, posargs...; eltype=eltype, celldims=celldims, celltype=celltype, blocklength=blocklength))
 end
 
 
@@ -97,10 +97,10 @@ See also: [`@zeros`](@ref), [`@ones`](@ref), [`@rand`](@ref), [`@trues`](@ref), 
 """
 @doc FALSES_DOC
 macro falses(args...)
-    check_initialized()
+    check_initialized(__module__)
     posargs, kwargs_expr = split_args(args)
     celldims, blocklength = extract_kwargvalues(kwargs_expr, (:celldims, :blocklength), "@falses")
-    esc(_falses(posargs...; celldims=celldims, blocklength=blocklength))
+    esc(_falses(__module__, posargs...; celldims=celldims, blocklength=blocklength))
 end
 
 
@@ -120,10 +120,10 @@ See also: [`@zeros`](@ref), [`@ones`](@ref), [`@rand`](@ref), [`@falses`](@ref),
 """
 @doc TRUES_DOC
 macro trues(args...)
-    check_initialized()
+    check_initialized(__module__)
     posargs, kwargs_expr = split_args(args)
     celldims, blocklength = extract_kwargvalues(kwargs_expr, (:celldims, :blocklength), "@trues")
-    esc(_trues(posargs...; celldims=celldims, blocklength=blocklength))
+    esc(_trues(__module__, posargs...; celldims=celldims, blocklength=blocklength))
 end
 
 
[…]
