### 2023-12-18 3924c80
move to per module PS parameters
