### 2023-12-18 43fd2c0
move to per module PS parameters
