+++ src/ParallelKernel/kernel_language.jl
@@ -9,7 +9,7 @@ const GRIDDIM_DOC = """
 Return the grid size (or "dimension") in x, y and z dimension. The grid size in a specific dimension is commonly retrieved directly as in this example in x dimension: `@gridDim().x`.
 """
 @doc GRIDDIM_DOC
-macro gridDim(args...) check_initialized(); checknoargs(args...); esc(gridDim(args...)); end
+macro gridDim(args...) check_initialized(__module__); checknoargs(args...); esc(gridDim(__module__, args...)); end
 
 
 ##
@@ -19,7 +19,7 @@ const BLOCKIDX_DOC = """
 Return the block ID in x, y and z dimension within the grid. The block ID in a specific dimension is commonly retrieved directly as in this example in x dimension: `@blockIdx().x`.
 """
 @doc BLOCKIDX_DOC
-macro blockIdx(args...) check_initialized(); checknoargs(args...); esc(blockIdx(args...)); end
+macro blockIdx(args...) check_initialized(__module__); checknoargs(args...); esc(blockIdx(__module__, args...)); end
 
 
 ##
@@ -29,7 +29,7 @@ const BLOCKDIM_DOC = """
 Return the block size (or "dimension") in x, y and z dimension. The block size in a specific dimension is commonly retrieved directly as in this example in x dimension: `@blockDim().x`.
 """
 @doc BLOCKDIM_DOC
-macro blockDim(args...) check_initialized(); checknoargs(args...); esc(blockDim(args...)); end
+macro blockDim(args...) check_initialized(__module__); checknoargs(args...); esc(blockDim(__module__, args...)); end
 
 
 ##
@@ -39,7 +39,7 @@ const THREADIDX_DOC = """
 Return the thread ID in x, y and z dimension within the block. The thread ID in a specific dimension is commonly retrieved directly as in this example in x dimension: `@threadIdx().x`.
 """
 @doc THREADIDX_DOC
-macro threadIdx(args...) check_initialized(); checknoargs(args...); esc(threadIdx(args...)); end
+macro threadIdx(args...) check_initialized(__module__); checknoargs(args...); esc(threadIdx(__module__, args...)); end
 
 
 ##
@@ -49,7 +49,7 @@ const SYNCTHREADS_DOC = """
 Synchronize the threads of the block: wait until all threads in the block have reached this point and all global and shared memory accesses made by these threads prior to the `sync_threads()` call are visible to all threads in the block.
 """
 @doc SYNCTHREADS_DOC
-macro sync_threads(args...) check_initialized(); checknoargs(args...); esc(sync_threads(args...)); end
+macro sync_threads(args...) check_initialized(__module__); checknoargs(args...); esc(sync_threads(__module__, args...)); end
 
 
 ##
@@ -63,7 +63,7 @@ When multiple shared memory arrays are created within a kernel, then all arrays 
     The amount of shared memory needs to be specified when launching the kernel (keyword argument `shmem`).
 """
 @doc SHAREDMEM_DOC
-macro sharedMem(args...) check_initialized(); checkargs_sharedMem(args...); esc(sharedMem(args...)); end
+macro sharedMem(args...) check_initialized(__module__); checkargs_sharedMem(args...); esc(sharedMem(__module__, args...)); end
 
 
 ##
@@ -73,7 +73,7 @@ const PKSHOW_DOC = """
 Call a macro analogue to `Base.@show`, compatible with the package for parallelization selected with [`@init_parallel_kernel`](@ref) (Base.@show for Threads and CUDA.@cushow for CUDA).
 """
 @doc PKSHOW_DOC
-macro pk_show(args...) check_initialized(); esc(pk_show(args...)); end
[…]
