### 2023-12-18 3132bb1
move to per module PS parameters
