### 2023-12-18 4e52b07
move to per module PS parameters
