### 2023-12-18 0b3cb2f
move to per module PS parameters
