### 2023-12-18 2b7be8a
move to per module PS parameters
