+++ src/reset_parallel_stencil.jl
@@ -9,9 +9,9 @@ macro reset_parallel_stencil() esc(reset_parallel_stencil(__module__)) end
 
 function reset_parallel_stencil(caller::Module)
     ParallelKernel.reset_parallel_kernel(caller)
-    set_initialized(false)
-    set_package(PKG_NONE)
-    set_numbertype(NUMBERTYPE_NONE)
-    set_ndims(NDIMS_NONE)
+    set_initialized(caller, false)
+    set_package(caller, PKG_NONE)
+    set_numbertype(caller, NUMBERTYPE_NONE)
+    set_ndims(caller, NDIMS_NONE)
     return nothing
 end
