### 2023-12-18 69832c1
adjust unit tests to per module PS parameters
