### 2023-12-18 f0dbc01
adjust unit tests to per module PS parameters
