+++ test/test_projects/Diffusion/src/backends/memcopyXD.jl
@@ -1,0 +1,4 @@
+@parallel_indices (ix) function Diffusion.memcopy!(B::Data.Array, A::Data.Array)
+    B[ix] = A[ix]
+    return
+end
