+++ test/test_projects/Diffusion/test/localtest_diffusion_AMDGPU.jl
@@ -1,0 +1,14 @@
+push!(LOAD_PATH, "@stdlib")  # NOTE: this is needed to enable this test to run from the Pkg manager
+push!(LOAD_PATH, joinpath(@__DIR__, ".."))
+using Test
+using Pkg
+Pkg.activate(joinpath(@__DIR__, ".."))
+Pkg.instantiate()
+import AMDGPU
+using Diffusion
+@test diffusion2D(AMDGPUBackend) <: AMDGPU.ROCArray
+@test diffusion3D(AMDGPUBackend, Float32) <: AMDGPU.ROCArray
+B = AMDGPU.zeros(4, 4)
+A = AMDGPU.ones(4, 4)
+memcopy!(B, A)
+@test all(B .== A)
