### 2023-12-20 a868a12
improve error message
