### 2023-12-20 afce17c
enable overriding number of dimensions in need colonel definition
