### 2023-12-20 a51db22
improve error message
