+++ src/shared.jl
@@ -28,8 +28,8 @@ Return an expression that evaluates to `true` if the indices generated by @paral
 const SUPPORTED_NDIMS           = [1, 2, 3]
 const NDIMS_NONE                = 0
 const ERRMSG_KERNEL_UNSUPPORTED = "unsupported kernel statements in @parallel kernel definition: @parallel is only applicable to kernels that contain exclusively array assignments using macros from FiniteDifferences{1|2|3}D or from another compatible computation submodule. @parallel_indices supports any kind of statements in the kernels."
-const ERRMSG_CHECK_NDIMS        = "ndims must be noted LITERALLY (NOT a variable containing the ndims) and has to be one of the following: $(join(SUPPORTED_NDIMS,", "))"
-const ERRMSG_CHECK_MEMOPT       = "memopt must be a evaluatable at parse time (e.g. literal or constant) and has to be of type Bool."
+const ERRMSG_CHECK_NDIMS        = "ndims must be evaluatable at parse time (e.g. literal or constant) and has to be one of the following Integers: $(join(SUPPORTED_NDIMS,", "))"
+const ERRMSG_CHECK_MEMOPT       = "memopt must be evaluatable at parse time (e.g. literal or constant) and has to be of type Bool."
 const PSNumber                  = PKNumber
 const LOOPSIZE                  = 16
 const LOOPDIM_NONE              = 0
