### 2023-12-20 338d884
add unit tests for ndims omission
