### 2023-12-20 a23dc88
add unit test for defining ndims in kernel
