+++ test/test_projects/Diffusion/ext/Diffusion_AMDGPUExt.jl
@@ -24,7 +24,7 @@ module Diffusion_AMDGPUExt
         import AMDGPU
         using ParallelStencil
         @init_parallel_stencil(package=AMDGPU)
-        # Diffusion.memcopy!(A::Data.Array, B::Data.Array) = (@info "using AMDGPU backend"; memcopy!(A, B))
+        Diffusion.memcopy!(B::Data.Array, A::Data.Array) = (@info "using AMDGPU backend"; @parallel memcopy!(B, A))
         include(joinpath(@__DIR__, "..", "src", "backends", "memcopyXD.jl"))
     end
 end
