### 2023-12-20 751b96e
Correct ParallelStencil compatibility in diffusion test project
