+++ test/test_projects/Diffusion/Project.toml
@@ -7,12 +7,12 @@ version = "0.1.0"
 ParallelStencil = "94395366-693c-11ea-3b26-d9b7aac5d958"
 
 [compat]
-ParallelStencil = ">= 0.4.0"
+ParallelStencil = ">= 0.11.0"
 
 [weakdeps]
 AMDGPU = "21141c5a-9bdb-4563-92ae-f87d6854732e"
 CUDA = "052768ef-5323-5732-b1bb-66c8b64840ba"
 
 [extensions]
-#Diffusion_AMDGPUExt = "AMDGPU"
+Diffusion_AMDGPUExt = "AMDGPU"
 Diffusion_CUDAExt = "CUDA"
