### 2023-12-20 82a10cf
activate minimal extension test project
