+++ test/test_extensions.jl
@@ -10,7 +10,7 @@ end
     if !AMDGPU.functional() TEST_PACKAGES = filter!(x->x≠PKG_AMDGPU, TEST_PACKAGES) end
 end
 exename = joinpath(Sys.BINDIR, Base.julia_exename())
-const TEST_PROJECTS = ["Diffusion"] # ["Diffusion3D_minimal", "Diffusion3D", "Diffusion"]
+const TEST_PROJECTS = ["Diffusion3D_minimal"] # ["Diffusion3D_minimal", "Diffusion3D", "Diffusion"]
 
 @static for package in TEST_PACKAGES  eval(:(
     @testset "$(basename(@__FILE__)) (package: $(nameof($package)))" begin
