### 2024-01-02 c02b696
add expansion of ndims tuple
