+++ src/parallel.jl
@@ -111,15 +111,20 @@ parallel_async(source::LineNumberNode, caller::Module, args::Union{Symbol,Expr}.
 function parallel(source::LineNumberNode, caller::Module, args::Union{Symbol,Expr}...; package::Symbol=get_package(caller), async::Bool=false)
     if is_kernel(args[end])
         posargs, kwargs_expr, kernelarg = split_parallel_args(args, is_call=false)
-        kwargs     = extract_kwargs(caller, kwargs_expr, (:ndims, :inbounds, :memopt, :optvars, :loopdim, :loopsize, :optranges, :useshmemhalos, :optimize_halo_read, :metadata_module, :metadata_function), "@parallel <kernel>"; eval_args=(:ndims, :inbounds, :memopt, :loopdim, :optranges, :useshmemhalos, :optimize_halo_read, :metadata_module))
-        numbertype = get_numbertype(caller)
-        if !haskey(kwargs, :metadata_module)
-            get_name(kernelarg)
-            metadata_module, metadata_function = create_metadata_storage(source, caller, kernelarg)
+        kwargs = extract_kwargs(caller, kwargs_expr, (:ndims, :inbounds, :memopt, :optvars, :loopdim, :loopsize, :optranges, :useshmemhalos, :optimize_halo_read, :metadata_module, :metadata_function), "@parallel <kernel>"; eval_args=(:ndims, :inbounds, :memopt, :loopdim, :optranges, :useshmemhalos, :optimize_halo_read, :metadata_module))
+        ndims = haskey(kwargs, :ndims) ? kwargs.ndims : get_ndims(caller)
+        if typeof(ndims) <: Tuple
+            expand_ndims_tuple(ndims, true, kernelarg, kwargs, posargs...)
         else
-            metadata_module, metadata_function = kwargs.metadata_module, kwargs.metadata_function
+            numbertype = get_numbertype(caller)
+            if !haskey(kwargs, :metadata_module)
+                get_name(kernelarg)
+                metadata_module, metadata_function = create_metadata_storage(source, caller, kernelarg)
+            else
+                metadata_module, metadata_function = kwargs.metadata_module, kwargs.metadata_function
+            end
+            parallel_kernel(metadata_module, metadata_function, caller, package, ndims, numbertype, kernelarg, posargs...; kwargs)
         end
-        parallel_kernel(metadata_module, metadata_function, caller, package, numbertype, kernelarg, posargs...; kwargs)
     elseif is_call(args[end])
         posargs, kwargs_expr, kernelarg = split_parallel_args(args)
         kwargs, backend_kwargs_expr = extract_kwargs(caller, kwargs_expr, (:memopt, :configcall, :∇, :ad_mode, :ad_annotations), "@parallel <kernelcall>", true; eval_args=(:memopt,))
@@ -146,25 +151,30 @@ function parallel_indices(source::LineNumberNode, caller::Module, args::Union{Sy
     posargs, kwargs_expr, kernelarg = split_parallel_args(args, is_call=false)
     kwargs = extract_kwargs(caller, kwargs_expr, (:ndims, :inbounds, :memopt, :optvars, :loopdim, :loopsize, :optranges, :useshmemhalos, :optimize_halo_read, :metadata_module, :metadata_function), "@parallel_indices"; eval_args=(:ndims, :inbounds, :memopt, :loopdim, :optranges, :useshmemhalos, :optimize_halo_read, :metadata_module))
     indices_expr = posargs[1]
-    if is_splatarg(indices_expr)
-        parallel_indices_splatarg(caller, package, kwargs_expr, posargs..., kernelarg; kwargs)
+    ndims = haskey(kwargs, :ndims) ? kwargs.ndims : get_ndims(caller)
+    if typeof(ndims) <: Tuple
[…]
