### 2024-01-02 7759cb2
add functions to deal with signatures
