### 2024-01-02 9446497
enable expand of ndims symbol
