### 2024-01-03 2286198
move to interpretation syntax
