### 2024-01-03 1ad7e1b
use ndims expansion syntax
