### 2024-01-03 08b8088
Add XD test for third dimension
