+++ test/test_projects/Diffusion/test/localtest_diffusion_CUDA.jl
@@ -12,3 +12,7 @@ B = CUDA.zeros(4, 4)
 A = CUDA.ones(4, 4)
 memcopy!(B, A)
 @test all(B .== A)
+D = CUDA.zeros(4, 4, 4)
+C = CUDA.ones(4, 4, 4)
+memcopy!(D, C)
+@test all(D .== C)
