### 2024-01-03 f56aafb
rename expand to interpolate
