### 2024-01-03 7504424
removes CI for Julia 1.8
