+++ .github/workflows/ci.yml
@@ -15,7 +15,7 @@ jobs:
           #- '1.8' # Minimum required Julia version (due to CellArrays' AMDGPU dependency 1.7 and due to Enzyme 1.8).
           - '1.9'
           - '1'   # Latest stable 1.x release of Julia
-          - 'nightly'
+          #- 'nightly'
         os:
           - ubuntu-latest
           - macOS-latest
