### 2024-01-03 7152bdc
add documentation on ndims expansion and interpolation
