### 2024-01-03 6ac1d0d
Make type aliases in data module constant
