### 2024-01-04 e979e9d
Remove multiple extensions test again
