### 2024-01-04 a030c1a
enable N substitution instead of interpolation of ndims
