+++ test/test_parallel.jl
@@ -76,6 +76,40 @@ import ParallelStencil.@gorgeousexpand
                 @test @prettystring(1, @parallel ∇=(V.x->V̄.x, V.y->V̄.y) f!(V.x, V.y, a)) == "@parallel configcall = f!(V.x, V.y, a) ParallelStencil.ParallelKernel.AD.autodiff_deferred!(Enzyme.Reverse, f!, (EnzymeCore.DuplicatedNoNeed)(V.x, V̄.x), (EnzymeCore.DuplicatedNoNeed)(V.y, V̄.y), (EnzymeCore.Const)(a))"
             end;
             @testset "@parallel <kernel>" begin
+                @testset "N substitution | ndims tuple expansion" begin
+                    @testset "N substitution (N=3)" begin
+                        expansion = @prettystring(1, @parallel N=3 f(A::Data.Array{N}, B::Data.Array{N}, c::Integer) where T <: PSNumber = (@all(A) = @all(B)^c; return))
+                        @test occursin("f(A::Data.Array{3}, B::Data.Array{3},", expansion)
+                    end;
+                    @testset "N substitution (N=2)" begin
+                        expansion = @prettystring(1, @parallel N=2 f(A::Data.Array{N}, B::Data.Array{N}, c::Integer) where T <: PSNumber = (@all(A) = @all(B)^c; return))
+                        @test occursin("f(A::Data.Array{2}, B::Data.Array{2},", expansion)
+                    end;
+                    @testset "N substitution (N=1)" begin
+                        expansion = @prettystring(1, @parallel N=1 f(A::Data.Array{N}, B::Data.Array{N}, c::Integer) where T <: PSNumber = (@all(A) = @all(B)^c; return))
+                        @test occursin("f(A::Data.Array{1}, B::Data.Array{1},", expansion)
+                    end;
+                    @testset "N substitution (N=ndims)" begin
+                        expansion = @prettystring(1, @parallel N=ndims f(A::Data.Array{N}, B::Data.Array{N}, c::Integer) where T <: PSNumber = (@all(A) = @all(B)^c; return))
+                        @test occursin("f(A::Data.Array{3}, B::Data.Array{3},", expansion)
+                    end;
+                    @testset "N substitution (ndims=2, N=ndims+2)" begin
+                        expansion = @prettystring(1, @parallel ndims=2 N=ndims+2 f(A::Data.Array{N}, B::Data.Array{N}, c::Integer) where T <: PSNumber = (@all(A) = @all(B)^c; return))
+                        @test occursin("f(A::Data.Array{4}, B::Data.Array{4},", expansion)
+                    end;
+                    @testset "ndims tuple expansion (ndims=(1,2,3), N=ndims)" begin
+                        expansion = @prettystring(2, @parallel ndims=(1,2,3) N=ndims f(A::Data.Array{T,N}, B::Data.Array{T,N}, c::Integer) where T <: PSNumber = (@all(A) = @all(B)^c; return))
+                        @test occursin("@parallel ndims = 1 function f(A::Data.Array{T, 1}, B::Data.Array{T, 1},", expansion)
+                        @test occursin("@parallel ndims = 2 function f(A::Data.Array{T, 2}, B::Data.Array{T, 2},", expansion)
+                        @test occursin("@parallel ndims = 3 function f(A::Data.Array{T, 3}, B::Data.Array{T, 3},", expansion)
+                    end;
+                    @testset "ndims tuple expansion (ndims=(1,3), N=ndims.+1)" begin
[…]
