### 2024-01-05 7acd83a
update docstring for kwarg N
