### 2024-01-05 975f9ca
enable Data.Index as eltype in allocators
