### 2024-01-08 f745411
add unit tests for Data.Index array creation
