+++ test/ParallelKernel/test_allocators.jl
@@ -15,7 +15,7 @@ end
     import ParallelStencil.ParallelKernel.AMDGPU
     if !AMDGPU.functional() TEST_PACKAGES = filter!(x->x≠PKG_AMDGPU, TEST_PACKAGES) end
 end
-
+const DATA_INDEX = ParallelStencil.INT_THREADS # TODO: using Data.Index does not work in combination with @reset_parallel_kernel, because the macros from module Test alternate the order of evaluation, resulting in the Data module being replaced with an empty module before Data.Index is evaluated. If at some point the indexing varies depending on the used package, then something more sophisticated is needed here (e.g., wrapping the test for each package in a module and using then Data.Index everywhere).
 
 @static for package in TEST_PACKAGES  eval(:(
     @testset "$(basename(@__FILE__)) (package: $(nameof($package)))" begin
@@ -71,6 +71,11 @@ end
                 @test occursin("xx::Float32", call)
                 @test occursin("zz::Float32", call)
                 @test occursin("xz::Float32", call)
+                call = @prettystring(1, @CellType SymmetricTensor2D_Index fieldnames=(xx, zz, xz) eltype=Data.Index)
+                @test occursin("struct SymmetricTensor2D_Index <: ParallelStencil.ParallelKernel.FieldArray{Tuple{3}, Data.Index, length([3])}", call)
+                @test occursin("xx::Data.Index", call)
+                @test occursin("zz::Data.Index", call)
+                @test occursin("xz::Data.Index", call)
             end;
             @reset_parallel_kernel()
         end;
@@ -85,50 +90,65 @@ end
                 @CellType SymmetricTensor2D_T fieldnames=(xx, zz, xz) parametric=true
                 @CellType SymmetricTensor2D_Float32 fieldnames=(xx, zz, xz) eltype=Float32
                 @CellType SymmetricTensor2D_Bool fieldnames=(xx, zz, xz) eltype=Bool
+                @CellType SymmetricTensor2D_Index fieldnames=(xx, zz, xz) eltype=DATA_INDEX
                 @test SymmetricTensor2D <: FieldArray
                 @test SymmetricTensor3D <: FieldArray
                 @test Tensor2D <: FieldArray
                 @test SymmetricTensor2D_T <: FieldArray
                 @test SymmetricTensor2D_Float32 <: FieldArray
                 @test SymmetricTensor2D_Bool <: FieldArray
+                @test SymmetricTensor2D_Index <: FieldArray
             end;
             @testset "mapping to package (no celldims/celltype)" begin
                 @static if $package == $PKG_CUDA
-                    @test typeof(@zeros(2,3))                   == typeof(CUDA.CuArray(zeros(Float16,2,3)))
-                    @test typeof(@zeros(2,3, eltype=Float32))   == typeof(CUDA.CuArray(zeros(Float32,2,3)))
-                    @test typeof(@ones(2,3))                    == typeof(CUDA.CuArray(ones(Float16,2,3)))
-                    @test typeof(@ones(2,3, eltype=Float32))    == typeof(CUDA.CuArray(ones(Float32,2,3)))
-                    @test typeof(@rand(2,3))                    == typeof(CUDA.CuArray(rand(Float16,2,3)))
-                    @test typeof(@rand(2,3, eltype=Float64))    == typeof(CUDA.CuArray(rand(Float64,2,3)))
[…]
