### 2024-01-08 8275b59
add test for index symbol in initialization
