### 2024-01-08 d6557b8
just minimal julia version and remove windows CI
