+++ src/ParallelKernel/EnzymeExt/autodiff_gpu.jl
@@ -1,0 +1,12 @@
+import ParallelStencil
+import Enzyme
+
+function ParallelStencil.ParallelKernel.AD.autodiff_deferred!(args...)
+    Enzyme.autodiff_deferred(args...)
+    return
+end
+
+function ParallelStencil.ParallelKernel.AD.autodiff_deferred_thunk!(args...)
+    Enzyme.autodiff_deferred_thunk(args...)
+    return
+end
