### 2024-01-11 6596ed4
remove dependencies from test packages
