+++ Project.toml
@@ -31,4 +31,4 @@ TOML = "fa267f1f-6049-4f14-aa54-33bafae1ed76"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
-test = ["Test", "TOML", "AMDGPU", "CUDA", "CellArrays", "Enzyme", "StaticArrays"]
+test = ["Test", "TOML", "Enzyme"]
