### 2024-01-12 476dc00
add empty init_AD and default methods with error message
