### 2024-01-12 eaf4298
add empty init_AD and default methods with error message
