### 2024-01-12 e30d967
add minimal specialization and add init_AD
