### 2024-01-12 790ac4f
remove Enzyme import and replace Enzyme init with init_AD call
