### 2024-01-12 f7920ad
turn all Enzyme constants into expressions
