+++ src/init_parallel_stencil.jl
@@ -53,14 +53,14 @@ function init_parallel_stencil(caller::Module, package::Symbol, numbertype::Data
     if (numbertype == NUMBERTYPE_NONE) datadoc_call = :(@doc replace(ParallelStencil.ParallelKernel.DATA_DOC_NUMBERTYPE_NONE, "@init_parallel_kernel" => "@init_parallel_stencil") Data)
     else                               datadoc_call = :(@doc replace(ParallelStencil.ParallelKernel.DATA_DOC,                 "@init_parallel_kernel" => "@init_parallel_stencil") Data)
     end
-    ParallelKernel.init_parallel_kernel(caller, package, numbertype, inbounds; datadoc_call=datadoc_call)
+    return_expr = ParallelKernel.init_parallel_kernel(caller, package, numbertype, inbounds; datadoc_call=datadoc_call)
     set_package(caller, package)
     set_numbertype(caller, numbertype)
     set_ndims(caller, ndims)
     set_inbounds(caller, inbounds)
     set_memopt(caller, memopt)
     set_initialized(caller, true)
-    return nothing
+    return return_expr
 end
 
 
