### 2024-01-12 b701605
move exceptions module to the beginning
