### 2024-01-12 b3ba215
adjust import and tests in AD unit tests
