### 2024-01-12 cc4144b
Fix AD initialization command
