+++ src/ParallelKernel/init_parallel_kernel.jl
@@ -42,7 +42,7 @@ function init_parallel_kernel(caller::Module, package::Symbol, numbertype::DataT
         data_module_shared = Data_shared(numbertype, indextype)
         pkg_import_cmd     = :()
     end
-    ad_init_cmd = :(ParallelStencil.ParallelKernel.AD.init_AD(Symbol("$PKG_THREADS")))
+    ad_init_cmd = :(ParallelStencil.ParallelKernel.AD.init_AD(ParallelStencil.ParallelKernel.PKG_THREADS))
     if !isdefined(caller, :Data) || (@eval(caller, isa(Data, Module)) &&  length(symbols(caller, :Data)) == 1)  # Only if the module Data does not exist in the caller or is empty, create it.
         if (datadoc_call==:())
             if (numbertype == NUMBERTYPE_NONE) datadoc_call = :(@doc ParallelStencil.ParallelKernel.DATA_DOC_NUMBERTYPE_NONE Data) 
