### 2024-01-12 fa55220
remove info about special AD initialization for threads
