+++ src/ParallelKernel/EnzymeExt/AD.jl
@@ -1,7 +1,7 @@
 """
 Module AD
 
-Provides GPU-compatible wrappers for automatic differentiation functions of the Enzyme.jl package. Consult the Enzyme documentation to learn how to use the wrapped functions.
+Provides GPU-compatible wrappers for automatic differentiation functions of the Enzyme.jl package. Enzyme needs to be imported before ParallelStencil in order to have it load the corresponding extension. Consult the Enzyme documentation to learn how to use the wrapped functions.
 
 # Usage
     import ParallelKernel.AD
