### 2024-02-14 aaff305
add support for CellArrays 0.2
