+++ Project.toml
@@ -20,7 +20,7 @@ ParallelStencil_EnzymeExt = "Enzyme"
 [compat]
 AMDGPU = "0.6, 0.7, 0.8"
 CUDA = "3.12, 4, 5"
-CellArrays = "0.1"
+CellArrays = "0.2"
 Enzyme = "0.11"
 MacroTools = "0.5"
 StaticArrays = "1"
