### 2024-02-14 3963132
add support for CellArrays 0.2
