### 2024-02-14 3e0a34f
add support for CellArrays 0.2
