### 2024-02-15 16e1624
define CellArrays aliases in unit tests
