+++ miniapps/acoustic-pml-2D.jl
@@ -1,0 +1,101 @@
+const USE_GPU = false  # Use GPU? If this is set false, then no GPU needs to be available
+using ParallelStencil
+using ParallelStencil.FiniteDifferences2D
+@static if USE_GPU
+    @init_parallel_stencil(CUDA, Float64, 2)
+else
+    @init_parallel_stencil(Threads, Float64, 2)
+end
+using Plots, Printf, Statistics
+
+#=
+This variation of the acoustic2D miniapp implements a perfectly-matched-layer as described in:
+
+Treeby, Bradley E., and Benjamin T. Cox. "k-Wave: MATLAB toolbox for the simulation 
+and reconstruction of photoacoustic wave fields." Journal of biomedical optics 
+15.2 (2010): 021314-021314.
+
+Cox, Benjamin T., et al. "k-space propagation models for acoustically heterogeneous media: 
+Application to biomedical photoacoustics." The Journal of the Acoustical Society of America 
+121.6 (2007): 3453-3464.
+
+and the k-wave user manual: http://www.k-wave.org/manual/k-wave_user_manual_1.1.pdf
+=#
+
+@parallel function compute_V!(Vx::Data.Array, Vy::Data.Array, P::Data.Array, αx::Data.Array, αy::Data.Array, dt::Data.Number, ρ::Data.Number, dx::Data.Number, dy::Data.Number)
+    @inn(Vx) = @inn(Vx) - dt*(1/ρ*@d_xi(P)/dx + @inn(αx)*@inn(Vx))
+    @inn(Vy) = @inn(Vy) - dt*(1/ρ*@d_yi(P)/dy + @inn(αy)*@inn(Vy))
+    return
+end
+
+@parallel function compute_P!(P::Data.Array, Vx::Data.Array, Vy::Data.Array, αx::Data.Array, αy::Data.Array, dt::Data.Number, ρ::Data.Number, dx::Data.Number, dy::Data.Number)
+    @all(P) = @all(P) - dt*(ρ*(@d_xa(Vx)/dx + @d_ya(Vy)/dy)+(@all(αx) + @all(αy))*@all(P)) 
+    return
+end
+
+##################################################
+@views function acoustic2D()
+    # Physics
+    lx, ly    = 40.0, 40.0  # domain extends
+    k         = 1.0         # bulk modulus
+    ρ         = 1.0         # density
+    t         = 0.0         # physical time
+    # Numerics
+    nx, ny    = 255, 255    # numerical grid resolution; should be a mulitple of 32-1 for optimal GPU perf
+    nt        = 1000        # number of timesteps
+    nout      = 10          # plotting frequency
+    # Derived numerics
+    dx, dy    = lx/(nx-1), ly/(ny-1) # cell sizes
+    # Array allocations
+    P         = @zeros(nx  ,ny  )
+    Vx        = @zeros(nx+1,ny  )
+    Vy        = @zeros(nx  ,ny+1)
+    αx        = @zeros(nx, ny)
+    αy        = @zeros(nx, ny)
+
+    # Initial conditions
+    P        .= Data.Array([exp(-((ix-1)*dx-0.5*lx)^2 -((iy-1)*dy-0.5*ly)^2) for ix=1:size(P,1), iy=1:size(P,2)])
+    dt        = min(dx,dy)/sqrt(k/ρ)/4.1
+    
+    c       = √(k/ρ) # Sound speed
+    αmax    = 4*c/dx
+    ξ0      = 1
+    ξmax    = 32
+    m       = 4
+    ξ       = ξ0:ξmax
+    α       = @. αmax * ((ξ-ξ0)/(ξmax-ξ0))^m # Equation (2.28) from the k-wave user manual.
+
[…]
