### 2024-03-13 979f900
move types and more in shared.jl to extensions; handle supported packages
