### 2024-03-13 91d7e80
move code in allocators to extensions
