### 2024-03-13 930ef1b
add default for extensions
