### 2024-03-13 baf46b6
adjust gpu packages import in unit tests
