### 2024-03-13 f6f9aee
adjust gpu packages import in unit tests
