### 2024-03-13 d93885a
adjust gpu packages import in unit tests
