### 2024-03-13 f77a34a
adjust gpu packages import in unit tests
