### 2024-03-13 31d8507
adjust gpu packages import in unit tests
