### 2024-03-13 f6beabc
adjust gpu packages import in unit tests
