### 2024-03-13 ded0a77
adjust gpu packages import in unit tests
