### 2024-03-13 9161c4c
adjust gpu packages import in unit tests
