### 2024-03-13 a9b3691
adjust gpu packages import in unit tests
