### 2024-03-13 4a84100
adjust gpu packages import in unit tests
