### 2024-03-14 3c44361
Improve comment on retry loading extensions
