### 2024-03-14 6df996f
add error if package needed for extension is not installed
