+++ src/ParallelKernel/Exceptions.jl
@@ -1,12 +1,13 @@
 module Exceptions
-export @ModuleInternalError, @MethodPluginError, @IncoherentCallError, @NotInitializedError, @NotLoadedError, @IncoherentArgumentError, @KeywordArgumentError, @ArgumentEvaluationError, @ArgumentError
-export ModuleInternalError, MethodPluginError, IncoherentCallError, NotInitializedError, NotLoadedError, IncoherentArgumentError, KeywordArgumentError, ArgumentEvaluationError
+export @ModuleInternalError, @MethodPluginError, @IncoherentCallError, @NotInitializedError, @NotLoadedError, @NotInstalledError, @IncoherentArgumentError, @KeywordArgumentError, @ArgumentEvaluationError, @ArgumentError
+export ModuleInternalError, MethodPluginError, IncoherentCallError, NotInitializedError, NotLoadedError, NotInstalledError, IncoherentArgumentError, KeywordArgumentError, ArgumentEvaluationError
 
 macro ModuleInternalError(msg) esc(:(throw(ModuleInternalError($msg)))) end
 macro MethodPluginError(msg) esc(:(throw(MethodPluginError($msg)))) end
 macro IncoherentCallError(msg) esc(:(throw(IncoherentCallError($msg)))) end
 macro NotInitializedError(msg) esc(:(throw(NotInitializedError($msg)))) end
 macro NotLoadedError(msg)  esc(:(throw(NotLoadedError($msg)))) end
+macro NotInstalledError(msg)  esc(:(throw(NotInstalledError($msg)))) end
 macro IncoherentArgumentError(msg) esc(:(throw(IncoherentArgumentError($msg)))) end
 macro KeywordArgumentError(msg) esc(:(throw(KeywordArgumentError($msg)))) end
 macro ArgumentEvaluationError(msg) esc(:(throw(ArgumentEvaluationError($msg)))) end
@@ -36,6 +37,11 @@ struct NotLoadedError <: Exception
     msg::String
 end
 Base.showerror(io::IO, e::NotLoadedError) = print(io, "NotLoadedError: ", e.msg)
+
+struct NotInstalledError <: Exception
+    msg::String
+end
+Base.showerror(io::IO, e::NotInstalledError) = print(io, "NotInstalledError: ", e.msg)
 
 struct IncoherentArgumentError <: Exception
     msg::String
