+++ src/ParallelKernel/init_parallel_kernel.jl
@@ -27,11 +27,13 @@ end
 
 function init_parallel_kernel(caller::Module, package::Symbol, numbertype::DataType, inbounds::Bool; datadoc_call=:())
     if package == PKG_CUDA
+        if (!CUDA_IS_INSTALLED) @NotInstalledError("CUDA was selected as package for parallelization, but CUDA.jl is not installed. CUDA functionality is provided with an extension of ParallelStencil and CUDA.jl needs therefore to be installed independently.") end
         indextype          = INT_CUDA
         data_module        = Data_cuda(numbertype, indextype)
         data_module_shared = Data_shared(numbertype, indextype)
         pkg_import_cmd     = :(import CUDA)
     elseif package == PKG_AMDGPU
+        if (!AMDGPU_IS_INSTALLED) @NotInstalledError("AMDGPU was selected as package for parallelization, but AMDGPU.jl is not installed. AMDGPU functionality is provided with an extension of ParallelStencil and AMDGPU.jl needs therefore to be installed independently.") end
         indextype          = INT_AMDGPU
         data_module        = Data_amdgpu(numbertype, indextype)
         data_module_shared = Data_shared(numbertype, indextype)
