### 2024-03-14 26a1e92
add error if package needed for extension is not installed
