### 2024-03-20 1daaa7d
make not installed warning dynamic
