### 2024-03-21 #138 @lucaferranti
The error above is expecetd and known. The new infrastructure is mainly in place, the only piece missing is the generation of the xml file, which is what the error is comlaining about
