### 2024-03-22 44319dd
create module name dynamically and remove problematic constructors
