### 2024-03-22 cbec560
add missing exports of MarkerTools functions
