### 2024-03-25 #138 @lucaferranti
> The error might be due to missing DOIs. @lucaferranti is my understanding correct?

sorry for the delay in replying, but I think it's correct, ping me if htere are still issues with editorialbot
