+++ Project.toml
@@ -13,11 +13,13 @@ StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 AMDGPU = "21141c5a-9bdb-4563-92ae-f87d6854732e"
 CUDA = "052768ef-5323-5732-b1bb-66c8b64840ba"
 Enzyme = "7da242da-08ed-463a-9acd-ee780be4f1d9"
+Polyester = "f517fe37-dbe3-4b94-8317-1923a5111588"
 
 [extensions]
 ParallelStencil_AMDGPUExt = "AMDGPU"
 ParallelStencil_CUDAExt = "CUDA"
 ParallelStencil_EnzymeExt = "Enzyme"
+ParallelStencil_PolyesterExt = "Polyester"
 
 [compat]
 AMDGPU = "0.6, 0.7, 0.8"
@@ -25,6 +27,7 @@ CUDA = "3.12, 4, 5"
 CellArrays = "0.2.1"
 Enzyme = "0.11"
 MacroTools = "0.5"
+Polyester = "0.7"
 StaticArrays = "1"
 julia = "1.9" # Minimum version supporting extensions
 
@@ -33,4 +36,4 @@ TOML = "fa267f1f-6049-4f14-aa54-33bafae1ed76"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
-test = ["Test", "TOML", "AMDGPU", "CUDA", "Enzyme"]
+test = ["Test", "TOML", "AMDGPU", "CUDA", "Enzyme", "Polyester"]
