+++ src/ParallelKernel/EnzymeExt/autodiff_gpu.jl
@@ -1,9 +1,9 @@
 import ParallelStencil
-import ParallelStencil: PKG_THREADS
+import ParallelStencil: PKG_THREADS, PKG_POLYESTER
 import Enzyme
 
 function ParallelStencil.ParallelKernel.AD.init_AD(package::Symbol)
-    if package == PKG_THREADS
+    if package in [PKG_THREADS, PKG_POLYESTER]
         Enzyme.API.runtimeActivity!(true) # NOTE: this is currently required for Enzyme to work correctly with threads
     end
 end
