+++ src/ParallelKernel/kernel_language.jl
@@ -87,6 +87,45 @@ macro pk_println(args...) check_initialized(__module__); esc(pk_println(__module
 
 
 ##
+const FORALL_DOC = """
+    @∀ x ∈ X statement
+    @∀ x in X statement
+    @∀ x = X statement
+
+Expand the `statement` for all `x` in `X`.
+
+# Arguments
+- `x`: the name of the variable the `statement` is to be expanded with.
+- `X`: the set of names or range of values `x` is to be expanded over.
+- `statement`: the statement to be expanded for all `x` in `X`.
+
+# Examples
+    @∀ i ∈ (x,z) @all(C.i) = @all(A.i) + @all(B.i)                     # Equivalent to: @all(C.x) = @all(A.x) + @all(B.x)
+                                                                       #                @all(C.z) = @all(A.z) + @all(B.z)
+
+    @∀ i ∈ (y,z) C.i[ix,iy,iz] = A.i[ix,iy,iz] + B.i[ix,iy,iz]         # Equivalent to: C.y[ix,iy,iz] = A.y[ix,iy,iz] + B.y[ix,iy,iz]
+                                                                       #                C.z[ix,iy,iz] = A.z[ix,iy,iz] + B.z[ix,iy,iz]
+
+    @∀ (ij,i,j) ∈ ((xy,x,y), (xz,x,z), (yz,y,z)) @all(C.ij) = @all(A.i) + @all(B.j)  # Equivalent to: @all(C.xy) = @all(A.x) + @all(B.y)
+                                                                                     #                @all(C.xz) = @all(A.x) + @all(B.z)
+                                                                                     #                @all(C.yz) = @all(A.y) + @all(B.z)
+
+    @∀ i ∈ 1:N @all(C[i]) = @all(A[i]) + @all(B[i])                    # Equivalent to: for i in 1:N
+                                                                       #                  @all(C[i]) = @all(A[i]) + @all(B[i])
+                                                                       #                end
+
+    @∀ i ∈ 1:N C[i][ix,iy,iz] = A[i][ix,iy,iz] + B[i][ix,iy,iz]        # Equivalent to: for i in 1:N
+                                                                       #                  C[i][ix,iy,iz] = A[i][ix,iy,iz] + B[i][ix,iy,iz]
+                                                                       #                end
+
+!!! note "Performance note"
+    Besides enabling a concise notation for certain sets of equations, the `@∀` macro is also designed to replace for loops over a smaller range of values in compute kernels, leading often to better performance.
+"""
+@doc FORALL_DOC
+macro ∀(args...) check_initialized(__module__); checkforallargs(args...); esc(∀(__module__, args...)); end
+
+
+##
 macro return_value(args...) check_initialized(__module__); checksinglearg(args...); esc(return_value(args...)); end
 
 
@@ -106,6 +145,11 @@ end
 
 function checkargs_sharedMem(args...)
     if !(2 <= length(args) <= 3) @ArgumentError("wrong number of arguments.") end
+end
+
+function checkforallargs(args...)
+    if !(length(args) == 2) @ArgumentError("wrong number of arguments.") end
+    if !((args[1].head == :call && args[1].args[1] in [:∈, :in]) || args[1].head == :(=)) @ArgumentError("the first argument must be of the form `x ∈ X, `x in X` or `x = X`.") end
 end
 
 
@@ -189,7 +233,31 @@ function pk_println(caller::Module, args...; package::Symbol=get_package(caller)
 end
 
 
-## FUNCTION FOR DIVERSE TASKS
+## FUNCTIONS FOR MATH SYNTAX
[…]
