### 2024-05-06 e83c779
add gateway macro for for all
