### 2024-05-06 abc00a1
make for all docstring clearer
