+++ src/ParallelKernel/allocators.jl
@@ -3,7 +3,7 @@ const ZEROS_DOC = """
     @zeros(args...)
     @zeros(args..., <keyword arguments>)
 
-Call `zeros(eltype, args...)`, where `eltype` is by default the `numbertype` selected with [`@init_parallel_kernel`](@ref) and the function `zeros` is chosen to be compatible with the package for parallelization selected with [`@init_parallel_kernel`](@ref) (zeros for Threads, CUDA.zeros for CUDA and AMDGPU.zeros for AMDGPU).
+Call `zeros(eltype, args...)`, where `eltype` is by default the `numbertype` selected with [`@init_parallel_kernel`](@ref) and the function `zeros` is chosen to be compatible with the package for parallelization selected with [`@init_parallel_kernel`](@ref) (zeros for Threads or Polyester, CUDA.zeros for CUDA and AMDGPU.zeros for AMDGPU).
 
 !!! note "Advanced"
     The `eltype` can be explicitly passed as keyword argument in order to be used instead of the default `numbertype` chosen with [`@init_parallel_kernel`](@ref). If no default `numbertype` was chosen [`@init_parallel_kernel`](@ref), then the keyword argument `eltype` is mandatory. This needs to be used with care to ensure that no datatype conversions occur in performance critical computations.
@@ -31,7 +31,7 @@ const ONES_DOC = """
     @ones(args...)
     @ones(args..., <keyword arguments>)
 
-Call `ones(eltype, args...)`, where `eltype` is by default the `numbertype` selected with [`@init_parallel_kernel`](@ref) and the function `ones` is chosen to be compatible with the package for parallelization selected with [`@init_parallel_kernel`](@ref) (ones for Threads, CUDA.ones for CUDA and AMDGPU.ones for AMDGPU).
+Call `ones(eltype, args...)`, where `eltype` is by default the `numbertype` selected with [`@init_parallel_kernel`](@ref) and the function `ones` is chosen to be compatible with the package for parallelization selected with [`@init_parallel_kernel`](@ref) (ones for Threads or Polyester, CUDA.ones for CUDA and AMDGPU.ones for AMDGPU).
 
 !!! note "Advanced"
     The `eltype` can be explicitly passed as keyword argument in order to be used instead of the default `numbertype` chosen with [`@init_parallel_kernel`](@ref). If no default `numbertype` was chosen [`@init_parallel_kernel`](@ref), then the keyword argument `eltype` is mandatory. This needs to be used with care to ensure that no datatype conversions occur in performance critical computations.
@@ -226,27 +226,34 @@ end
 
 ## MACROS FORCING PACKAGE, IGNORING INITIALIZATION
 
-macro zeros_cuda(args...)     check_initialized(__module__); esc(_zeros(__module__, args...; package=PKG_CUDA)); end
-macro ones_cuda(args...)      check_initialized(__module__); esc(_ones(__module__, args...; package=PKG_CUDA)); end
-macro rand_cuda(args...)      check_initialized(__module__); esc(_rand(__module__, args...; package=PKG_CUDA)); end
-macro falses_cuda(args...)    check_initialized(__module__); esc(_falses(__module__, args...; package=PKG_CUDA)); end
-macro trues_cuda(args...)     check_initialized(__module__); esc(_trues(__module__, args...; package=PKG_CUDA)); end
-macro fill_cuda(args...)      check_initialized(__module__); esc(_fill(__module__, args...; package=PKG_CUDA)); end
-macro fill!_cuda(args...)     check_initialized(__module__); esc(_fill!(__module__, args...; package=PKG_CUDA)); end
[…]
