### 2024-05-09 a4f02f3
update initialization for polyester
