### 2024-05-09 bc81321
add macro iscpu
