### 2024-05-09 451675e
add architecture agnostic threads marco
