+++ Project.toml
@@ -19,7 +19,6 @@ Polyester = "f517fe37-dbe3-4b94-8317-1923a5111588"
 ParallelStencil_AMDGPUExt = "AMDGPU"
 ParallelStencil_CUDAExt = "CUDA"
 ParallelStencil_EnzymeExt = "Enzyme"
-ParallelStencil_PolyesterExt = "Polyester"
 
 [compat]
 AMDGPU = "0.6, 0.7, 0.8"
