+++ test/test_projects/Diffusion3D_Polyester/src/Diffusion3D_Polyester.jl
@@ -1,0 +1,6 @@
+module Diffusion3D_Polyester
+    using ParallelStencil
+    using ParallelStencil.FiniteDifferences3D
+    @init_parallel_stencil(Polyester, Float64, 3)
+    include(joinpath(@__DIR__, "..", "..", "shared", "diffusion3D.jl"))
+end
