+++ test/test_projects/Diffusion3D_Polyester/Project.toml
@@ -1,0 +1,10 @@
+name = "Diffusion3D_Polyester"
+uuid = "21dd7ca1-9249-4d3e-b366-1b777d29ed82"
+authors = ["Samuel Omlin <samuel.omlin@cscs.ch>"]
+version = "0.1.0"
+
+[deps]
+ParallelStencil = "94395366-693c-11ea-3b26-d9b7aac5d958"
+
+[compat]
+ParallelStencil = ">= 0.4.0"
