### 2024-05-23 b729d84
add for all macro and change return nothing
