### 2024-05-23 e52eeef
add for all macro and change return nothing
