### 2024-05-23 0fe4a5b
add for all macro and change return nothing
