+++ src/ParallelKernel/kernel_language.jl
@@ -224,20 +224,20 @@ end
 
 ## FUNCTIONS FOR DIVERSE TASKS
 
-function threads(caller::Module, args...; package::Symbol=get_package(caller))
-    if     (package == PKG_THREADS)   return :(Base.Threads.@threads($(args...)))
-    elseif (package == PKG_POLYESTER) return :(Polyester.@batch($(args...)))
-    elseif isgpu(package)             return :(begin end)
-    else                              @KeywordArgumentError("$ERRMSG_UNSUPPORTED_PACKAGE (obtained: $package).")
-    end
-end
-
 function return_value(value)
     return :(return $value)
 end
 
 function return_nothing()
     return :(return)
+end
+
+function threads(caller::Module, args...; package::Symbol=get_package(caller))
+    if     (package == PKG_THREADS)   return :(Base.Threads.@threads($(args...)))
+    elseif (package == PKG_POLYESTER) return :(Polyester.@batch($(args...)))
+    elseif isgpu(package)             return :(begin end)
+    else                              @KeywordArgumentError("$ERRMSG_UNSUPPORTED_PACKAGE (obtained: $package).")
+    end
 end
 
 
