### 2024-05-24 e4c6896
remove polyester support temporarily
