+++ src/ParallelKernel/shared.jl
@@ -15,7 +15,7 @@ const PKG_AMDGPU                   = :AMDGPU
 const PKG_THREADS                  = :Threads
 const PKG_POLYESTER                = :Polyester
 const PKG_NONE                     = :PKG_NONE
-const SUPPORTED_PACKAGES           = [PKG_THREADS, PKG_POLYESTER, PKG_CUDA, PKG_AMDGPU]
+const SUPPORTED_PACKAGES           =  [PKG_THREADS, PKG_CUDA, PKG_AMDGPU] # [PKG_THREADS, PKG_POLYESTER, PKG_CUDA, PKG_AMDGPU]
 const INT_CUDA                     = Int64 # NOTE: unsigned integers are not yet supported (proper negative offset and range is dealing missing)
 const INT_AMDGPU                   = Int64 # NOTE: ...
 const INT_POLYESTER                = Int64 # NOTE: ...
