### 2024-06-14 a602409
fix and improve unit tests
