+++ Project.toml
@@ -1,7 +1,7 @@
 name = "ParallelStencil"
 uuid = "94395366-693c-11ea-3b26-d9b7aac5d958"
 authors = ["Samuel Omlin", "Ludovic Räss"]
-version = "0.12.0"
+version = "0.13.0"
 
 [deps]
 CellArrays = "d35fcfd7-7af4-4c67-b1aa-d78070614af4"
@@ -21,10 +21,10 @@ ParallelStencil_CUDAExt = "CUDA"
 ParallelStencil_EnzymeExt = "Enzyme"
 
 [compat]
-AMDGPU = "0.6, 0.7, 0.8, 0.9"
+AMDGPU = "0.6, 0.7, 0.8"
 CUDA = "3.12, 4, 5"
 CellArrays = "0.2.1"
-Enzyme = "0.11, 0.12"
+Enzyme = "0.11"
 MacroTools = "0.5"
 Polyester = "0.7"
 StaticArrays = "1"
