### 2024-07-05 e27d1f3
introduce nthreads_x_max for amdgpu
