### 2024-07-05 7bdb7ca
introduce nthreads_max for amdgpu
