### 2024-07-05 b8be2d9
adjust unit tests for introduced nthreads max
