### 2024-07-19 6631233
move expansion to begin and refactor into functions
