### 2024-07-19 5b25699
refactor into functions
