### 2024-07-25 #138 @omlins
Verifying that generating PDF still works before pushing changes
