### 2024-07-25 b254e6a
address remaining issues of review 1 and 2
