+++ README.md
@@ -29,7 +29,6 @@ Beyond traditional high-performance computing, ParallelStencil supports automati
 * [Installation](#installation)
 * [Questions, comments and discussions](#questions-comments-and-discussions)
 * [Your contributions](#your-contributions)
-* [Contributions](#contributions)
 * [References](#references)
 
 ## Parallelization and optimization with one macro call
@@ -525,12 +524,6 @@ To discuss numerical/domain-science issues, please post on Julia Discourse in th
 ## Your contributions
 This project welcomes your contribution! Have you developed an application with ParallelStencil that could be featured as a mini-app? Please contribute it to share it with the world! Would you like to use other methods than finite differences with math-close notation in ParallelStencil kernels? Then check out the tiny `ParallelStencil.FiniteDifferences1D` submodule as an example for enabling math-close notation for a method and contribute your own submodule! Are you missing a great feature in the core of ParallelStencil? Maybe you can contribute yourself! 
 Please open an issue to discuss your idea for a contribution beforehand. Furthermore, note that pull request should always address a significant issue in its completeness. Moreover, pull requests should blend nicely into the existing project; common sense is the primary guide in this regard (community guideline documents, e.g. [ColPrac](https://github.com/SciML/ColPrac), can be consulted in addition for inspiration). We are looking forward to your contribution!
-
-## Contributors
-The principal contributors to this project have been so far (ordered by the significance of the relative contributions):
-- Dr. Samuel Omlin ([@omlins](https://github.com/omlins)), Swiss National Supercomputing Centre (CSCS), ETH Zurich
-- Dr. Ludovic Räss ([@luraess](https://github.com/luraess)), Swiss Geocomputing Centre, University of Lausanne
-
 
 ## References
 \[1\] [Omlin, S., Räss, L., Kwasniewski, G., Malvoisin, B., & Podladchikov, Y. Y. (2020). Solving Nonlinear Multi-Physics on GPU Supercomputers with Julia. JuliaCon Conference, virtual.][JuliaCon20a]
