+++ paper/ref.bib
@@ -23,10 +23,9 @@
 
 @article{implicitglobalgrid2022,
   author={Omlin, S. and R{\"{a}}ss, L. and Utkin, I.},
-  doi = {???},
+  doi = {10.48550/arXiv.2211.15716},
   journal = {Proc. JuliaCon Conf.},
   title = {{D}istributed {P}arallelization of x{PU} {S}tencil {C}omputations in {J}ulia},
-  volume = {?},
   pages = {2},
   year = {2022},
 }
