+++ paper/figure/julia_xpu_perf.jl
@@ -7,21 +7,21 @@ using DelimitedFiles, Statistics
 function prepare_data()
 
      regex = Regex("out_diffusion3D_clean_benchm.")
-     
-     ne   = length(readdir(joinpath(@__DIR__,"../data")))
-     data = Array{Union{Float64,String}, 2}(undef, ne, 4)
-     
+
+     ne = length(readdir(joinpath(@__DIR__, "../data")))
+     data = Array{Union{Float64,String},2}(undef, ne, 4)
+
      i = 1
-     for fl in readdir(joinpath(@__DIR__,"../data"))
-          if splitext(fl)[end]!=".txt"
+     for fl in readdir(joinpath(@__DIR__, "../data"))
+          if splitext(fl)[end] != ".txt"
                continue
           end
-          data[i,1] = replace(splitext(fl)[1], regex=>"")
-          data[i,2] =  median(readdlm(joinpath(@__DIR__,"../data",fl))[:,end])
+          data[i, 1] = replace(splitext(fl)[1], regex => "")
+          data[i, 2] = median(readdlm(joinpath(@__DIR__, "../data", fl))[:, end])
           # 95% of confidence interval
-          tmp       = sort(readdlm(joinpath(@__DIR__,"../data",fl))[:,end]; rev=true)
-          data[i,3] = tmp[5]  # 5th rank  <=   (n-1.96*n^(1/2))/2 =  5.617306764100412
-          data[i,4] = tmp[16] # 16th rank <= 1+(n+1.96*n^(1/2))/2 = 15.382693235899588
+          tmp = sort(readdlm(joinpath(@__DIR__, "../data", fl))[:, end]; rev=true)
+          data[i, 3] = tmp[5]  # 5th rank  <=   (n-1.96*n^(1/2))/2 =  5.617306764100412
+          data[i, 4] = tmp[16] # 16th rank <= 1+(n+1.96*n^(1/2))/2 = 15.382693235899588
           # previous naive approach
           # data[i,3] = minimum(readdlm(joinpath(@__DIR__,"../data",fl))[:,end])
           # data[i,4] = maximum(readdlm(joinpath(@__DIR__,"../data",fl))[:,end])
@@ -32,63 +32,68 @@ end
 
 data = prepare_data()
 
-TpeakA100 = 1355
-TpeakP100 = 561
+TpeakA100 = 1370
+TpeakP100 = 559
 TpeakEPYC = 170 # vendor announced (2x85)
 TpeakXeon = 57
 
 using StatsPlots, LaTeXStrings, Plots.Measures
 
-default(fontfamily="Computer Modern",margin=5mm)
-scalefontsizes(); scalefontsizes(1.1)
+default(fontfamily="Computer Modern", margin=5mm); scalefontsizes(); scalefontsizes(1.1)
 
 # NOTE: I only used the BenchmarkTools generated results here
-exp_gpu  = [3,2,1]#["broadcasting", "explicit", "math-close"]
-nam_gpu  = repeat(exp_gpu, outer = 2)
-sx_gpu   = repeat(["P100", "A100"], inner = length(exp_gpu))
-id_A100  = [22,16,2]
-id_P100  = [21,15,1]
-data_gpu     = [Float64.(data[id_P100,2]); Float64.(data[id_A100,2])]
-data_gpu_min = [Float64.(data[id_P100,3]); Float64.(data[id_A100,3])]
[…]
