### 2024-07-31 #138 @georgebisbas
Thank you @omlins!
After having a pass over the applied edits, I think that all the main points have been successfully addressed.
