### 2024-10-02 #138 @luraess
And as version number we could set and use the `v1` tag from the paper.
