### 2024-10-07 #138 @luraess
@fcdimitr could we get the final actions done here in order to have the paper out? Thanks!
