### 2024-10-31 #138 @luraess
@fcdimitr Can you set the following:
- the code DOI to https://doi.org/10.5281/zenodo.13847579
- the paper DOI to https://doi.org/10.5281/zenodo.13882411
- the paper version to v2

and then trigger the `@editorialbot reaccept` (maybe running first, if it would work, the `recommend-accept` command to produce the proof I could check)
